% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.modflow_2d_array.R
\name{as.modflow_2d_array}
\alias{as.modflow_2d_array}
\title{Add modflow_2d_array class to object}
\usage{
as.modflow_2d_array(obj)
}
\arguments{
\item{obj}{object to add class to}
}
\description{
Add modflow_2d_array class to object
}

