% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cell_centers}
\alias{cell_centers}
\alias{cell_centers.dis}
\title{Generic function to get cell center z coordinates}
\usage{
cell_centers(...)

\method{cell_centers}{dis}(dis)
}
\value{
\code{NULL}
}
\description{
Generic function to get cell center z coordinates

Get cell center z coordinates from a dis file
}

