% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_coordinates.dis.R, R/cell_coordinates.huf.R
\name{cell_coordinates.dis}
\alias{cell_coordinates.dis}
\alias{cell_coordinates.huf}
\title{Get cell x, y and z coordinates from a dis object}
\usage{
\method{cell_coordinates}{dis}(dis, include_faces = FALSE)

\method{cell_coordinates}{huf}(huf, dis = NULL)
}
\arguments{
\item{dis}{dis object}

\item{include_faces}{logical; should face coordinates be included?}

\item{huf}{huf object}
}
\value{
list with with cell coordinate 3d arrays

3d array with cell coordinates
}
\description{
Get cell x, y and z coordinates from a dis object

Get cell coordinates from a huf object
}

