% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_dimensions.R, R/cell_dimensions.dis.R, R/cell_dimensions.huf.R
\name{cell_dimensions}
\alias{cell_dimensions}
\alias{cell_dimensions.dis}
\alias{cell_dimensions.huf}
\title{Generic function to get cell dimensions}
\usage{
cell_dimensions(...)

\method{cell_dimensions}{dis}(dis, hed = NULL, include_volume = FALSE,
  include_faces = FALSE)

\method{cell_dimensions}{huf}(huf, dis = NULL, hed = NULL,
  include_volume = FALSE, include_faces = FALSE)
}
\arguments{
\item{dis}{dis object}

\item{hed}{hed object, used for calculating the saturated thickness; if not specified, the regular cell thickness is returned}

\item{huf}{huf object}

\item{hed}{hed object, used for calculating the saturated thickness; if not specified, the regular cell thickness is returned}
}
\value{
list with cell dimension 3d arrays

list with cell dimension 3d arrays
}
\description{
Generic function to get cell dimensions

Get cell dimensions from a dis object

Get cell dimensions from a huf object
}

