% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_info.R, R/cell_info.dis.R, R/cell_info.huf.R
\name{cell_info}
\alias{cell_info}
\alias{cell_info.dis}
\alias{cell_info.huf}
\title{Generic function to get information at a certain grid cell}
\usage{
cell_info(...)

\method{cell_info}{dis}(dis, i, j)

\method{cell_info}{huf}(huf, i, j)
}
\arguments{
\item{dis}{a discretization file object}

\item{i}{row number}

\item{j}{column number}

\item{huf}{a hydrogeologic unit file object}

\item{i}{row number}

\item{j}{column number}
}
\value{
\code{NULL}

\code{NULL}
}
\description{
Generic function to get information at a certain grid cell

Get information from a dis object at a certain grid cell

Get information from a huf object at a certain grid cell
}

