% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cell_thickness.R, R/cell_thickness.dis.R, R/cell_thickness.huf.R
\name{cell_thickness}
\alias{cell_thickness}
\alias{cell_thickness.dis}
\alias{cell_thickness.huf}
\title{Generic function to get cell thicknesses}
\usage{
cell_thickness(...)

\method{cell_thickness}{dis}(dis, hed = NULL)

\method{cell_thickness}{huf}(huf, hed = NULL)
}
\arguments{
\item{dis}{dis object}

\item{hed}{hed object, used for calculating the saturated thickness; if not specified, the regular cell thickness is returned}

\item{huf}{huf object}

\item{hed}{hed object, used for calculating the saturated thickness; if not specified, the regular cell thickness is returned}
}
\value{
3d array with cell thicknesses

3d array with cell thicknesses
}
\description{
Generic function to get cell thicknesses

Get cell thicknesses from a dis object

Get cell thicknesses from a huf object
}

