% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflowDataFuncs.R
\name{chooseBudgetTerms1Col}
\alias{chooseBudgetTerms1Col}
\title{Choose Modflow Cell-by-cell Budget Terms}
\usage{
chooseBudgetTerms1Col(CBCterms)
}
\arguments{
\item{CBCterms}{vector created by \code{listBinHeaders}}
}
\description{
GUI choices for Modflow Budget Terms
}
