% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflowDataFuncs.R
\name{chooseDataSource}
\alias{chooseDataSource}
\title{Choose Source Data}
\usage{
chooseDataSource(matrixSource)
}
\arguments{
\item{matrixSource}{defines source of data}
}
\value{
list(Array3D=Array3D,rasType=rasTyp))
}
\description{
\code{chooseDataSource} is a function to choose data:  Source of input is either read from asciiFile
     or taken from previously generated DiffArray3D R dataset [see \code{ReadCBCbyLayer.R}]
}
