% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflowDataFuncs.R
\name{chooseModel}
\alias{chooseModel}
\title{Choose Modflow Model}
\usage{
chooseModel()
}
\value{
\code{model} name from vector of available models in MFModels
}
\description{
\code{chooseModel} Provides Radio button choices of available Modflow models.
     The implementation is capable of a variable number of radioButtons
     depending on length of MFModels vector via a loop
}
\examples{
     MFmodel.Params <- defineMFmodel()
     model <- chooseModel()
     M <- as.data.frame(MFmodel.Params[model,])
     fontHeading is defined with [family = "Arial",size = 24,weight = "bold",slant = "italic"]
}
