% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_bud_to_darcy}
\alias{convert_bud_to_darcy}
\title{Convert bud object fluxes to darcy velocities}
\usage{
convert_bud_to_darcy(bud, dis)
}
\arguments{
\item{bud}{bud object}

\item{dis}{dis object}
}
\value{
list of four 3d arrays: right, front, lower and magnitude
}
\description{
Convert bud object fluxes to darcy velocities
}

