% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_dis_to_real}
\alias{convert_dis_to_real}
\title{Convert modflow coordinates to real world coordinates}
\usage{
convert_dis_to_real(x = NULL, y = NULL, z = NULL, i = NULL, j = NULL,
  k = NULL, prj, dis = NULL)
}
\description{
Convert modflow coordinates to real world coordinates
}

