% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_dis_to_saturated_dis.R
\name{convert_dis_to_saturated_dis}
\alias{convert_dis_to_saturated_dis}
\title{Convert a dis object to correspond to the saturated volume}
\usage{
convert_dis_to_saturated_dis(dis, hed)
}
\arguments{
\item{dis}{dis object}

\item{hed}{hed object}
}
\value{
dis object
}
\description{
Convert a dis object to correspond to the saturated volume
}

