% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_grid_to_xyz.R
\name{convert_grid_to_xyz}
\alias{convert_grid_to_xyz}
\title{Convert modflow coordinates to real world coordinates}
\usage{
convert_grid_to_xyz(x = NULL, y = NULL, z = NULL, i = NULL, j = NULL,
  k = NULL, prj, dis = NULL)
}
\arguments{
\item{x}{modflow x coordinate}

\item{y}{modflow y coordinate}

\item{z}{modflow z coordinate}

\item{i}{modflow row number}

\item{j}{modflow column number}

\item{k}{modflow layer number}

\item{prj}{prj object}

\item{dis}{dis object}
}
\value{
data frame with real world x and y coordinates
}
\description{
Convert modflow coordinates to real world coordinates
}
\details{
Provide either xyz or ijk
}

