% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-hob-to-time-series.R
\name{convert_hob_to_time_series}
\alias{convert_hob_to_time_series}
\title{Convert a hob object to a time series data frame}
\usage{
convert_hob_to_time_series(hob, dis, prj)
}
\arguments{
\item{hob}{hob object}

\item{dis}{dis object}

\item{prj}{prj object}
}
\value{
time series data frame containing name, time and head columns
}
\description{
Convert a hob object to a time series data frame
}

