% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_huf_to_dis}
\alias{convert_huf_to_dis}
\title{Convert a parameter defined on the HUF grid to the numerical grid}
\usage{
convert_huf_to_dis(values, huf, dis, ibound, type = "arithmetic")
}
\arguments{
\item{values}{Vector of parameter values, corresponding to HGUNAM}

\item{huf}{huf object}

\item{dis}{dis object}
}
\value{
3d array
}
\description{
Convert a parameter defined on the HUF grid to the numerical grid
}

