% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-huf-to-mask.R
\name{convert_huf_to_mask}
\alias{convert_huf_to_mask}
\title{Convert a huf to a mask object}
\usage{
convert_huf_to_mask(huf, dis, bas)
}
\arguments{
\item{huf}{huf object}

\item{dis}{dis object, corresponding to the huf object}

\item{bas}{bas object, corresponding to the huf object}
}
\value{
mask rmodflow_3d_array
}
\description{
Convert a huf to a mask object
}

