% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-huf-to-nlay.R
\name{convert_huf_to_nlay}
\alias{convert_huf_to_nlay}
\title{Convert a huf object to an rmodflow_3d_array with the number of numerical layers per hydrogeological unit}
\usage{
convert_huf_to_nlay(huf, dis, bas)
}
\arguments{
\item{huf}{huf object}

\item{dis}{dis object, corresponding to the huf object}

\item{bas}{bas object, corresponding to the huf object}
}
\value{
nlay rmodflow_3d_array
}
\description{
Convert a huf object to an rmodflow_3d_array with the number of numerical layers per hydrogeological unit
}

