% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_id_to_ijk.R
\name{convert_id_to_ijk}
\alias{convert_id_to_ijk}
\title{Convert id to ijk}
\usage{
convert_id_to_ijk(id, dis)
}
\arguments{
\item{id}{cell id, providing the place of the number in an input file 3d array}

\item{dis}{a discretisation file object}
}
\description{
Convert id to ijk
}

