% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_real_to_dis}
\alias{convert_real_to_dis}
\title{Convert real world coordinates to modflow coordinates}
\usage{
convert_real_to_dis(x, y, prj, z = NULL, dis = NULL)
}
\description{
Convert real world coordinates to modflow coordinates
}

