% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coord_real_to_grid}
\alias{coord_real_to_grid}
\title{Convert real world coordinates to modflow coordinates}
\usage{
coord_real_to_grid(x, y, prj, z = NULL, dis = NULL)
}
\description{
Convert real world coordinates to modflow coordinates
}

