% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create.hob}
\alias{create.hob}
\title{Create an \code{RMODFLOW} hob object}
\usage{
create.hob(xFilter, yFilter, topFilter, bottomFilter, head, i, j, xGrid, yGrid,
  multiLayer = F, dis, useTransmissivity = F, huf = NA, mlt = NA)
}
\arguments{
\item{xFilter}{x coordinates of the monitoring well filter}

\item{yFilter}{y coordinates of the monitoring well filter}

\item{topFilter}{z coordinate of the top of the filter}

\item{bottomFilter}{z coordinate of the bottom of the filter}

\item{head}{}

\item{i}{}

\item{j}{}

\item{xGrid}{}

\item{yGrid}{}

\item{multiLayer}{}

\item{dis}{}

\item{useTransmissivity}{}

\item{huf}{}

\item{mlt}{}
}
\value{
Object of class hob
}
\description{
\code{create.hob} creates in a MODFLOW budget ascii file created from the binary file (typically *.bud) using \code{Read_budget.exe} and returns it as an \code{\link{RMODFLOW}} bud object.
}
\seealso{
\code{\link{read.hob}} and \code{\link{write.hob}}
}

