% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_array.R
\name{create_array}
\alias{create_array}
\title{Add rmodflow array class to object based on object dimensions}
\usage{
create_array(obj, dim = NULL)
}
\arguments{
\item{obj}{object to add class to}
}
\description{
Add rmodflow array class to object based on object dimensions
}

