% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bas.R
\name{create_bas}
\alias{create_bas}
\title{Create an \code{RMODFLOW} bas object}
\usage{
create_bas(dis = create_dis(), xsection = FALSE, chtoch = FALSE,
  free = TRUE, printtime = FALSE, showprogress = FALSE,
  stoperror = FALSE, stoper = 1, ibound = create_rmodflow_array(1, dim =
  c(dis$nrow, dis$ncol, dis$nlay)), hnoflo = -999,
  strt = create_rmodflow_array(0, dim = c(dis$nrow, dis$ncol, dis$nlay)))
}
\arguments{
\item{dis}{RMODFLOW dis object}

\item{xsection}{logical; is the model a 1-row cross-section?}

\item{chtoch}{logical; should flow between adjacent constant-head cells be calculated?}

\item{free}{logical; is free format used?}

\item{printtime}{logical; should start, end and elapsed times be written to the global output file?}

\item{showprogress}{logical; should progress information be displayed?}

\item{stoperror}{logical; should the model be stopped based on budget percent discrepancy?}

\item{stoper}{numeric; threshold budget percent discrepancy}

\item{ibound}{3d array specifying active (1), inactive (0) or constant head (0) status of all cells}

\item{strt}{3d array specifying starting heads}
}
\value{
Object of class bas
}
\description{
\code{create_bas} creates an \code{RMODFLOW} bas object.
}
\seealso{
\code{\link{read_bas}}, \code{\link{write_bas}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?bas.htm}
}

