% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_dis.R
\name{create_dis}
\alias{create_dis}
\title{Create an \code{RMODFLOW} dis object}
\usage{
create_dis(NLAY = 3, NROW = 10, NCOL = 10, NPER = 1, ITMUNI = 0,
  LENUNI = 0, LAYCBD = rep(0, NLAY), DELR = rep(100, NCOL),
  DELC = rep(100, NROW), TOP = matrix(0, nrow = NROW, ncol = NCOL),
  BOTM = array(rep(seq(0, -100, length = NLAY + 1)[2:(NLAY + 1)], each = NROW
  * NCOL), dim = c(NROW, NCOL, NLAY)), PERLEN = rep(1, NPER), NSTP = rep(1,
  NPER), TSMULT = rep(1, NPER), SSTR = c("SS", rep("TS", NPER - 1)))
}
\arguments{
\item{NLAY}{number of layers; defaults to 3}

\item{NROW}{number of rows; defaults to 10}

\item{NCOL}{number of columns; defaults to 10}

\item{NPER}{number of stress periods; defaults to 1}

\item{ITMUNI}{time unit; defaults to 0 (undefined)}

\item{LENUNI}{length unit; defaults to 0 (undefined)}

\item{LAYCBD}{vector of quasi-3D confining bed flags; defaults to 0 for all layers}

\item{DELR}{vector of cell widths along rows; defaults to 100 for all columns}

\item{DELC}{vector of cell widths along columns; defaults to 100 for all rows}

\item{TOP}{matrix with the top elevation of layer 1; defaults to 0 for all NROW x NCOL cells}

\item{BOTM}{3D array with the bottom elevations of all layers; defaults to NLAY layers, equally spaced between 0 (top first layer) and -100 (bottom last layer)}

\item{PERLEN}{vector of stress period lengths}

\item{NSTP}{vector of stress period time steps}

\item{TSMULT}{vector of successive time step length multipliers}

\item{SSTR}{character vector with steady state ('SS') or transient ('TS') stress period indicator}
}
\value{
Object of class dis
}
\description{
\code{create_dis} creates an \code{RMODFLOW} dis object.
}
\seealso{
\code{\link{read_dis}}, \code{\link{write_dis}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?dis.htm}
}

