% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_hob.R
\name{create_hob}
\alias{create_hob}
\title{Create an \code{RMODFLOW} hob object}
\usage{
create_hob(name, x, y, z = NULL, top = NULL, bottom = NULL, head_name,
  head, time = x * 0, STATISTIC = rep(1, length(x)), STAT_FLAG = rep(0,
  length(x)), PLOT_SYMBOL = rep(1, length(x)), dis, MAXM = dis$NLAY,
  IUHOBSV = 0, HOBDRY = -999, NOPRINT = TRUE, TOMULTH = 1, prj = NULL,
  useTransmissivity = F, huf = NA, mlt = NA)
}
\arguments{
\item{head}{}

\item{dis}{}

\item{useTransmissivity}{}

\item{huf}{}

\item{mlt}{}

\item{xFilter}{x coordinates of the monitoring well filter}

\item{yFilter}{y coordinates of the monitoring well filter}

\item{topFilter}{z coordinate of the top of the filter}

\item{bottomFilter}{z coordinate of the bottom of the filter}

\item{i}{}

\item{j}{}

\item{xGrid}{}

\item{yGrid}{}

\item{multiLayer}{}
}
\value{
Object of class hob
}
\description{
\code{create_hob} creates an RMODFLOW hob object
}
\seealso{
\code{\link{read.hob}} and \code{\link{write.hob}}
}

