% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_modflow.R
\name{create_modflow}
\alias{create_modflow}
\title{Create an \code{RMODFLOW} modflow object}
\usage{
create_modflow(...)
}
\arguments{
\item{...}{RMODFLOW objects to be included in the modflow model; if a nam object is not provided, it as added automatically}
}
\value{
Object of class modflow
}
\description{
\code{create_modflow} creates an \code{RMODFLOW} modflow object.
}
\seealso{
\code{\link{read_modflow}}, \code{\link{write_modflow}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html}
}

