% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nam.R
\name{create_nam}
\alias{create_nam}
\title{Create an \code{RMODFLOW} nam object}
\usage{
create_nam(...)
}
\arguments{
\item{...}{RMODFLOW objects to be included in the nam file}
}
\value{
Object of class nam
}
\description{
\code{create_nam} creates an \code{RMODFLOW} nam object.
}
\seealso{
\code{\link{read_nam}}, \code{\link{write_nam}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?name_file.htm}
}

