% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_oc.R
\name{create_oc}
\alias{create_oc}
\title{Create an \code{RMODFLOW} oc object}
\usage{
create_oc(dis = create_dis(), ihedfm = NA, chedfm = NA, ihedun = 666,
  iddnfm = NA, cddnfm = NA, iddnun = NA, cboufm = NA, ibouun = NA,
  compact_budget = TRUE, aux = FALSE, head_label = TRUE,
  drawdown_label = TRUE, ibound_label = TRUE, iperoc = rep(1:dis$nper,
  dis$nstp), itsoc = unlist(apply(data.frame(1, dis$nstp), 1, FUN =
  function(x) x[1]:x[2])), print_head = rep(FALSE, length(iperoc)),
  print_drawdown = rep(FALSE, length(iperoc)), print_budget = rep(FALSE,
  length(iperoc)), save_head = rep(TRUE, length(iperoc)),
  save_drawdown = rep(FALSE, length(iperoc)), save_ibound = rep(FALSE,
  length(iperoc)), save_budget = rep(TRUE, length(iperoc)))
}
\arguments{
\item{dis}{dis object}

\item{ihedfm}{code for the format in which heads will be printed; defaults to NA (heads are not printed)}

\item{chedfm}{code for the format in which heads will be saved; defaults to NA (heads are saved in binary file)}

\item{ihedun}{unit number on which heads will be saved; defaults to 666}

\item{iddnfm}{code for the format in which drawdowns will be printed; defaults to NA (drawdowns are not printed)}

\item{cddnfm}{code for the format in which drawdowns will be saved; defaults to NA (drawdowns are saved in binary file)}

\item{iddnun}{unit number on which drawdowns will be saved; defaults to NA (drawdowns are not saved)}

\item{cboufm}{code for the format in which ibound will be saved; defaults to NA (ibound is saved in binary file)}

\item{ibouun}{unit number on which ibound will be saved; defaults to NA (ibound is not saved)}

\item{compact_budget}{logical; should the COMPACT BUDGET option be used}

\item{aux}{logical; should the auxiliary data be included in the budget file}

\item{head_label}{logical; should labels be included in the unformatted head file?}

\item{drawdown_label}{logical; should labels be included in the unformatted drawdown file?}

\item{ibound_label}{logical; should labels be included in the unformatted ibound file?}

\item{iperoc}{vector of stress period numbers}

\item{itsoc}{vector of time step numbers (within the corresponding stress period)}

\item{print_head}{logical vector; should heads be printed for the time step corresponding to iperoc and itsoc?}

\item{print_drawdown}{logical vector; should drawdowns be printed for the time step corresponding to iperoc and itsoc?}

\item{print_budget}{logical vector; should budget be printed for the time step corresponding to iperoc and itsoc?}

\item{save_head}{logical vector; should heads be saved for the time step corresponding to iperoc and itsoc?}

\item{save_drawdown}{logical vector; should drawdowns be saved for the time step corresponding to iperoc and itsoc?}

\item{save_ibound}{logical vector; should ibound be saved for the time step corresponding to iperoc and itsoc?}

\item{save_budget}{logical vector; should budget be saved for the time step corresponding to iperoc and itsoc?}
}
\value{
Object of class oc
}
\description{
\code{create_oc} creates an \code{RMODFLOW} oc object.
}
\seealso{
\code{\link{read_oc}}, \code{\link{write_oc}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?oc.htm}
}

