% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rmodflow_array.R
\name{create_rmodflow_array}
\alias{create_rmodflow_array}
\title{Add rmodflow array class to object based on object dimensions}
\usage{
create_rmodflow_array(obj = NA, dim = NULL)
}
\arguments{
\item{obj}{object to add class to}

\item{dim}{the dim attribute for the array to be created; by default, dim(obj) is used}
}
\description{
Add rmodflow array class to object based on object dimensions
}

