% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflowDataFuncs.R
\name{defineMFmodel}
\alias{defineMFmodel}
\title{Defines SFWMD Modflow Model characteristics}
\usage{
defineMFmodel()
}
\value{
MFmodel.Params Data.frame providing model characteristics
}
\description{
\code{defineMFmodel} defines several SFWMD Modflow Models characteristics
}
