% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflowDataFuncs.R
\name{definePlotOpts}
\alias{definePlotOpts}
\title{Define Plot options}
\usage{
definePlotOpts()
}
\value{
list(matrixSource=matrixSrc, printingOn=printingOn, printAnnualOn=printAnnualOn,
      Animform=Animform))
}
\description{
\code{definePlotOpts} provides radio and checkboxes for Ploting options
}
\examples{

}
