% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export_animation.4d_array.R, R/export_animation.R
\name{export_animation.4d_array}
\alias{export_animation}
\alias{export_animation.4d_array}
\title{Export animations from rmodflow 4d arrays}
\usage{
\method{export_animation}{4d_array}(array, i = NULL, j = NULL, k = NULL,
  l = 1:dim(array)[4], title = "totim_unit", dis = dis, ...,
  file = "rmodflow_animation.gif", type = tools::file_ext(file),
  plot_type = "fill", width = c(600, 600, 8)[type == c("gif", "html",
  "pdf")], height = width, clean = TRUE, interval = 0.5, overlay = NULL,
  background = NULL)

export_animation(...)
}
\arguments{
\item{array}{rmodflow 4d array}

\item{i}{row number}

\item{j}{column number}

\item{k}{layer number}

\item{l}{vector of time step numbers; defaults to 1 to the size of the fourth dimension of array}

\item{title}{animation title; if equal to 'totim', the total time is shown for each frame, as taken from the 4d_array attributes, if equal to 'totim_unit', the time unit obtained from dis$itmuni is pasted to the totim times}

\item{dis}{dis object}

\item{...}{other parameters passed to plot.4d_array}

\item{file}{animation filename}

\item{type}{animation filetype; options are 'gif' (default), 'html' and 'pdf'; defaults to file extension}

\item{plot_type}{type of plot for plot.4d_array}

\item{width}{animation width; defaults to 600 pixels or 8 inch in case of pdf}

\item{height}{animation height; default equal to width}

\item{clean}{logical; should files other than the animation (e.g. frames) be removed?}

\item{interval}{time interval between frames, in seconds; can be a vector}

\item{overlay}{additional ggplot2 layers to add to all frames}

\item{background}{background ggplot2 plot for all frames}
}
\description{
Export animations from rmodflow 4d arrays

Generic function to export animations from RMODFLOW arrays
}

