% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_raster.R
\name{export_raster}
\alias{export_raster}
\title{Generic function to export GIS raster layers from RMODFLOW arrays}
\usage{
export_raster(...)
}
\description{
Generic function to export GIS raster layers from RMODFLOW arrays
}

