% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export_vector.R
\name{export_vector}
\alias{export_vector}
\title{Generic function to export GIS vector layers from RMODFLOW arrays}
\usage{
export_vector(...)
}
\description{
Generic function to export GIS vector layers from RMODFLOW arrays
}

