% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_table.R, R/export_table.rmodflow_4d_array.R
\name{export_table}
\alias{export_table}
\alias{export_table.rmodflow_4d_array}
\title{Generic function to export tables from RMODFLOW arrays}
\usage{
export_table(...)

\method{export_table}{rmodflow_4d_array}(array, k, l, dis, bas = NULL,
  mask = ifelse0(is.null(bas), array * 0 + 1, bas$ibound[, , 1]),
  prj = NULL, crs = NULL, file = "rmodflow_export.csv", type = "csv")
}
\description{
Generic function to export tables from RMODFLOW arrays

Generic function to export tables from RMODFLOW arrays
}

