% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export_vector.2d_array.R
\name{export_vector.2d_array}
\alias{export_vector.2d_array}
\title{Generic function to export vectors}
\usage{
\method{export_vector}{2d_array}(array, dis, bas = NULL,
  mask = ifelse0(is.null(bas), array * 0 + 1, bas$ibound[, , 1]),
  prj = NULL, crs = NULL, file = "rmodflow_vector",
  type = "ESRI Shapefile")
}
\description{
Generic function to export vectors
}

