% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{geomean}
\alias{geomean}
\title{Calculate a geometric mean}
\usage{
geomean(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{further arguments passed to \code{\link{prod}}}
}
\description{
Calculate a geometric mean
}
\seealso{
\code{\link{harmean}} and \code{\link{mean}}
}

