% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{harmean}
\alias{harmean}
\title{Calculate a harmonic mean}
\usage{
harmean(x, ...)
}
\arguments{
\item{x}{An invertable R object.}

\item{...}{further arguments passed to \code{\link{mean}}}
}
\description{
Calculate a harmonic mean
}
\seealso{
\code{\link{geomean}} and \code{\link{mean}}
}

