% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{ifelse0}
\alias{ifelse0}
\title{Conditional return}
\usage{
ifelse0(test, yes, no)
}
\arguments{
\item{test}{an object which can be coerced to logical mode.}

\item{yes}{return value for \code{test==TRUE}}

\item{no}{return value for \code{test==FALSE}}
}
\description{
\code{ifelse0} returns \code{yes} if \code{test} is \code{TRUE}. If \code{test} is \code{FALSE}, it returns \code{no}.
}

