% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/int_get_modflow_array.R
\name{int_get_modflow_array}
\alias{int_get_modflow_array}
\title{Get an array specified by a free-format control record from the text lines analyzed in an \code{\link{RMODFLOW}} \code{read.*} function}
\usage{
int_get_modflow_array(mfarray.lines, NROW, NCOL, NLAY)
}
\arguments{
\item{object}{MODFLOW input file text object, starting with the free-format control record}
}
\value{
A list containing the array and the remaining text of the MODFLOW input file
}
\description{
Get an array specified by a free-format control record from the text lines analyzed in an \code{\link{RMODFLOW}} \code{read.*} function
}

