% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflowDataFuncs.R
\name{listBinHeaders}
\alias{listBinHeaders}
\title{Creates list of Budget Term}
\usage{
listBinHeaders(filPtr)
}
\arguments{
\item{filPtr}{file pointer}
}
\value{
CBCTermSet as list c(firstHeader$TEXT,firstHeader$K,firstHeader$NR,firstHeader$NC,CBCterms)
}
\description{
\code{listBinHeaders} creates a list of Budget Term Headers available in
     Modflow binary cell-by-cell file
}
