% Generated by roxygen2 (4.0.2): do not edit by hand
\name{node_info}
\alias{node_info}
\alias{node_info.dis}
\alias{node_info.huf}
\title{Generic function to get information at a certain grid node}
\usage{
node_info(...)

\method{node_info}{dis}(dis, i, j)

\method{node_info}{huf}(huf, i, j)
}
\value{
\code{NULL}

\code{NULL}
}
\description{
Generic function to get information at a certain grid node

Get information from a dis object at a certain grid node

Get information from a huf object at a certain grid node
}

