% Generated by roxygen2 (4.0.2): do not edit by hand
\name{performance}
\alias{performance}
\alias{performance.hpr}
\title{Generic function to get model performance measures}
\usage{
performance(...)

\method{performance}{hpr}(hpr, savePNG = F, file = NULL, residuals = F,
  doPlot = T)
}
\value{
\code{NULL}
}
\description{
Generic function to get model performance measures

Generic function to get model performance measures
}

