% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{performance_measures}
\alias{performance_measures}
\title{Model performance measures}
\usage{
performance_measures(observations, predictions, print = F)
}
\description{
Model performance measures
}

