% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.2d_array.R
\name{plot.2d_array}
\alias{plot.2d_array}
\title{Plot a MODFLOW 2D array}
\usage{
\method{plot}{2d_array}(rmodflow_array, dis, bas = NULL,
  mask = ifelse0(is.null(bas), rmodflow_array * 0 + 1, bas$IBOUND[, , 1]),
  colour_palette = rev_rainbow, zlim = range(rmodflow_array, finite = TRUE),
  levels = pretty(zlim, nlevels), nlevels = 7, type = "fill",
  add = FALSE, height_exaggeration = 100, binwidth = round(diff(zlim)/20),
  label = TRUE, prj = NULL, target_CRS = NULL, alpha = 1,
  plot3d = FALSE, height = NULL)
}
\arguments{
\item{rmodflow_array}{an object of class 2d_array}

\item{dis}{discretization file object}

\item{bas}{basic file object; optional}

\item{mask}{a 2D array with 0 or F indicating inactive cells; optional; defaults to having all cells active or, if bas is provided, the first layer of bas$IBOUND}

\item{colour_palette}{a colour palette for imaging the array values}

\item{zlim}{vector of minimum and maximum value for the colour scale}

\item{levels}{levels to indicate on the colour scale; defaults to nlevels pretty breakpoints}

\item{nlevels}{number of levels for the colour scale; defaults to 7}

\item{type}{plot type: 'fill' (default), 'grid' or 'contour'}

\item{add}{logical; if TRUE, provide ggplot2 layers instead of object, or add 3D plot to existing rgl device; defaults to FALSE}

\item{height_exaggeration}{height exaggeration for 3D plot; optional}

\item{binwidth}{binwidth for contour plot; defaults to 1/20 of zlim}

\item{label}{logical; should labels be added to contour plot}

\item{prj}{projection file object}

\item{target_CRS}{coordinate reference system for the plot}

\item{alpha}{transparency value; defaults to 1}

\item{plot3d}{logical; should a 3D plot be made}

\item{height}{2D array for specifying the 3D plot z coordinate}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
\code{plot.2d_array} plots a MODFLOW 2D array.
}

