% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.3d_array.R
\name{plot.3d_array}
\alias{plot.3d_array}
\title{Plot a 2D section through a MODFLOW 3D array}
\usage{
\method{plot}{3d_array}(rmodflow_array, i = NULL, j = NULL, k = NULL, dis,
  bas = NULL, mask = ifelse0(is.null(bas), rmodflow_array * 0 + 1,
  bas$IBOUND), zlim = range(rmodflow_array[ifelse0(is.null(i),
  c(1:dim(rmodflow_array)[1]), i), ifelse0(is.null(j),
  c(1:dim(rmodflow_array)[2]), j), ifelse0(is.null(k),
  c(1:dim(rmodflow_array)[3]), k)], finite = TRUE),
  colour_palette = rev_rainbow, nlevels = 7, type = "fill", add = FALSE,
  ...)
}
\arguments{
\item{rmodflow_array}{an object of class 3d_array}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{dis}{discretization file object}

\item{bas}{basic file object; optional}

\item{mask}{a 3D array with 0 or F indicating inactive cells optional; defaults to having all cells active or, if bas is provided, bas$IBOUND}

\item{zlim}{vector of minimum and maximum value for the colour scale}

\item{colour_palette}{a colour palette for imaging the array values}

\item{nlevels}{number of levels for the colour scale; defaults to 7}

\item{...}{parameters provided to plot.2d_array}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
\code{plot.3d_array} plots a 2D section through a MODFLOW 3D array.
}

