% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.hpr}
\alias{plot.hpr}
\title{Plot a MODFLOW head predictions file}
\usage{
\method{plot}{hpr}(hpr, type = "scatter")
}
\arguments{
\item{hpr}{head predictions file object}

\item{type}{plot type: 'scatter' or 'residual'}
}
\description{
Plot a MODFLOW head predictions file
}

