% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.huf.R
\name{plot.huf}
\alias{plot.huf}
\title{Plot a 2D section through a MODFLOW 3D array}
\usage{
\method{plot}{huf}(huf, i = NULL, j = NULL, k = NULL, dis, bas = NULL,
  colour_palette = rev_rainbow, nlevels = 7, levels = huf$hgunam,
  type = "fill", grid = FALSE, add = FALSE, title = NULL, hed = NULL,
  l = NULL, ...)
}
\arguments{
\item{huf}{an object of class huf}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{dis}{discretization file object}

\item{bas}{basic file object; optional}

\item{colour_palette}{a colour palette for imaging the array values}

\item{nlevels}{number of levels for the colour scale; defaults to 7}

\item{levels}{labels that should be used on the factor legend; huf$hgunam is used by default}

\item{type}{plot type: 'fill' (default), 'factor' or 'grid'}

\item{grid}{logical; should grid lines be plotted? alternatively, provide colour of the grid lines.}

\item{title}{plot title}

\item{hed}{hed object for only plotting the saturated part of the grid; possibly subsetted with time step number; by default, last time step is used}

\item{l}{time step number for subsetting the hed object}

\item{...}{parameters provided to plot.2d_array}

\item{mask}{a 3D array with 0 or F indicating inactive cells optional; defaults to having all cells active or, if bas is provided, bas$ibound}

\item{zlim}{vector of minimum and maximum value for the colour scale}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
\code{plot.huf} plots a 2D section through a MODFLOW 3D array.
}

