% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.modflow_2d_array}
\alias{plot.modflow_2d_array}
\title{Plot a MODFLOW 2D array}
\usage{
\method{plot}{modflow_2d_array}(modflow_2d_array, dis, mask = modflow_2d_array
  * 0 + 1, color.palette = rev_rainbow, zlim = range(modflow_2d_array,
  finite = TRUE), levels = pretty(zlim, nlevels), nlevels = 7,
  main = "MF ARRAY plot", type = "fill", add = FALSE, xOrigin = 0,
  yOrigin = 0, height.exageration = 100, binwidth = round(diff(zlim)/20),
  label = TRUE, prj = NULL, target_CRS = NULL, alpha = 1,
  plot3d = FALSE, height = NULL)
}
\arguments{
\item{modflow_2d_array}{An object of class modflow_2d_array, or a 2D matrix}

\item{mask}{An ibound array with 1 or TRUE indicating active cells, and 0 or F indicating inactive cells}

\item{color.palette}{A color palette for imaging the parameter values}

\item{zlim}{}

\item{levels}{}

\item{nlevels}{}

\item{main}{}
}
\value{
None
}
\description{
\code{plot.modflow_2d_array} plots a MODFLOW 2D array.
}

