% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.modflow_3d_array.R
\name{plot.modflow_3d_array}
\alias{plot.modflow_3d_array}
\title{Plot a 2D section through a MODFLOW 3D array}
\usage{
\method{plot}{modflow_3d_array}(modflow_3d_array, i = NULL, j = NULL,
  k = NULL, dis, ba6 = NULL, mask = ifelse0(is.null(ba6), modflow_3d_array
  * 0 + 1, ba6$IBOUND), zlim = range(modflow_3d_array[ifelse0(is.null(i),
  c(1:dim(modflow_3d_array)[1]), i), ifelse0(is.null(j),
  c(1:dim(modflow_3d_array)[2]), j), ifelse0(is.null(k),
  c(1:dim(modflow_3d_array)[3]), k)], finite = TRUE),
  colour_palette = rev_rainbow, nlevels = 7, ...)
}
\arguments{
\item{modflow_3d_array}{an object of class modflow_3d_array}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{dis}{discretization file object}

\item{ba6}{basic file object; optional}

\item{mask}{a 3D array with 0 or F indicating inactive cells optional; defaults to having all cells active or, if ba6 is provided, ba6$IBOUND}

\item{zlim}{vector of minimum and maximum value for the colour scale}

\item{colour_palette}{a colour palette for imaging the array values}

\item{nlevels}{number of levels for the colour scale; defaults to 7}

\item{...}{parameters provided to plot.modflow_2d_array}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
\code{plot.modflow_3d_array} plots a 2D section through a MODFLOW 3D array.
}

