% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rmodflow_4d_array.R
\name{plot.rmodflow_4d_array}
\alias{plot.rmodflow_4d_array}
\title{Plot a 2D section through a MODFLOW 4D array}
\usage{
\method{plot}{rmodflow_4d_array}(array, i = NULL, j = NULL, k = NULL,
  l = NULL, ...)
}
\arguments{
\item{array}{an object of class rmodflow_3d_array}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{l}{time step number to plot}

\item{...}{parameters provided to plot.rmodflow_3d_array}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
\code{plot.rmodflow_4d_array} plots a 2D section through a MODFLOW 4D array.
}

