% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.sen.R
\name{plot.sen}
\alias{plot.sen}
\title{Plot a MODFLOW sensitivity analysis object}
\usage{
\method{plot}{sen}(sen, type = "css")
}
\arguments{
\item{sen}{sensitivity analysis object}

\item{plot}{type: 'css' or 'dss'}
}
\description{
Plot a MODFLOW sensitivity analysis object
}

