% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{prj_auxiliary}
\alias{prj_auxiliary}
\alias{rmf_get_prj}
\alias{rmf_get_prj.dis}
\alias{rmf_get_prj.modflow}
\alias{rmf_has_prj}
\alias{rmf_has_prj.dis}
\alias{rmf_has_prj.modflow}
\alias{rmf_set_prj}
\alias{rmf_set_prj.character}
\alias{rmf_set_prj.dis}
\alias{rmf_set_prj.modflow}
\alias{rmf_transform_prj}
\alias{rmf_transform_prj.prj}
\alias{rmf_transform_prj.dis}
\alias{rmf_transform_prj.modflow}
\title{Functions to get, set, transform and check presence of prj objects}
\usage{
rmf_get_prj(...)

\method{rmf_get_prj}{dis}(dis)

\method{rmf_get_prj}{modflow}(modflow)

rmf_has_prj(...)

\method{rmf_has_prj}{dis}(dis)

\method{rmf_has_prj}{modflow}(modflow)

rmf_set_prj(...)

\method{rmf_set_prj}{character}(file, dis, prj = rmf_get_prj(dis))

\method{rmf_set_prj}{dis}(dis, prj)

\method{rmf_set_prj}{modflow}(modflow, prj)

rmf_transform_prj(...)

\method{rmf_transform_prj}{prj}(prj, crs)

\method{rmf_transform_prj}{dis}(dis, crs)

\method{rmf_transform_prj}{modflow}(modflow, crs)
}
\arguments{
\item{dis}{\code{RMODFLOW} dis object}

\item{modflow}{\code{RMODFLOW} modflow object}

\item{file}{path to discretization file; typically "*.dis"}

\item{prj}{\code{RMODFLOW} prj object}

\item{crs}{coordinate reference system to transform to. Input for \code{sf::st_crs}.}
}
\value{
\code{rmf_get_prj} returns a \code{RMODFLOW} prj object if present; otherwise \code{NULL}

\code{rmf_has_prj} returns a logical depending on whether or not a \code{RMODFLOW} prj object is present

\code{rmf_set_prj} returns either a \code{RMODFLOW} dis or modflow object with the prj set or nothing when writing directly to a file.

\code{rmf_transform_prj} returns an \code{RMODFLOW} prj object with transformed crs
}
\description{
Functions to get, set, transform and check presence of prj objects
}
\details{
If \code{prj} information is already present, a warnings is raised when overwriting.

\code{rmf_set_prj.character} writes the projection information of \code{prj} directly into the header comments of the discretization file

\code{rmf_transform_prj} transforms the origin coordinates to the new crs. If no \code{prj} was set, an error is raised.
}
\examples{
dis <- rmf_read_dis(rmf_example_file('water-supply-problem.dis'))
rmf_get_prj(dis)

m <- rmf_read(rmf_example_file('example-model.nam'), verbose = FALSE)
rmf_get_prj(m)

# return NULL
rmf_get_prj(rmf_create_dis())

rmf_has_prj(dis)
rmf_has_prj(m)
rmf_has_prj(rmf_create_dis())

dis <- rmf_create_dis()
prj <- rmf_create_prj(origin = c(100, -150))
rmf_set_prj(dis, prj)

# write directly to header comments of file
f <- tempfile()
rmf_write_dis(dis, file = f)
rmf_set_prj(f, dis, prj)
rmf_read_dis(f)

prj <- rmf_create_prj(origin = c(152082, 168000.2), rotation = -12, crs = 31370)
dis <- rmf_create_dis(prj = prj)

rmf_transform_prj(prj, crs = 4326)
rmf_transform_prj(dis, crs = 3044)

\dontrun{
# error when no prj is present
rmf_transform_prj(rmf_create_dis(), 3044)
}

}
