% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.bud}
\alias{read.bud}
\title{Read a MODFLOW budget file}
\usage{
read.bud(file)
}
\arguments{
\item{file}{Filename; typically "Budget.out"}
}
\value{
Object of class bud
}
\description{
\code{read.bud} reads in a MODFLOW budget file
}

