% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.dis}
\alias{read.dis}
\title{Read a MODFLOW discretization file}
\usage{
read.dis(file)
}
\arguments{
\item{file}{Filename; typically *.dis}
}
\value{
Object of class dis
}
\description{
\code{read.dis} reads in a MODFLOW discretization file and returns it as an \code{\link{RMODFLOW}} dis object.
}

