\name{read.gms2dgrid}
\alias{read.gms2dgrid}
\title{Read a GMS 2D grid file}
\usage{
read.gms2dgrid(file)
}
\arguments{
  \item{file}{Filename}
}
\value{
Object of class gms2dgrid
}
\description{
\code{read.gms2dgrid} reads in a GMS 2D grid file and
returns it as an \code{\link{RMODFLOW}} gms2dgrid object.
}

