\name{read.hob}
\alias{read.hob}
\title{Read a MODFLOW head observations file}
\usage{
read.hob(file)
}
\arguments{
  \item{file}{Filename; typically "*.hob"}
}
\value{
Object of class hob
}
\description{
\code{read.hob} reads in a MODFLOW head observations file
and returns it as an \code{\link{RMODFLOW}} hob object.
}

