\name{read.pval}
\alias{read.pval}
\title{Read a MODFLOW parameter value file}
\usage{
read.pval(file, read.all = F)
}
\arguments{
  \item{file}{Filename; typically "*.pval"}
}
\value{
Object of class pval
}
\description{
\code{read.pval} reads in a MODFLOW parameter value file
and returns it as an \code{\link{RMODFLOW}} pval object.
}

