% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_ba6}
\alias{read_ba6}
\title{Read a MODFLOW basic file}
\usage{
read_ba6(file, dis = read_dis(paste(substring(file, 1, nchar(file) - 4),
  ".dis", sep = "")))
}
\arguments{
\item{file}{Filename; typically "*.ba6"}
}
\value{
Object of class ba6
}
\description{
\code{read_ba6} reads in a MODFLOW basic file and returns it as an \code{\link{RMODFLOW}} ba6 object.
}

