% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_bud}
\alias{read_bud}
\title{Read a MODFLOW budget file}
\usage{
read_bud(file, binary = TRUE)
}
\arguments{
\item{file}{Filename; typically "Budget.out"}
}
\value{
Object of class bud
}
\description{
\code{read_bud} reads in a MODFLOW budget file
}

