% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_dis}
\alias{read_dis}
\title{Read a MODFLOW discretization file}
\usage{
read_dis(file)
}
\arguments{
\item{file}{filename; typically '*.dis'}
}
\value{
object of class dis
}
\description{
\code{read_dis} reads in a MODFLOW discretization file and returns it as an \code{\link{RMODFLOW}} dis object.
}

