% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_gms_2d_grid}
\alias{read_gms_2d_grid}
\title{Read a GMS 2D grid file}
\usage{
read_gms_2d_grid(file)
}
\arguments{
\item{file}{filename}
}
\value{
object of class gms2dgrid
}
\description{
\code{read_gms_2d_grid} reads in a GMS 2D grid file and returns it as an \code{\link{RMODFLOW}} gms2dgrid object.
}

