% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_hed}
\alias{read_hed}
\title{Read a MODFLOW head file}
\usage{
read_hed(file, dis = read_dis(paste(substring(file, 1, nchar(file) - 4),
  ".dis", sep = "")), ba6 = read_ba6(paste(substring(file, 1, nchar(file) -
  4), ".ba6", sep = "")), convert_HNOFLO_to_NA = TRUE)
}
\arguments{
\item{file}{Filename; typically "*.hed"}
}
\value{
Object of class hed
}
\description{
\code{read_hed} reads in a MODFLOW head file and returns it as an \code{\link{RMODFLOW}} hed object.
}

