% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_huf}
\alias{read_huf}
\title{Read a MODFLOW hydrogeologic unit flow file}
\usage{
read_huf(file, dis = read_dis(paste(substring(file, 1, nchar(file) - 4),
  ".dis", sep = "")))
}
\arguments{
\item{file}{Filename; typically *.huf}

\item{dis}{Corresponding discretization file; typically *.dis}
}
\value{
Object of class huf
}
\description{
\code{read_ba6} reads in a MODFLOW hydrogeologic unit flow file and returns it as an \code{\link{RMODFLOW}} huf object.
}

