% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_kdep}
\alias{read_kdep}
\title{Read a MODFLOW hydraulic conductivity depth-dependence capability file}
\usage{
read_kdep(file, dis = read_dis(paste(substring(file, 1, nchar(file) - 4),
  ".dis", sep = "")), huf = read_huf(paste(substring(file, 1, nchar(file) -
  4), ".huf", sep = ""), dis))
}
\arguments{
\item{file}{Filename; typically *.kdep}

\item{dis}{discretization file object; defaults to that with the same filename but with extension '.dis'}

\item{huf}{hydrogeologic unit file object; defaults to that with the same filename but with extension '.huf'}
}
\value{
object of class kdep
}
\description{
\code{read_kdep} reads in a MODFLOW Hydraulic-Conductivity Depth-Dependence Capability file and returns it as an \code{\link{RMODFLOW}} kdep object.
}

