% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_mlt}
\alias{read_mlt}
\title{Read a MODFLOW multiplier file}
\usage{
read_mlt(file, dis = read_dis(paste(substring(file, 1, nchar(file) - 4),
  ".dis", sep = "")))
}
\arguments{
\item{file}{filename; typically '*.mlt'}

\item{dis}{discretization file object; defaults to that with the same filename but with extension '.dis'}
}
\value{
object of class mlt
}
\description{
\code{read_mlt} reads in a MODFLOW multiplier file and returns it as an \code{\link{RMODFLOW}} mlt object.
}

