% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_modflow_array}
\alias{read_modflow_array}
\title{Get an array specified by a free-format control record from the text lines analyzed in an \code{\link{RMODFLOW}} \code{read.*} function}
\usage{
read_modflow_array(remaining_lines, NROW, NCOL, NLAY)
}
\arguments{
\item{object}{MODFLOW input file text object, starting with the free-format control record}
}
\value{
A list containing the array and the remaining text of the MODFLOW input file
}
\description{
Get an array specified by a free-format control record from the text lines analyzed in an \code{\link{RMODFLOW}} \code{read.*} function
}

