% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{read_modflow_comments}
\alias{read_modflow_comments}
\title{Read comments
Internal function used in the read_* functions to read comments}
\usage{
read_modflow_comments(remaining_lines)
}
\description{
Read comments
Internal function used in the read_* functions to read comments
}

