% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_nam}
\alias{read_nam}
\title{Read a MODFLOW name file}
\usage{
read_nam(file)
}
\arguments{
\item{file}{Filename; typically "*.nam"}
}
\value{
Object of class nam
}
\description{
\code{read_nam} reads in a MODFLOW name file and returns it as an \code{\link{RMODFLOW}} nam object.
}

