% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_pval}
\alias{read_pval}
\title{Read a MODFLOW parameter value file}
\usage{
read_pval(file, read_all = F)
}
\arguments{
\item{file}{Filename; typically "*.pval"}

\item{read_all}{A logical value indicating if \code{NP} parameters should be read, or the full parameter table (only relevant if external codes use the pval file for storing additional parameters).}
}
\value{
Object of class pval
}
\description{
\code{read_pval} reads in a MODFLOW parameter value file and returns it as an \code{\link{RMODFLOW}} pval object.
}

