% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{read_variables}
\alias{read_variables}
\title{Read modflow variables
If all are numbers, returns numeric, otherwise returns character vector}
\usage{
read_variables(remaining_lines)
}
\description{
Read modflow variables
If all are numbers, returns numeric, otherwise returns character vector
}

