% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflowDataFuncs.R
\name{readCBCHeader}
\alias{readCBCHeader}
\title{Read cell-by-cell budget header record}
\usage{
readCBCHeader(filPtr)
}
\arguments{
\item{filPtr}{file pointer}
}
\value{
\code{header} as a list c(KSTP,KPER,TEXT,NC,NR,K)
}
\description{
\code{readCBCHeader} Reads just the Header record from a binary cell-by-Cell budget file
}
