% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflowDataFuncs.R
\name{readgridPoints}
\alias{readgridPoints}
\title{Read Model Mesh Points}
\usage{
readgridPoints(Modelgrd.Path, Model.Shape)
}
\value{
ModelGridCoords Data.frame providing model row, column and coordinates
}
\description{
\code{defineMFmodel} Reads arcGIS shape file of model mesh
   -Converts mesh polygons to points referencing the center of the model cells
   -Changes spatial reference if necessary
   -Converts point coordinates into a data frame
}
\examples{
     ModelGridCoords<-readgridPoints(Modelgrd.Path,Model.Shape)

     ---Using futures:
     g <- future({readgridPoints(Modelgrd.Path,Model.Shape)})
     ModelGridCoords <-future::value(g)
}
