% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflowDataFuncs.R
\name{readHeadsbin}
\alias{readHeadsbin}
\title{Read Heads by Stress Period}
\usage{
readHeadsbin(filPtr, SP_rng)
}
\arguments{
\item{filPtr}{file pointer}

\item{SP_rng}{[stress period range]}
}
\value{
bigVector
}
\description{
\code{readHeadsbin} searchs for a Modflow binary Heads Layers
     and returns a vector of values by stress periods
     identified in range of values
}
