% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflowDataFuncs.R
\name{readHeadsbinAtPnts}
\alias{readHeadsbinAtPnts}
\title{Read Heads for a vector of points at specified Stress Periods}
\usage{
readHeadsbinAtPnts(filPtr, SP_rng, PointVector)
}
\arguments{
\item{filPtr}{file pointer}

\item{SP_rng}{range of stress periods}

\item{PointVector}{List of points}
}
\value{
listOfPnts [with values for stress period range]
}
\description{
\code{readHeadsbinAtPnts} searchs for Heads by Layer
     and returns a vector of values at each Point
     by Stress Periods identified in range of values
}
