% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflowDataFuncs.R
\name{readHeadsbinByLay}
\alias{readHeadsbinByLay}
\title{Read Heads by Layer}
\usage{
readHeadsbinByLay(to.Read, selectLayer, maxSP)
}
\arguments{
\item{filPtr, }{selectLayer}
}
\value{
bigVector
}
\description{
\code{readHeadsbinByLay} searchs for a Modflow binary Heads Layer
     and returns a vector of values by stress periods
     identified in range of values
}
