% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflowDataFuncs.R
\name{readHeadsHeader}
\alias{readHeadsHeader}
\title{Read Heads header record}
\usage{
readHeadsHeader(filPtr)
}
\arguments{
\item{filPtr}{file pointer}
}
\value{
\code{header} as a list c(KSTP,KPER,PERTIM,TOTTIM,TEXT,NC,NR,K)
}
\description{
\code{readHeadsHeader} Reads just the Header record from a binary Heads file
}
