% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{remove_comments_end_of_line}
\alias{remove_comments_end_of_line}
\title{Remove comments at the end of a string}
\usage{
remove_comments_end_of_line(line)
}
\arguments{
\item{line}{A string.}
}
\value{
The string, without the commented part.
}
\description{
Remove comments at the end of a string
}

