% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{remove_comments_from_lines}
\alias{remove_comments_from_lines}
\title{Remove comments at the beginning of a vector of strings}
\usage{
remove_comments_from_lines(lines)
}
\arguments{
\item{lines}{Vector of strings, typically obtained from using \code{scan} with \code{sep='\n'}.}
}
\value{
Vector of strings without the commented lines.
}
\description{
Remove comments at the beginning of a vector of strings
}

