% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{remove_empty_strings}
\alias{remove_empty_strings}
\title{Remove empty elements from a vector of strings.}
\usage{
remove_empty_strings(vector_of_strings)
}
\arguments{
\item{vector_of_strings}{Vector of strings.}
}
\value{
Vector of strings without the empty items.
}
\description{
Remove empty elements from a vector of strings.
}

