% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{rmf_analyze}
\alias{rmf_analyze}
\title{Analyze a MODFLOW model}
\usage{
rmf_analyze(
  path,
  code = "2005",
  evaluate = NULL,
  include = NULL,
  transform = NULL,
  backup = TRUE,
  restore = FALSE,
  visualize = interactive(),
  ...
)
}
\arguments{
\item{path}{Path to the NAM file. Typically with extension \code{.nam}.}

\item{code}{Name of the MODFLOW variant to use, or path to the executable.}

\item{evaluate}{Vector of PVAL file parameter values to evaluate. This should
be a named vector if not all parameters are provided in their order of
occurrence, where the names (can be regular expressions to) match the
parameter names. Parameters that are not mentioned take the value from the
PVAL file. If \code{NULL} (default), no values are changed in the original
PVAL file.}

\item{include}{Character vector indicating which PVAL file parameters should
be included. Regular expressions can be used to include multiple parameters
at once. Parameters that are not mentioned are not included. If \code{NULL}
(default), all parameters are included.}

\item{transform}{Character vector of transformations. This should be a named
vector if not all parameters are transformed, and listed in their order of
occurrence, where the names (can be regular expressions to) match the
parameter names. Parameters that are not mentioned are not transformed. If
\code{NULL} (default), no transformations are done. Currently only \code{"log"} is
supported.}

\item{backup}{Logical. Should a backup (with \code{.old} suffix) of the original
PVAL file be created? Defaults to \code{TRUE}.}

\item{restore}{Logical. Should the original PVAL file be restored? Defaults
to \code{FALSE}.}

\item{visualize}{Logical. Should the results be visualized during the
analysis? Defaults to \code{TRUE} in an interactive session; \code{FALSE} otherwise.}

\item{...}{Optional arguments passed to \code{\link[=rmf_execute]{rmf_execute()}}.}
}
\value{
An \code{rmf_analyze} object, which is a list with dimensionless scaled
sensitivities (dss) and composite scaled sensitivies (css).
}
\description{
This function performs local sensitivity analysis of a MODFLOW model .
}
\details{
Only works with models using MODFLOW parameters and having a head
predictions output file as defined in the HOB object
}
\seealso{
\code{\link[=rmf_execute]{rmf_execute()}} for executing a MODFLOW model.\cr
\code{\link[=rmf_optimize]{rmf_optimize()}} for optimizing a MODFLOW model.
}
