% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-animate.R, R/rmf-animate.rmf-4d-array.R
\name{rmf_animate}
\alias{rmf_animate}
\alias{rmf_animate.rmf_4d_array}
\title{Generic function to make animations from RMODFLOW arrays}
\usage{
rmf_animate(...)

\method{rmf_animate}{rmf_4d_array}(array, i = NULL, j = NULL, k = NULL,
  l = 1:dim(array)[4], title = "totim_unit", dis = dis, ...,
  file = "rmf_animation.gif", type = tools::file_ext(file),
  plot_type = "fill", width = c(600, 600, 8)[type == c("gif", "html",
  "pdf")], height = width, clean = TRUE, interval = 0.5, overlay = NULL,
  background = NULL)
}
\arguments{
\item{...}{other parameters passed to plot.rmf_4d_array}

\item{array}{rmf 4d array}

\item{i}{row number}

\item{j}{column number}

\item{k}{layer number}

\item{l}{vector of time step numbers; defaults to 1 to the size of the fourth dimension of array}

\item{title}{animation title; if equal to 'totim', the total time is shown for each frame, as taken from the rmf_4d_array attributes, if equal to 'totim_unit', the time unit obtained from dis$itmuni is pasted to the totim times}

\item{dis}{dis object}

\item{file}{animation filename}

\item{type}{animation filetype; options are 'gif' (default), 'html' and 'pdf'; defaults to file extension}

\item{plot_type}{type of plot for plot.rmf_4d_array}

\item{width}{animation width; defaults to 600 pixels or 8 inch in case of pdf}

\item{height}{animation height; default equal to width}

\item{clean}{logical; should files other than the animation (e.g. frames) be removed?}

\item{interval}{time interval between frames, in seconds; can be a vector}

\item{overlay}{additional ggplot2 layers to add to all frames}

\item{background}{background ggplot2 plot for all frames}
}
\description{
Generic function to make animations from RMODFLOW arrays

Animations from rmf 4d arrays
}
