% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_as_array}
\alias{rmf_as_array}
\alias{rmf_as_array.rmf_list}
\title{Convert an object to a RMODFLOW array}
\usage{
rmf_as_array(...)

\method{rmf_as_array}{rmf_list}(obj, dis, select = 4, na_value = 0,
  sparse = TRUE, kper = attr(obj, "kper"), fun = sum, ...)
}
\arguments{
\item{...}{additional arguments passed to fun.}

\item{obj}{a \code{rmf_list}}

\item{dis}{a \code{RMODFLOW dis} object}

\item{select}{either a single integer or character specifying which column to select from the \code{rmf_list}. Defaults to the 4th column, i.e. the first variable.}

\item{na_value}{value of the cells in the array which are not specified in obj; defaults to 0}

\item{sparse}{logical; indicating if the returned array should be 2D or 3D. See details. Defaults to TRUE.}

\item{kper}{sets the kper attribute of the returned array; defaults to the kper attribute of obj}

\item{fun}{function to compute values in the case multiple values are defined for the same MODFLOW cell. Typically either \code{mean} or \code{sum}. Defaults to sum.}
}
\value{
either a \code{rmf_2d_array} or a \code{rmf_3d_array}
}
\description{
Convert an object to a RMODFLOW array

Convert a rmf_list to a RMODFLOW array
}
\details{
the dimension of the returned array is guessed from the supplied dis object. If there is only one unique k value in obj, 
 \code{sparse}, determines the dimensions of the returned array. If \code{sparse = TRUE}, a 2D array is returned. If \code{sparse = FALSE}, a 3D array is returned.
 When there is more than one unique k value in obj, a 3D array is always returned.
}
\seealso{
\code{\link{rmf_as_list}}
}
