% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rmf_as_array.sf}
\alias{rmf_as_array.sf}
\title{Title}
\usage{
\method{rmf_as_array}{sf}(obj, dis, select, k = NULL, prj = NULL,
  sparse = TRUE, op = sf::st_intersects, kper = attr(obj, "kper"),
  ...)
}
\arguments{
\item{obj}{}

\item{dis}{}

\item{select}{}

\item{k}{}

\item{prj}{}

\item{sparse}{}

\item{op}{}

\item{kper}{}

\item{...}{}
}
\description{
Title
}
