% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rmf_as_array.stars}
\alias{rmf_as_array.stars}
\title{Title}
\usage{
\method{rmf_as_array}{stars}(obj, dis, prj = NULL, crs = NULL,
  select = 1, resample = TRUE, method = "bilinear",
  kper = attr(obj, "kper"))
}
\arguments{
\item{obj}{}

\item{dis}{}

\item{prj}{}

\item{crs}{}

\item{select}{}

\item{resample}{}

\item{method}{}

\item{kper}{}
}
\description{
Title
}
