% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_as_list}
\alias{rmf_as_list}
\alias{rmf_as_list.rmf_3d_array}
\alias{rmf_as_list.rmf_2d_array}
\alias{rmf_as_list.rmf_4d_array}
\title{Convert objects to a rmf_list}
\usage{
rmf_as_list(...)

\method{rmf_as_list}{rmf_3d_array}(obj, dis, name = "value", mask = 1 +
  obj * 0, ijk = NULL, kper = attr(obj, "kper"))

\method{rmf_as_list}{rmf_2d_array}(obj, ...)

\method{rmf_as_list}{rmf_4d_array}(obj, l, ...)
}
\arguments{
\item{obj}{a \code{rmf_2d_array} or \code{rmf_3d_array}}

\item{dis}{a \code{RMODFLOW dis} object; only used when ijk is suppplied}

\item{name}{character; name of the resulting column which holds the extracted values from the array; defaults to "value"}

\item{mask}{a logical array with the same dimensions as obj. Used to select cells from obj; defaults to all cells}

\item{ijk}{optional; a data.frame with i, j and k columns used to select the cells. Overwrites the use of mask.}

\item{kper}{sets the kper attribute of the returned list; defaults to the kper attribute of obj}

\item{l}{integer index used to subset the 4th dimension of a \code{rmf_4d_array}.}
}
\value{
a \code{rmf_list}
}
\description{
Convert objects to a rmf_list

Converts a rmf_2d_array or rmf_3d_array to a rmf_list
}
\details{
\code{\link{rmf_convert_id_to_ijk}} can be used to obtain a ijk data.frame from either MODFLOW cell id's or R cell id's.

will set k to 1
}
\seealso{
\code{\link{rmf_as_array}}
}
