% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_as_sf.rmf_2d_array}
\alias{rmf_as_sf.rmf_2d_array}
\title{Title}
\usage{
\method{rmf_as_sf}{rmf_2d_array}(array, dis, mask = array * 0 + 1,
  prj = NULL, crs = NULL)
}
\arguments{
\item{array}{}

\item{dis}{}

\item{mask}{}

\item{prj}{}

\item{crs}{}
}
\description{
Title
}
