% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rmf_as_sf.rmf_list}
\alias{rmf_as_sf.rmf_list}
\title{Title}
\usage{
\method{rmf_as_sf}{rmf_list}(obj, dis, prj = NULL, crs = NULL,
  as_points = FALSE)
}
\arguments{
\item{obj}{}

\item{dis}{}

\item{prj}{}

\item{crs}{}

\item{as_points}{}
}
\description{
Title
}
