% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rmf_as_stars}
\alias{rmf_as_stars}
\alias{rmf_as_stars.rmf_2d_array}
\alias{rmf_as_stars.rmf_3d_array}
\alias{rmf_as_stars.rmf_4d_array}
\alias{rmf_as_stars.rmf_list}
\title{Functions to convert rmf_array and rmf_list objects to stars objects}
\usage{
rmf_as_stars(...)

\method{rmf_as_stars}{rmf_2d_array}(
  array,
  dis,
  mask = array * 0 + 1,
  prj = rmf_get_prj(dis),
  name = "value",
  id = "r",
  ...
)

\method{rmf_as_stars}{rmf_3d_array}(
  array,
  dis,
  mask = array * 0 + 1,
  prj = rmf_get_prj(dis),
  name = "value",
  id = "r",
  ...
)

\method{rmf_as_stars}{rmf_4d_array}(
  array,
  dis,
  mask = array(1, dim = dim(array)[1:3]),
  prj = rmf_get_prj(dis),
  name = "value",
  id = "r",
  ...
)

\method{rmf_as_stars}{rmf_list}(
  obj,
  dis,
  select,
  prj = rmf_get_prj(dis),
  name = "value",
  id = "r",
  ...
)
}
\arguments{
\item{...}{additional arguments passed to \code{rmf_as_array} when converting a \code{rmf_list} object. Otherwise, ignored.}

\item{array}{\code{rmf_2d_array}, \code{rmf_3d_array} or \code{rmf_4d_array} object}

\item{dis}{\code{RMODFLOW} dis object}

\item{mask}{a 2d array when \code{array} is 2d or a 3d array when \code{array} is 3d or 4d that can be coerced to logical. Used to specify which cells to convert. Defaults to all cells.}

\item{prj}{\code{RMODFLOW} prj object}

\item{name}{character specifying the name of the resulting variable in the stars object. Defaults to \code{'value'}}

\item{id}{either \code{'r'} (default) or \code{'modflow'}. Specifies which type of cell id is returned. R uses column-major array ordering whereas MODFLOW uses row-major ordering.}

\item{obj}{\code{rmf_list} object}

\item{select}{integer or character specifying which column of the \code{rmf_list} object to convert to \code{stars} variable.}
}
\value{
a \code{stars} object with x and y dimensions when \code{array} is 2d, x, y and layer (integer representing MODFLOW layer; similar to bands) when \code{array} is 3d,
x, y, layer and time dimensions when \code{array} is 4d. When converting a \code{rmf_list} object, it is first converted to a \code{rmf_array} using \code{rmf_as_array}.
Two variables are present in the returned \code{stars} object, one with the array values and on with the cell id (when \code{id} is \code{'r'} or \code{'modflow'}).
}
\description{
Functions to convert rmf_array and rmf_list objects to stars objects
}
\details{
The crs is taken from the \code{prj} argument.
}
\examples{
dis <- rmf_create_dis()

# 2d array
r <- rmf_create_array(1:prod(dis$nrow, dis$ncol), dim = c(dis$nrow, dis$ncol))
rmf_as_stars(r, dis = dis)

# 3d array
r <- rmf_create_array(1:prod(dis$nrow, dis$ncol, dis$nlay), dim = c(dis$nrow, dis$ncol, dis$nlay))
rmf_as_stars(r, dis = dis, id = 'modflow')

# 4d array
r <- rmf_create_array(1:prod(dis$nrow, dis$ncol, dis$nlay, 2), dim = c(dis$nrow, dis$ncol, dis$nlay, 2))
rmf_as_stars(r, dis = dis, id = FALSE)

# rmf_list
l <- rmf_create_list(data.frame(i = 1, j = 1:2, k = c(3, 2), q = c(-500, -400), d = 35))
rmf_as_stars(l, dis = dis, select = 'q')

}
