% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rmf_as_stars.rmf_2d_array}
\alias{rmf_as_stars.rmf_2d_array}
\title{Title}
\usage{
\method{rmf_as_stars}{rmf_2d_array}(array, dis, mask = array * 0 + 1,
  prj = NULL, crs = NULL, name = "value")
}
\arguments{
\item{array}{}

\item{dis}{}

\item{mask}{}

\item{prj}{}

\item{crs}{}

\item{name}{}
}
\description{
Title
}
