% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-as-tibble.rmf-3d-array.R
\name{rmf_as_tibble.rmf_3d_array}
\alias{rmf_as_tibble.rmf_3d_array}
\title{Title}
\usage{
\method{rmf_as_tibble}{rmf_3d_array}(array, i = NULL, j = NULL,
  k = NULL, dis, mask = array * 0 + 1, prj = NULL, crs = NULL)
}
\arguments{
\item{array}{}

\item{i}{}

\item{j}{}

\item{k}{}

\item{dis}{}

\item{mask}{}

\item{prj}{}

\item{crs}{}
}
\description{
Title
}
