% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{rmf_browse}
\alias{rmf_browse}
\title{Browse the RMODFLOW pkgdown website}
\usage{
rmf_browse()
}
\description{
This function can be used to launch the \href{https://rogiersbart.github.io/RMODFLOW/}{RMODFLOW pkgdown website} from the R console.
}
\examples{
rmf_browse()
}
\seealso{
\code{\link[=rmf_guide]{rmf_guide()}} for browsing the online guide pages.
}
