% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-calculate-array.R
\name{rmf_calculate_array}
\alias{rmf_calculate_array}
\title{Calculate a \code{rmf_2d_array} from multiplier arrays, zone arrays and/or parameter values}
\usage{
rmf_calculate_array(dis, mltarr, zonarr, iz = NULL,
  parameter_value = 1)
}
\arguments{
\item{dis}{dis object; used to set the dimensions of the array}

\item{mltarr}{either a list of multiplier arrays or a single multiplier array. The keyword \code{"NONE"} indicates no multiplier array is present.}

\item{zonarr}{either a list of the zone arrays or a single zone array. The keyword \code{"ALL"} indicates no zone array is present.}

\item{iz}{only read when zonarr is not \code{"ALL"}. A list where each element is a numeric vector with the zone numbers for the corresponding zone array.}

\item{parameter_value}{vector of parameter values corresponding to mltarr and/or zonarr which are multiplied with the final array. Typically used when the array represents a parameter.}
}
\value{
\code{rmf_2d_array} with the values calculated from the multiplier and/or zone arrays.
}
\description{
Given a multiplier array and/or zone array with corresponding zone numbers, calculate a \code{rmf_2d_array}. Parameter values can be used to multiply the arrays as well.
}
\details{
if mltarr (zonarr) is a list, certain elements are allowed to be set to \code{"NONE"} (\code{"ALL"}) indicating the multiplier (zone) array is not active for this cluster.
         if mltarr is \code{"NULL"} and zonarr is \code{"ALL"}, all the cells in the returned array will be equal to the parameter value.
         Multiple multiplier arrays, zone arrays and/or parameter values can be used to calculate the final values. Cells that correspond to multiple multiplier arrays, zone arrays and/or parameter values will be summed.
}
