% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-cell-dimensions.R,
%   R/rmf-cell-dimensions.dis.R, R/rmf-cell-dimensions.huf.R
\name{rmf_cell_dimensions}
\alias{rmf_cell_dimensions}
\alias{rmf_cell_dimensions.dis}
\alias{rmf_cell_dimensions.huf}
\title{Generic function to get cell dimensions}
\usage{
rmf_cell_dimensions(...)

\method{rmf_cell_dimensions}{dis}(dis, hed = NULL,
  include_volume = FALSE, include_faces = FALSE)

\method{rmf_cell_dimensions}{huf}(huf, dis = NULL, hed = NULL,
  include_volume = FALSE, include_faces = FALSE)
}
\arguments{
\item{dis}{dis object}

\item{hed}{hed object, used for calculating the saturated thickness; if not specified, the regular cell thickness is returned}

\item{include_volume}{logical; should the cell volumes be included?}

\item{include_faces}{logical; should face areas be included?}

\item{huf}{huf object}

\item{hed}{hed object, used for calculating the saturated thickness; if not specified, the regular cell thickness is returned}
}
\value{
list with cell dimension 3d arrays

list with cell dimension 3d arrays
}
\description{
Generic function to get cell dimensions

Get cell dimensions from a dis object

Get cell dimensions from a huf object
}
