% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_convert_cbc_to_darcy}
\alias{rmf_convert_cbc_to_darcy}
\title{Convert cbc object fluxes to darcy velocities}
\usage{
rmf_convert_cbc_to_darcy(cbc, dis, hed = NULL, porosity = NULL)
}
\arguments{
\item{cbc}{\code{RMODFLOW} cbc object}

\item{dis}{\code{RMODFLOW} dis object}

\item{hed}{\code{RMODFLOW} hed object; optional; if specified, the saturated cell thickness is used}

\item{porosity}{optional 3d array with porosity values. If used, the average linear groundwater flow velocities are returned.}
}
\value{
list of 4d arrays: right, front, lower, left, back, upper, qx, qy, qz and q; all represent Darcy velocities (or average linear groundwater flow velocities if porosity is specified): the first six at the different cell faces, the last four represent the components and magnitude at the cell center
}
\description{
Convert cbc object fluxes to darcy velocities
}
