% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-convert-grid-to-xyz.R
\name{rmf_convert_grid_to_xyz}
\alias{rmf_convert_grid_to_xyz}
\alias{convert_grid_to_xyz}
\title{Convert modflow coordinates to real world coordinates}
\usage{
rmf_convert_grid_to_xyz(x = NULL, y = NULL, z = NULL, i = NULL,
  j = NULL, k = NULL, prj, dis = NULL)

convert_grid_to_xyz(...)
}
\arguments{
\item{x}{modflow x coordinate}

\item{y}{modflow y coordinate}

\item{z}{modflow z coordinate}

\item{i}{modflow row number}

\item{j}{modflow column number}

\item{k}{modflow layer number}

\item{prj}{prj object}

\item{dis}{dis object}
}
\value{
data frame with real world x and y coordinates
}
\description{
Convert modflow coordinates to real world coordinates
}
\details{
Provide either xyz or ijk
}
\section{Functions}{
\itemize{
\item \code{convert_grid_to_xyz}: Deprecated function name
}}

