% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_convert_hob_to_locations}
\alias{rmf_convert_hob_to_locations}
\title{Convert a hob object to a locations data frame}
\usage{
rmf_convert_hob_to_locations(hob, dis, prj = rmf_get_prj(dis))
}
\arguments{
\item{hob}{\code{RMODFLOW} hob object}

\item{dis}{\code{RMODFLOW} dis object}

\item{prj}{prj object}
}
\value{
a data frame containing name, screened layer proportion pr, cell indices k, i, j and x, y & z coordinates
}
\description{
Convert a hob object to a locations data frame
}
\details{
returned z coordinate represents the center of the cell, not the top or bottom of the well screen
}
\seealso{
\code{\link{rmf_create_hob}}, \code{\link{rmf_convert_hob_to_time_series}}
}
