% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_convert_hob_to_time_series}
\alias{rmf_convert_hob_to_time_series}
\title{Convert a hob object to a time series data frame}
\usage{
rmf_convert_hob_to_time_series(hob, dis, starttime = dis$starttime)
}
\arguments{
\item{hob}{\code{RMODFLOW} hob object}

\item{dis}{\code{RMODFLOW} dis object}
}
\value{
time series data frame containing name, time and head columns
}
\description{
Convert a hob object to a time series data frame
}
\seealso{
\code{\link{rmf_create_hob}}, \code{\link{rmf_convert_hob_to_locations}}
}
