% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-convert-huf-to-dis.R
\name{rmf_convert_huf_to_dis}
\alias{rmf_convert_huf_to_dis}
\alias{convert_huf_to_dis}
\title{Convert a huf to a dis object (for plotting)}
\usage{
rmf_convert_huf_to_dis(huf, dis)

convert_huf_to_dis(...)
}
\arguments{
\item{huf}{huf object}

\item{dis}{dis object, corresponding to the huf object}
}
\value{
dis object containing the layers of the huf object
}
\description{
Convert a huf to a dis object (for plotting)
}
\section{Functions}{
\itemize{
\item \code{convert_huf_to_dis}: Deprecated function name
}}

