% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_convert_huf_to_lpf}
\alias{rmf_convert_huf_to_lpf}
\title{Convert a huf to a lpf object}
\usage{
rmf_convert_huf_to_lpf(huf, dis, mask = NULL,
  vka = ifelse(all(huf$hguvani == 0), "VK", "VANI"), averaging = c(HK =
  "arithmetic", HANI = "arithmetic", VK = "harmonic", VANI = "arithmetic",
  SS = "arithmetic", SY = "arithmetic"), pvl = NULL, ...)
}
\arguments{
\item{huf}{\code{RMODFLOW} huf object}

\item{dis}{\code{RMODFLOW} dis object}

\item{mask}{masking 3d array for averaging \code{\link{rmf_convert_huf_to_grid}}, typically the \code{ibound} array, to speed up grid conversion; defaults to including all cells}

\item{vka}{character indicating what variable the VKA array in the resulting lpf object represents. Possible values are \code{'VK'} or \code{'VANI'}. If all HGUVANI values are the same, the default vka is set correspondingly. If HGUVANI varies between hgu's, the default vka is \code{'VANI'}.}

\item{averaging}{named character vector of weighted averaging to use in \code{\link{rmf_convert_huf_to_grid}}. Possible values are 'arithmetic', 'harmonic' and 'geometric'. Names should correspond to the \code{partyp} defined in the huf object. Defaults to 'arithmetic' for every parameter type except for 'VK' ('harmonic').}

\item{pvl}{optional \code{RMODFLOW} pvl object; used to overwrite huf parameter values in \code{\link{rmf_convert_huf_to_grid}}. Defaults to NULL.}

\item{...}{arguments passed to \code{\link{rmf_create_lpf}}}
}
\value{
a \code{RMODFLOW} lpf object
}
\description{
Convert a huf to a lpf object
}
\details{
Huf parameters are converted to non-parameter data averaged over the \code{dis} grid using \code{\link{rmf_convert_huf_to_grid}}.
The resulting lpf object will therefore not have any flow parameters.
If HGUVANI varies per hgu, a correct conversion to a lpf vka array is not possible (i.e. because part of the layer might represent VANI from hgu A and another part VK from hgu B). Therefore, the user has to decide whether vka represents vk or vani in its entirety.
}
\seealso{
\code{\link{rmf_convert_huf_to_grid}}, \code{\link{rmf_create_lpf}}
}
