% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-convert-huf-to-nlay.R
\name{rmf_convert_huf_to_nlay}
\alias{rmf_convert_huf_to_nlay}
\alias{convert_huf_to_nlay}
\title{Convert a huf object to an rmf_3d_array with the number of numerical layers per hydrogeological unit}
\usage{
rmf_convert_huf_to_nlay(huf, dis, bas)

convert_huf_to_nlay(...)
}
\arguments{
\item{huf}{huf object}

\item{dis}{dis object, corresponding to the huf object}

\item{bas}{bas object, corresponding to the huf object}
}
\value{
nlay rmf_3d_array
}
\description{
Convert a huf object to an rmf_3d_array with the number of numerical layers per hydrogeological unit
}
\section{Functions}{
\itemize{
\item \code{convert_huf_to_nlay}: Deprecated function name
}}

