% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_convert_id_to_id}
\alias{rmf_convert_id_to_id}
\title{Convert id to id}
\usage{
rmf_convert_id_to_id(id, dis, from = "modflow", to = "r")
}
\arguments{
\item{id}{cell id, providing the place of the number in an input file 2d or 3d array}

\item{dis}{a discretisation file object}

\item{from}{'r' or 'modflow'. The type of id to convert from. Defaults to 'modflow'}

\item{to}{'r' or 'modflow'. The type of id to convert to. Defaults to 'r'}
}
\description{
Convert id to id
}
\details{
a modflow id provides the place of the number in an input file 3d array (not like the way R uses ids for arrays or matrices; rows and columns are switched)
}
