% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-convert-ijk-to-id.R
\name{rmf_convert_ijk_to_id}
\alias{rmf_convert_ijk_to_id}
\alias{convert_ijk_to_id}
\title{Convert ijk to id}
\usage{
rmf_convert_ijk_to_id(i, j, k, dis, type = "r")

convert_ijk_to_id(...)
}
\arguments{
\item{i}{vector of row numbers}

\item{j}{vector of column numbers}

\item{k}{vector of layer numbers}

\item{dis}{a discretization file object}
}
\value{
cell ids, providing the place of the cell in an input file 3d array
}
\description{
Convert ijk to id
}
\details{
a modflow id provides the place of the number in an input file 3d array (not like the way R uses ids for arrays or matrices; rows and columns are switched)
}
\section{Functions}{
\itemize{
\item \code{convert_ijk_to_id}: Deprecated function name
}}

