% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_convert_lpf_to_upw}
\alias{rmf_convert_lpf_to_upw}
\title{Convert a lpf to a upw object}
\usage{
rmf_convert_lpf_to_upw(lpf, iphdry = TRUE)
}
\arguments{
\item{lpf}{\code{RMODFLOW} lpf object}

\item{iphdry}{logical; indicating if head will be set to hdry when it's less than 1E-4 above the cell bottom; defaults to TRUE}
}
\value{
Object of class upw
}
\description{
Convert a lpf to a upw object
}
\note{
upw input structure is nearly identical to lpf but calculations are done differently. Differences include the addition of the iphdry value and the ommision of optional keywords. Layer wetting capabilities are also not supported by upw.

upw must be used with the Newton solver. See also \code{\link{rmf_create_nwt}}.
}
\seealso{
\code{\link{rmf_create_upw}}, \code{\link{rmf_convert_upw_to_lpf}} and \url{https://water.usgs.gov/ogw/modflow-nwt/MODFLOW-NWT-Guide/}
}
