% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_convert_upw_to_lpf}
\alias{rmf_convert_upw_to_lpf}
\title{Convert a upw to a lpf object}
\usage{
rmf_convert_upw_to_lpf(
  upw,
  storagecoefficient = FALSE,
  constantcv = FALSE,
  thickstrt = FALSE,
  nocvcorrection = FALSE,
  novfc = FALSE,
  noparcheck = FALSE,
  laywet = upw$laywet,
  wetfct = 0.1,
  iwetit = 1,
  ihdwet = 0,
  wetdry = NULL
)
}
\arguments{
\item{upw}{\code{RMODFLOW} upw object}

\item{storagecoefficient}{logical; should STORAGECOEFFICIENT keyword be included?; defaults to FALSE}

\item{constantcv}{logical; should CONSTANTCV keyword be included?; defaults to FALSE}

\item{thickstrt}{logical; should THICKSTRT keyword be included?; defaults to FALSE}

\item{nocvcorrection}{logical; should NOCVCORRECTION keyword be included?; defaults to FALSE}

\item{novfc}{logical; should NOVFC keyword be included?; defaults to FALSE}

\item{noparcheck}{logical; should NOPARCHECK keyword be included?; defaults to FALSE}

\item{laywet}{vector of flags for each layer, indicating if wetting is active; defaults to 0 for each layer}

\item{wetfct}{is a factor that is included in the calculation of the head that is initially established at a cell when it is converted from dry to wet; defaults to 0.1}

\item{iwetit}{is the iteration interval for attempting to wet cells; defaults to 1}

\item{ihdwet}{is a flag that determines which equation is used to define the initial head at cells that become wet; defaults to 0}

\item{wetdry}{3d array with a wetting threshold and flag indicating which neighboring cells can cause a cell to become wet; defaults to NULL. If not read for a specific layer, set all values in that layer to NA.}
}
\value{
object of class lpf
}
\description{
Convert a upw to a lpf object
}
\note{
upw input structure is nearly identical to lpf but calculations are done differently. Differences include the addition of the iphdry value and the omission of optional keywords. Layer wetting capabilities are also not supported by upw.
}
\seealso{
\code{\link{rmf_create_lpf}}, \code{\link{rmf_convert_lpf_to_upw}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?lpf.htm}
}
