% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-copy-to-wd.R
\name{rmf_copy_to_wd}
\alias{rmf_copy_to_wd}
\title{Copy files from specified paths to current working directory}
\usage{
rmf_copy_to_wd(filenames, ...)
}
\arguments{
\item{filenames}{character vector of filenames}

\item{...}{additional arguments provided to file.copy}
}
\description{
Copy files from specified paths to current working directory
}
