% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-create-bas.R
\name{rmf_create_bas}
\alias{rmf_create_bas}
\alias{create_bas}
\title{Create an \code{RMODFLOW} bas object}
\usage{
rmf_create_bas(dis = rmf_create_dis(), xsection = FALSE,
  chtoch = FALSE, free = TRUE, printtime = FALSE,
  showprogress = FALSE, stoperror = FALSE, stoper = 1,
  ibound = rmf_create_array(1L, dim = c(dis$nrow, dis$ncol, dis$nlay)),
  hnoflo = -999, strt = rmf_create_array(0, dim = c(dis$nrow, dis$ncol,
  dis$nlay)))

create_bas(...)
}
\arguments{
\item{dis}{RMODFLOW dis object}

\item{xsection}{logical; is the model a 1-row cross-section?}

\item{chtoch}{logical; should flow between adjacent constant-head cells be calculated?}

\item{free}{logical; is free format used?}

\item{printtime}{logical; should start, end and elapsed times be written to the global output file?}

\item{showprogress}{logical; should progress information be displayed?}

\item{stoperror}{logical; should the model be stopped based on budget percent discrepancy?}

\item{stoper}{numeric; threshold budget percent discrepancy}

\item{ibound}{3d array specifying active (1), inactive (0) or constant head (0) status of all cells}

\item{strt}{3d array specifying starting heads}
}
\value{
Object of class bas
}
\description{
\code{rmf_create_bas} creates an \code{RMODFLOW} bas object.
}
\section{Functions}{
\itemize{
\item \code{create_bas}: Deprecated function name
}}

\seealso{
\code{\link{rmf_read_bas}}, \code{\link{rmf_write_bas}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?bas.htm}
}
