% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-create-chd.R
\name{rmf_create_chd}
\alias{rmf_create_chd}
\title{Create an \code{RMODFLOW} chd object.}
\usage{
rmf_create_chd(npchd = NULL, mxl = NULL, mxactc = 10, option = NULL,
  parnam = NULL, parval = NULL, nlst = NULL, instances = NULL,
  numinst = NULL, instnam = NULL, layer_parm = NULL, row_parm = NULL,
  column_parm = NULL, shdfact_parm = NULL, ehdfact_parm = NULL,
  xyz_parm = NULL, itmp = 10, np = 0, layer_sp = list(rep(1, 10)),
  row_sp = list(1:10), column_sp = list(rep(1, 10)),
  shead_sp = list(rep(0, 10)), ehead_sp = list(rep(-1, 10)),
  xyz_sp = NULL, pname = NULL, iname = NULL)
}
\arguments{
\item{npchd}{number of chd parameters; defaults to NULL}

\item{mxl}{maximum number of constant-head boundaries that will be defined using parameters; defaults to NULL}

\item{mxactc}{maximum number of constant-head boundaries in use during any stress period defined with and without parameters; defaults to 10}

\item{option}{optional character vector specifying the auxiliary variables and/or \code{NOPRINT}. Do not include the keywords \code{AUXILIARY} or \code{AUX}; defaults to NULL}

\item{parnam}{vector of length \code{npchd} specifying the parameter names; defaults to NULL}

\item{parval}{vector of length \code{npchd} specifying the parameter values; defaults to NULL}

\item{nlst}{vector of length \code{npchd} specifying the number of constant-head cells that are included in a non-time-varying parameter or in each instance of a time-varying parameter; defaults to NULL}

\item{instances}{logical vector of length \code{npchd} indicating which parameters are time-varying; defaults to NULL}

\item{numinst}{vector of length \code{npchd} indicating the number instances that are included in the time-varying parameter; defaults to NULL}

\item{instnam}{list with \code{npchd} elements where each element \code{i} is a character vector of length \code{numinst} for parameter \code{i} specifying the names of the parameter instances. If not time-varying, set numinst dimension to 1. Defaults to NULL}

\item{layer_parm}{list with \code{npchd} elements where each element \code{i} is a numeric 2D array with dimensions \code{numinst x nlst} specifying the layer numbers of the model cells containing the constant-head boundaries specified by parameter \code{i}. If not time-varying, set numinst dimension to 1. Defaults to NULL}

\item{row_parm}{list with \code{npchd} elements where each element \code{i} is a numeric 2D array with dimensions \code{numinst x nlst} specifying the row numbers of the model cells containing the constant-head boundaries specified by parameter \code{i}. If not time-varying, set numinst dimension to 1. Defaults to NULL}

\item{column_parm}{list with \code{npchd} elements where each element \code{i} is a numeric 2D array with dimensions \code{numinst x nlst} specifying the column numbers of the model cells containing the constant-head boundaries specified by parameter \code{i}. If not time-varying, set numinst dimension to 1. Defaults to NULL}

\item{shdfact_parm}{list with \code{npchd} elements where each element \code{i} is a numeric 2D array with dimensions \code{numinst x nlst} specifying a factor used to calculate the head at the boundary at the start of the stress period for parameter \code{i}. The starting head at the boundary is the product of shdfact and the parameter value. If not time-varying, set numinst dimension to 1. Defaults to NULL}

\item{xyz_parm}{optional list with \code{npchd} elements where each element \code{i} is a character 2D array with dimensions \code{numinst x nlst} specifying the auxiliary variables for parameter \code{i}. If not time-varying, set numinst to 1. Defaults to NULL}

\item{itmp}{vector of length \code{dis$nper} specifying non-parameter constant-head data being read for each stress period; defaults to 10}

\item{np}{vector of length \code{dis$nper} specifying the number of constant-head parameters in use for each stress period; defaults to 0}

\item{layer_sp}{list with \code{dis$nper} elements where each element \code{i} is a numeric vector of length \code{itmp} for stress period \code{i} specifying the layer numbers of the model cells containing the constant-head boundaries; defaults to 1 for all constant-head boundaries}

\item{row_sp}{list with \code{dis$nper} elements where each element \code{i} is a numeric vector of length \code{itmp} for stress period \code{i} specifying the row numbers of the model cells containing the constant-head boundaries; defaults to 1:10}

\item{column_sp}{list with \code{dis$nper} elements where each element \code{i} is a numeric vector of length \code{itmp} for stress period \code{i} specifying the column numbers of the model cells containing the constant-head boundaries; defaults to 1 for all constant-head boundaries}

\item{shead_sp}{list with \code{dis$nper} elements where each element \code{i} is a numeric vector of length \code{itmp} for stress period \code{i} specifying the starting heads at the constant-head cells; defaults to 0 for all constant-head boundaries}

\item{ehead_sp}{list with \code{dis$nper} elements where each element \code{i} is a numeric vector of length \code{itmp} for stress period \code{i} specifying the ending heads at the constant-head cells; defaults to -1 for all constant-head boundaries}

\item{xyz_sp}{optional list with \code{dis$nper} elements where each element \code{i} is a character vector of length \code{itmp} for stress period \code{i} specifying the auxiliary variables of the model cells containing the constant-head boundaries; defaults to NULL}

\item{pname}{list with \code{dis$nper} elements where each element \code{i} is a character vector of length \code{np} for stress period \code{i} specifying the names of the parameters being used; defaults to NULL}

\item{iname}{list with \code{dis$nper} elements where each element \code{i} is a character vector of length \code{np} for stress period \code{i} specifying the names of the parameter instances being used; defaults to NULL}

\item{edhfact_parm}{list with \code{npchd} elements where each element \code{i} is a numeric 2D array with dimensions \code{numinst x nlst} specifying a factor used to calculate the head at the boundary at the end of the stress period for parameter \code{i}. The ending head at the boundary is the product of ehdfact and the parameter value. If not time-varying, set numinst dimension to 1. Defaults to NULL}
}
\value{
\code{RMODFLOW} chd object
}
\description{
\code{rmf_create_chd} creates an \code{RMODFLOW} chd object
}
\seealso{
\code{\link{rmf_read_chd}}, \code{\link{rmf_write_chd}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?chd.htm}
}
