% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-dis.R
\name{rmf_create_dis}
\alias{rmf_create_dis}
\alias{create_dis}
\title{Create an \code{RMODFLOW} dis object}
\usage{
rmf_create_dis(nlay = 3, nrow = 10, ncol = 10, nper = 1,
  itmuni = 1, lenuni = 2, laycbd = rep(0, nlay), delr = 100,
  delc = 100, top = matrix(0, nrow = nrow, ncol = ncol),
  botm = array(rep(seq(0, -10 * (nlay + sum(laycbd != 0)), length = nlay
  + sum(laycbd != 0) + 1)[2:(nlay + sum(laycbd != 0) + 1)], each = nrow *
  ncol), dim = c(nrow, ncol, nlay + sum(laycbd != 0))), perlen = rep(1,
  nper), nstp = rep(1, nper), tsmult = rep(1, nper), sstr = c("SS",
  rep("TR", nper - 1)))

create_dis(...)
}
\arguments{
\item{nlay}{number of layers; defaults to 3}

\item{nrow}{number of rows; defaults to 10}

\item{ncol}{number of columns; defaults to 10}

\item{nper}{number of stress periods; defaults to 1}

\item{itmuni}{time unit; defaults to 1 (seconds)}

\item{lenuni}{length unit; defaults to 2 (metres)}

\item{laycbd}{vector of Quasi-3D confining bed flags; defaults to 0 for all layers}

\item{delr}{vector of cell widths along rows; defaults to 100 for all columns}

\item{delc}{vector of cell widths along columns; defaults to 100 for all rows}

\item{top}{matrix with the top elevation of layer 1; defaults to 0 for all nrow x ncol cells}

\item{botm}{3D array with the bottom elevations of all layers and Quasi-3D confining beds; defaults to nlay layers, equally spaced between 0 (top first layer) and -100 (bottom last layer)}

\item{perlen}{vector of stress period lengths}

\item{nstp}{vector of stress period time steps}

\item{tsmult}{vector of successive time step length multipliers}

\item{sstr}{character vector with steady state ('SS') or transient ('TR') stress period indicator}
}
\value{
Object of class dis
}
\description{
\code{rmf_create_dis} creates an \code{RMODFLOW} dis object.
}
\section{Functions}{
\itemize{
\item \code{create_dis}: Deprecated function name
}}

\seealso{
\code{\link{rmf_read_dis}}, \code{\link{rmf_write_dis}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?dis.htm}
}
