% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-drn.R
\name{rmf_create_drn}
\alias{rmf_create_drn}
\title{Create an \code{RMODFLOW} drn object.}
\usage{
rmf_create_drn(..., dis, idrncb = 0, noprint = FALSE, aux = NULL)
}
\arguments{
\item{...}{\code{rmf_list} (possibly of class \code{rmf_parm}) objects or a single \code{list} with \code{rmf_list} objects (possibly of class \code{rmf_parm}) elements; defines the drains.}

\item{dis}{dis object}

\item{idrncb}{flag and unit number for writing cell-by-cell flow terms; defaults to 0 (cell-by-cell flow terms will not be written)}

\item{noprint}{logical, should the printing of DRN cells to the listing file be suppressed ? Defaults to \code{FALSE}}

\item{aux}{optional character vector specifying the names of the auxiliary variables. These variables should also be included in the \code{rmf_list} objects that are supplied; defaults to \code{NULL}}
}
\value{
\code{RMODFLOW} drn object
}
\description{
\code{rmf_create_drn} creates an \code{RMODFLOW} drn object
}
\seealso{
\code{\link{rmf_read_drn}}, \code{\link{rmf_write_drn}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?drn.htm}
}
