% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-create-flow-parameter.R
\name{rmf_create_flow_parameter}
\alias{rmf_create_flow_parameter}
\title{Create an input array flow parameter}
\usage{
rmf_create_flow_parameter(dis, name, value, type, layer,
  mlt_names = "NONE", zon_names = "ALL", iz = NULL, mlt = NULL,
  zon = NULL)
}
\arguments{
\item{dis}{\code{RMODFLOW} dis object; used to set the dimensions of the array}

\item{name}{character specifying the name of the parameter}

\item{value}{numeric specifying the value of the parameter which is used to multiply values in the array. Defaults to 1.0}

\item{type}{character specifying the parameter type. Possible values are \code{HK, HANI, VK, VANI, SS, SY and VKCB}.}

\item{layer}{integer vector specifying the the layer indices to which the parameter applies}

\item{mlt_names}{character vector specifying the names of multiplier arrays that are used to build the parameter. The keyword \code{"NONE"} indicates no multiplier array is present.}

\item{zon_names}{character vector specifying the names of zone arrays that are used to build the parameter. The keyword \code{"ALL"} indicates no multiplier array is present.}

\item{iz}{only read when the corresponding \code{zon_names} is not \code{"ALL"}. A list where each element is a numeric vector with the zone numbers for the corresponding zone array.}

\item{mlt}{\code{RMODFLOW} mlt object which holds the multiplier arrays specified in \code{mlt_names}}

\item{zon}{\code{RMODFLOW} zon object which holds the zone arrays specified in \code{zon_names}}

\item{kper}{integer vector with the stress period numbers during which the parameter is active}
}
\value{
an object of class \code{rmf_parm} and \code{rmf_2d_array}
}
\description{
Create a parameter for array input used in MODFLOW flow packages from multiplier and/or zone arrays
}
\details{
Typically, an array parameter is build from a single multiplier and/or zone array combination. However, multiple combinations can be used.
         If a parameter has multiple combinations of multiplier and/or zone arrays for a specific layer, the layer indices must be specified for each combination.
         If \code{mlt_names} is \code{"NONE"} and \code{zon_names} is \code{"ALL"}, \code{value} applies to all cells in the layer(s) and \code{mlt} and \code{zon} are not read.
}
\seealso{
\code{\link{rmf_create_array}}, \code{\link{rmf_create_array_parameter}}, \code{\link{rmf_create_list_parameter}}
}
