% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-create-ghb.R
\name{rmf_create_ghb}
\alias{rmf_create_ghb}
\title{Create an \code{RMODFLOW} ghb object.}
\usage{
rmf_create_ghb(npghb = 1, mxl = 10, mxactb = 10, ighbcb = 0,
  option = NULL, parnam = "GHB", parval = 1, nlst = 10,
  instances = NULL, numinst = NULL, instnam = NULL,
  layer_parm = list(array(rep(1, 10), dim = c(1, 10))),
  row_parm = list(array(1:10, dim = c(1, 10))),
  column_parm = list(array(rep(1, 10), dim = c(1, 10))),
  bhead_parm = list(array(5, dim = c(1, 10))),
  condfact_parm = list(array(10, dim = c(1, 10))), xyz_parm = NULL,
  itmp = 0, np = 1, layer_sp = NULL, row_sp = NULL, column_sp = NULL,
  bhead_sp = NULL, cond_sp = NULL, xyz_sp = NULL, pname = list("GHB"),
  iname = NULL)
}
\arguments{
\item{npghb}{number of ghb parameters; defaults to 1}

\item{mxl}{maximum number of general-head-boundary cells that will be defined using parameters; defaults to 10}

\item{mxactb}{maximum number of general-head-boundary cells in use during any stress period defined with and without parameters; defaults to 10}

\item{ighbcb}{flag and unit number for writing cell-by-cell flow terms; defaults to 0}

\item{option}{optional character vector specifying the auxiliary variables and/or \code{NOPRINT}. Do not include the keywords \code{AUXILIARY} or \code{AUX}; defaults to NULL}

\item{parnam}{vector of length \code{npghb} specifying the parameter names; defaults to 'GHB'}

\item{parval}{vector of length \code{npghb} specifying the parameter values; defaults to 1}

\item{nlst}{vector of length \code{npghb} specifying the number of general-head-boundary cells that are included in a non-time-varying parameter or in each instance of a time-varying parameter; defaults to 10}

\item{instances}{logical vector of length \code{npghb} indicating which parameters are time-varying; defaults to NULL}

\item{numinst}{vector of length \code{npghb} indicating the number instances that are included in the time-varying parameter; defaults to NULL}

\item{instnam}{list with \code{npghb} elements where each element \code{i} is a character vector of length \code{numinst} for parameter \code{i} specifying the names of the parameter instances. If not time-varying, set numinst dimension to 1. Defaults to NULL}

\item{layer_parm}{list with \code{npghb} elements where each element \code{i} is a numeric 2D array with dimensions \code{numinst x nlst} specifying the layer numbers of the model cells containing the general-head-boundaries specified by parameter \code{i}. If not time-varying, set numinst dimension to 1. Defaults to 1 for all general-head-boundary cells for the parameter.}

\item{row_parm}{list with \code{npghb} elements where each element \code{i} is a numeric 2D array with dimensions \code{numinst x nlst} specifying the row numbers of the model cells containing the general-head-boundaries specified by parameter \code{i}. If not time-varying, set numinst dimension to 1. Defaults to 1:10 for all general-head-boundary cells for the parameter}

\item{column_parm}{list with \code{npghb} elements where each element \code{i} is a numeric 2D array with dimensions \code{numinst x nlst} specifying the column numbers of the model cells containing the general-head-boundaries specified by parameter \code{i}. If not time-varying, set numinst dimension to 1. Defaults to 1 for the parameter}

\item{bhead_parm}{list with \code{npghb} elements where each element \code{i} is a numeric 2D array with dimensions \code{numinst x nlst} specifying the head on the boundary for parameter \code{i}. If not time-varying, set numinst dimension to 1. Defaults to 5 for all general-head-boundary cells for the parameter}

\item{condfact_parm}{list with \code{npghb} elements where each element \code{i} is a numeric 2D array with dimensions \code{numinst x nlst} specifying the conductance factor for parameter \code{i}. The conductance is the product of condfact and the parameter value. If not time-varying, set numinst dimension to 1. Defaults to 10 for all general-head-boundary cells for the parameter}

\item{xyz_parm}{optional list with \code{npghb} elements where each element \code{i} is a character 2D array with dimensions \code{numinst x nlst} specifying the auxiliary variables for parameter \code{i}. If not time-varying, set numinst to 1. Defaults to NULL}

\item{itmp}{vector of length \code{dis$nper} specifying non-parameter ghb data being read for each stress period; defaults to 0}

\item{np}{vector of length \code{dis$nper} specifying the number of ghb parameters in use for each stress period; defaults to 1}

\item{layer_sp}{list with \code{dis$nper} elements where each element \code{i} is a numeric vector of length \code{itmp} for stress period \code{i} specifying the layer numbers of the model cells containing the general-head-boundary cells; defaults to NULL}

\item{row_sp}{list with \code{dis$nper} elements where each element \code{i} is a numeric vector of length \code{itmp} for stress period \code{i} specifying the row numbers of the model cells containing the general-head-boundary cells; defaults to NULL}

\item{column_sp}{list with \code{dis$nper} elements where each element \code{i} is a numeric vector of length \code{itmp} for stress period \code{i} specifying the column numbers of the model cells containing the general-head-boundary cells; defaults to NULL}

\item{bhead_sp}{list with \code{dis$nper} elements where each element \code{i} is a numeric vector of length \code{itmp} for stress period \code{i} specifying the heads on the boundaries for model cells containing the general-head-boundary cells; defaults to NULL}

\item{cond_sp}{list with \code{dis$nper} elements where each element \code{i} is a numeric vector of length \code{itmp} for stress period \code{i} specifying the conductance of the model cells containing the general-head-boundary cells; defaults to NULL}

\item{xyz_sp}{optional list with \code{dis$nper} elements where each element \code{i} is a character vector of length \code{itmp} for stress period \code{i} specifying the auxiliary variables of the model cells containing the general-head-boundary cells; defaults to NULL}

\item{pname}{list with \code{dis$nper} elements where each element \code{i} is a character vector of length \code{np} for stress period \code{i} specifying the names of the parameters being used; defaults to 'GHB'}

\item{iname}{list with \code{dis$nper} elements where each element \code{i} is a character vector of length \code{np} for stress period \code{i} specifying the names of the parameter instances being used; defaults to NULL}
}
\value{
\code{RMODFLOW} ghb object
}
\description{
\code{rmf_create_ghb} creates an \code{RMODFLOW} ghb object
}
\seealso{
\code{\link{rmf_read_ghb}}, \code{\link{rmf_write_ghb}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?ghb.htm}
}
