% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-gmg.R
\name{rmf_create_gmg}
\alias{rmf_create_gmg}
\title{Create an \code{RMODFLOW} gmg object}
\usage{
rmf_create_gmg(rclose = 0.001, iiter = 100, hclose = 0.001,
  mxiter = 50, damp = 0.5, iadamp = 1, ioutgmg = 1, iunitmhc = 0,
  ism = 0, isc = 0, dup = 0.5, dlow = 0.01, chglimit = 0.001,
  relax = 1)
}
\arguments{
\item{rclose}{residual convergence criterion for the inner iteration; defaults to 0.001}

\item{iiter}{maximum number of PCG iterations for each linear solution; defaults to 100}

\item{hclose}{head-change convergence criterion for nonlinear problems; ignored of mxiter = 1. Defaults to 0.001}

\item{mxiter}{maximum number of outer iterations; defaults to 50}

\item{damp}{value of the damping parameter; defaults to 0.5}

\item{iadamp}{flag controlling adaptive damping. Either 0 (no damping), 1 (Cooley's method for first nonlinear iteration; default) or 2 (relative reduced residual damping)}

\item{ioutgmg}{flag controlling solver output. Either 0 (only solver input is printed), 1 (detailed solver output; default), 2 (basic solver output), 3 (same as 1 but printed to terminal) or 4 (same as 2 but printed to terminal)}

\item{iunitmhc}{flag and unit number for writing the maximum head change values. If > 0, these values are written to unit number iunitmhc which should be listed in the Name file. If <= 0 (default), no values are written.}

\item{ism}{type of smoother in preconditioner. Either 0 (ILU(0) smoothing; default) or 1 (Symmetric Gauss-Seidel smoothing)}

\item{isc}{flag controlling semi-coarsening in preconditioner. Either 0 (rows, columns & layers; default), 1 (rows & columns), 2 (columns & layers), 3 (rows & layers) or 4 (no coarsening)}

\item{dup}{maximum damping value that should be applied to any iteration when the solver is not oscillating. Only used when iadamp = 2; defaults to 0.5}

\item{chglimit}{maximum allowed head change at any cell between outer iterations. Only used when iadamp = 2; defaults to 0.001}

\item{relax}{relaxation parameter for the ILU PCG method. Defaults to 1. Only used when isc = 4.}

\item{dlowminimum}{damping value to be generated by the adaptive-damping procedure. Only used when iadamp = 2; defaults to 0.01}
}
\value{
Object of class gmg
}
\description{
\code{rmf_create_gmg} creates an \code{RMODFLOW} gmg object.
}
\seealso{
\code{\link{rmf_read_gmg}}, \code{\link{rmf_write_gmg}} and \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?gmg.htm}
}
