% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-create-hfb.R
\name{rmf_create_hfb}
\alias{rmf_create_hfb}
\title{Create an \code{RMODFLOW} hfb object.}
\usage{
rmf_create_hfb(..., dis, noprint = FALSE)
}
\arguments{
\item{...}{\code{rmf_list} (possibly of class \code{rmf_parm}) objects or a single \code{list} with \code{rmf_list} objects (possibly of class \code{rmf_parm}) elements; defines the horizontal-flow barriers.}

\item{dis}{dis object}

\item{noprint}{logical, should the printing of HFB cells to the listing file be suppressed ? Defaults to \code{FALSE}}
}
\value{
\code{RMODFLOW} hfb object
}
\description{
\code{rmf_create_hfb} creates an \code{RMODFLOW} hfb object
}
\details{
As an alternative to specifying \code{irow2} and \code{icol2}, a \code{direction} column can be present in the rmf_lists objects to specify the direction of the horizontal flow barrier with respect to \code{i & j}.
         Allowed values for the \code{direction} column are \code{"right"}, \code{"back"}, \code{"left"} and \code{"front"}.
}
\seealso{
\code{\link{rmf_read_hfb}}, \code{\link{rmf_write_hfb}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?hfb6.htm}
}
