% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-huf.R
\name{rmf_create_huf}
\alias{rmf_create_huf}
\title{Create an \code{RMODFLOW} huf object}
\usage{
rmf_create_huf(
  dis,
  ihufcb = 0,
  hdry = -888,
  nhuf,
  iohufheads = 0,
  iohufflows = 0,
  lthuf = rmfi_ifelse0(dis$nlay == 1, 1, c(1, rep(0, dis$nlay - 1))),
  laywt = rep(0, dis$nlay),
  wetfct = 0.1,
  iwetit = 1,
  ihdwet = 0,
  wetdry = NULL,
  hgunam,
  top,
  thck,
  hguhani = rep(1, nhuf),
  hguvani = rep(0, nhuf),
  parameters = NULL,
  hk = rep(1e-04, nhuf),
  hani = rep(1, nhuf),
  vk = hk,
  ss = rep(1e-05, nhuf),
  sy = rep(0.15, nhuf),
  sytp = NULL
)
}
\arguments{
\item{dis}{RMODFLOW dis object}

\item{ihufcb}{flag and unit number for writing cell-by-cell flow terms; defaults to 0}

\item{hdry}{head assigned to cells that are converted to dry cells; defaults to -888}

\item{nhuf}{number of hydrogeological units (hgu's)}

\item{iohufheads}{flag and unit number indicating whether interpolated heads should be written. Defaults to 0 (interpolated heads are not written). See details.}

\item{iohufflows}{flag and unit number indicating whether interpolated cell-by-cell flows should be written. Defaults to 0 (interpolated flows are not written). See details.}

\item{lthuf}{vector length \code{dis$nlay}, specifying layer type; defaults to all confined (0) except the first layer (1)}

\item{laywt}{vector length \code{dis$nlay},, indicating if wetting is active; defaults to 0 for each layer}

\item{wetfct}{is a factor that is included in the calculation of the head that is initially established at a cell when it is converted from dry to wet; defaults to 0.1}

\item{iwetit}{is the iteration interval for attempting to wet cells; defaults to 1}

\item{ihdwet}{is a flag that determines which equation is used to define the initial head at cells that become wet; defaults to 0}

\item{wetdry}{3d array with a wetting threshold and flag indicating which neighboring cells can cause a cell to become wet; defaults to NULL. If not read for a specific layer, set all values in that layer to NA.}

\item{hgunam}{character vector of length nhuf specifying the names of the hgu's}

\item{top}{3d array with the top elevations of each hgu (hgu is the third dimension)}

\item{thck}{3d array with the thicknesses of each hgu (hgu is the third dimension)}

\item{hguhani}{vector of flags indicating whether HANI parameters (0) or hguhani (> 0) is used to specify horizontal anisotropies for each hgu; defaults to 1 for each hgu}

\item{hguvani}{vector of flags indicating whether vk parameters will be specified (0) or HGUVANI is used for vertical anisotropy (> 0). VANI parameters may overwrite the the latter value; defaults to 0 for each hgu}

\item{parameters}{list of \code{rmf_parameter} as created by \code{\link{rmf_create_parameter}}. See details; defaults to NULL}

\item{hk}{(named) vector with hydraulic conductivity along rows; one value per hgu specified in the order of \code{hgunam}; defaults to 0.0001. See details.}

\item{hani}{(named) vector with the ratio of hydraulic conductivity along columns to that along rows; one value per hgu specified in the order of \code{hgunam}; defaults to 1. See details.}

\item{ss}{(named) vector with specific storage; only required when there are transient stress periods; one value per hgu specified in the order of \code{hgunam}; defaults to 1e-5. See details.}

\item{sy}{(named) vector with specific yield; only required when there are transient stress periods; one value per hgu specified in the order of \code{hgunam}; defaults to 0.15. See details.}

\item{sytp}{numeric value defining sytp; only required when there are transient stress periods and all layers are confined; defaults to NULL. See details.}

\item{vka}{(named) vector with vertical hydraulic conductivity or the ratio of horizontal to vertical conductivity depending on the value of hguvani; one value per hgu specified in the order of \code{hgunam}; defaults to NULL. See details.}
}
\value{
Object of class huf
}
\description{
\code{rmf_create_huf} creates an \code{RMODFLOW} huf object.
}
\details{
If iohufheads (iohufflows) > 0, heads (cell-by-cell flows)  will be interpolated to each hgu and saved, resulting in a single value per hgu at every row-column combination.
Flow variables are any of \code{HK, HANI, VK, VANI, SS, SY and SYTP}. Variables can be specified either through the use of parameters or direct vector input through hk, hani, vk, ss, sy and sytp arguments.
Specifying directly with parameters is useful if the parametrization is complex. If only a single value per hgu per partyp is needed, the direct vector input can be used.
If any parameters are specified, all flow variables have to be specified by parameters. If hk, hani, vk, ss, sy or sytp are not named, the hgu number will be added as a suffix to the type (e.g. 'HK_1').
Flow variable SYTP can not be specified per hgu, only for all hgu's. Using direct vector input, VANI can not be specified, only VK. VANI can be set through hguvani.

\if{html}{\out{<div class="sourceCode">}}\preformatted{     \code{\link{rmf_convert_huf_to_grid}} can be used to convert parameters defined on the HUF grid to the numerical grid
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link{rmf_create_parameter}}, \code{\link{rmf_convert_huf_to_grid}}, \code{\link{rmf_read_huf}}, \code{\link{rmf_write_huf}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?huf.htm}
}
