% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-huf.R
\name{rmf_create_kdep}
\alias{rmf_create_kdep}
\title{Create an \code{RMODFLOW} kdep object}
\usage{
rmf_create_kdep(parameters, rs = NULL, dis)
}
\arguments{
\item{parameters}{either a single \code{rmf_parameter} or a list of \code{rmf_parameters} specifying the depth-dependency coefficients. See details.}

\item{rs}{optional 2d array specifying the reference elevation surface}

\item{dis}{\code{RMODFLOW} dis object. Only used if \code{rs} is supplied.}
}
\value{
a \code{RMODFLOW} kdep object
}
\description{
\code{rmf_create_kdep} creates an \code{RMODFLOW} kdep object.
}
\details{
All parameters should have a hgunam attribute and their partyp attribute set to 'KDEP'.
The KDEP package can only be used in conjunction with the HUF package.
Note that the parameters are defined on the HUF grid, not the numerical grid
\code{\link{rmf_convert_huf_to_grid}} can be used to convert parameters defined on the HUF grid to the numerical grid
}
\seealso{
\code{link{rmf_convert_huf_to_grid}}, \code{\link{rmf_read_kdep}}, \code{\link{rmf_write_kdep}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?kdep.htm}
}
