% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-create-list-parameter.R
\name{rmf_create_list_parameter}
\alias{rmf_create_list_parameter}
\title{Create a List Data input parameter}
\usage{
rmf_create_list_parameter(rmf_list, parnam, parval = 1, instnam = NULL)
}
\arguments{
\item{rmf_list}{a rmf_list object}

\item{parnam}{character specifying the name of the parameter}

\item{parval}{numeric specifying the value of the parameter which is used to multiply the flux controlling variable in the data.frame. Defaults to 1.0}

\item{instnam}{optional character specying the instance name of the parameter is to be time-varying; defaults to NULL}
}
\value{
an object of class \code{rmf_parm} and \code{rmf_list}
}
\description{
Create a parameter for List Data input used in MODFLOW boundary condition packages.
}
\details{
the variable in the data.frame which is multiplied differs between boundary condition packages. 
         if the parameter is to be time-varying, a separate parameter should be created for each instance with a unique \code{instnam} but with the same \code{name}
}
\seealso{
\code{\link{rmf_create_list}}, \code{\link{rmf_create_array_parameter}}, \code{\link{rmf_create_flow_parameter}}
}
