% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-lmt.R
\name{rmf_create_lmt}
\alias{rmf_create_lmt}
\title{Create a \code{RMODFLOW} lmt object}
\usage{
rmf_create_lmt(
  fname = "output.ftl",
  inftl = 333,
  extended = TRUE,
  formatted = FALSE,
  package_flows = NULL
)
}
\arguments{
\item{fname}{character; name of the flow-transport link file produced. Defaults to \code{'output.ftl'}}

\item{inftl}{integer; unit number on which the flow-transport link file will be saved; defaults to 333}

\item{extended}{logical; should the extended header be used (instead of the standard header); defaults to TRUE}

\item{formatted}{logical; should the file be formatted. Under the default (FALSE), an unformatted flow-transport link file will be created by lmt}

\item{package_flows}{optional; additional flows to be saved in the flow-transport link file. Possible values are \code{'ALL'} or any combination of \code{'UZF', 'SFR' or 'LAK'}. Defaults to not including any of the above.}
}
\value{
object of class lmt
}
\description{
Create a \code{RMODFLOW} lmt object
}
\examples{
rmf_create_lmt(fname = file.path('..', 'transport', 'output.ftl'), formatted = TRUE, package_flows = c('UZF', 'LAK'))

}
\seealso{
\code{\link{rmf_read_lmt}}, \code{\link{rmf_write_lmt}} and \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?lmt6.htm}
}
