% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-lpf.R
\name{rmf_create_lpf}
\alias{rmf_create_lpf}
\alias{create_lpf}
\title{Create an \code{RMODFLOW} lpf object}
\usage{
rmf_create_lpf(dis, ilpfcb = 0, hdry = -888,
  storagecoefficient = FALSE, constantcv = FALSE, thickstrt = FALSE,
  nocvcorrection = FALSE, novfc = FALSE, noparcheck = FALSE,
  laytyp = rmfi_ifelse0(dis$nlay == 1, 1, c(1, rep(0, dis$nlay - 1))),
  layavg = rep(0, dis$nlay), chani = rep(1, dis$nlay),
  layvka = rep(0, dis$nlay), laywet = rep(0, dis$nlay), wetfct = 0.1,
  iwetit = 1, ihdwet = 0, parameters = NULL,
  hk = rmf_create_array(1e-04, dim = c(dis$nrow, dis$ncol, dis$nlay)),
  hani = rmfi_ifelse0(all(chani <= 0), hk * 0 + 1, NULL),
  vka = rmfi_ifelse0(all(layvka == 0), hk, NULL),
  ss = rmfi_ifelse0("TR" \%in\% dis$sstr, hk * 0 + 1e-05, NULL),
  sy = rmfi_ifelse0("TR" \%in\% dis$sstr && all(laytyp > 0), hk * 0 +
  0.15, NULL), vkcb = NULL, wetdry = NULL)

create_lpf(...)
}
\arguments{
\item{dis}{RMODFLOW dis object}

\item{ilpfcb}{flag and unit number for writing cell-by-cell flow terms; defaults to 0}

\item{hdry}{head assigned to cells that are converted to dry cells; defaults to -888}

\item{storagecoefficient}{logical; should STORAGECOEFFICIENT keyword be included?; defaults to FALSE}

\item{constantcv}{logical; should CONSTANTCV keyword be included?; defaults to FALSE}

\item{thickstrt}{logical; should THICKSTRT keyword be included?; defaults to FALSE}

\item{nocvcorrection}{logical; should NOCVCORRECTION keyword be included?; defaults to FALSE}

\item{novfc}{logical; should NOVFC keyword be included?; defaults to FALSE}

\item{noparcheck}{logical; should NOPARCHECK keyword be included?; defaults to FALSE}

\item{laytyp}{vector of flags for each layer, specifying layer type; defaults to all confined (0) except the first layer (1)}

\item{layavg}{vector of flags for each layer, specifying interblock transmissivity calculation method; defaults to 0 for each layer}

\item{chani}{vector of flags or horizontal anisotropies for each layer; defaults to 1 for each layer}

\item{layvka}{vector of flags for each layer, indicating whether vka is the vertical hydraulic conductivity or the ratio of horizontal to vertical; defaults to 0 for each layer}

\item{laywet}{vector of flags for each layer, indicating if wetting is active; defaults to 0 for each layer}

\item{wetfct}{is a factor that is included in the calculation of the head that is initially established at a cell when it is converted from dry to wet; defaults to 0.1}

\item{iwetit}{is the iteration interval for attempting to wet cells; defaults to 1}

\item{ihdwet}{is a flag that determines which equation is used to define the initial head at cells that become wet; defaults to 0}

\item{parameters}{list of \code{rmf_parameter} as created by \code{\link{rmf_create_parameter}}. See details; defaults to NULL}

\item{hk}{3d array with hydraulic conductivity along rows; defaults to 1. If not read for a specific layer, set all values in that layer to NA.}

\item{hani}{3d array with the ratio of hydraulic conductivity along columns to that along rows; defaults to 1. If not read for a specific layer, set all values in that layer to NA.}

\item{vka}{3d array with vertical hydraulic conductivity or the ratio of horizontal to vertical; defaults to hk. If not read for a specific layer, set all values in that layer to NA.}

\item{ss}{3d array with specific storage; only required when there are transient stress periods; defaults to 1E-5. If not read for a specific layer, set all values in that layer to NA.}

\item{sy}{3d array with specific yield; only required when there are transient stress periods; defaults to 0.15. If not read for a specific layer, set all values in that layer to NA.}

\item{vkcb}{3d array with vertical hydraulic conductivity of quasi-three-dimensional confining beds; defaults to NULL. If not read for a specific layer, set all values in that layer to NA.}

\item{wetdry}{3d array with a wetting threshold and flag indicating which neighboring cells can cause a cell to become wet; defaults to NULL. If not read for a specific layer, set all values in that layer to NA.}
}
\value{
Object of class lpf
}
\description{
\code{rmf_create_lpf} creates an \code{RMODFLOW} lpf object.
}
\details{
Flow variables are any of \code{HK, HANI, VK, VANI, SS, SY and VKCB}. A single variable can be specified either through the use of parameters or by using direct array input.
         When a flow variable for a specific layer is specified using parameters, all flow variables of the type must be specified by parameters. E.g. if a flow parameter defines HK for layer 1, HK must be defined for all layers using flow parameters instead of direct array input.
}
\section{Functions}{
\itemize{
\item \code{create_lpf}: Deprecated function name
}}

\seealso{
\code{\link{rmf_create_parameter}}, \code{\link{rmf_read_lpf}}, \code{\link{rmf_write_lpf}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?lpf.htm}
}
