% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-huf.R
\name{rmf_create_lvda}
\alias{rmf_create_lvda}
\title{Create an \code{RMODFLOW} lvda object}
\usage{
rmf_create_lvda(parameters)
}
\arguments{
\item{parameters}{either a single \code{rmf_parameter} or a list of \code{rmf_parameters} specifying the angle between the grid axis and the principal direction of horizontal hydraulic conductivity. See details.}
}
\value{
a \code{RMODFLOW} lvda object
}
\description{
\code{rmf_create_lvda} creates an \code{RMODFLOW} lvda object.
}
\details{
All parameters should have a hgunam attribute and their partyp attribute set to 'LVDA'.
         The LVDA package can only be used in conjunction with the HUF package.
}
\seealso{
\code{\link{rmf_read_lvda}}, \code{\link{rmf_write_lvda}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?lvda.htm}
}
