% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-nam.R
\name{rmf_create_nam}
\alias{rmf_create_nam}
\title{Create an \code{RMODFLOW} nam object}
\usage{
rmf_create_nam(..., basename = NULL)
}
\arguments{
\item{...}{RMODFLOW objects to be included in the nam file}

\item{basename}{character specifying the basename of the files. The default (\code{NULL}) sets input basenames to 'input' and output to 'output'.}
}
\value{
Object of class nam
}
\description{
\code{rmf_create_nam} creates an \code{RMODFLOW} nam object.
}
\details{
if a \code{RMODFLOW nam} object is present, it is recreated.
}
\seealso{
\code{\link{rmf_read_nam}}, \code{\link{rmf_write_nam}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?name_file.htm}
}
