% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_create_parameter.default}
\alias{rmf_create_parameter.default}
\title{Create an array parameter}
\usage{
\method{rmf_create_parameter}{default}(dis, kper = NULL, layer = NULL,
  parnam, parval = 1, mltnam = "NONE", zonnam = "ALL", iz = NULL,
  partyp = NULL, mlt = NULL, zon = NULL, instnam = NULL,
  hgunam = NULL)
}
\arguments{
\item{dis}{\code{RMODFLOW} dis object. Used to dimension the final array.}

\item{kper}{integer vector with the stress period numbers during which the parameter is active. Specifying kper indicates that this parameter represent boundary condition information.}

\item{layer}{integer vector denoting the layer indices represented by the parameter. Specifying layer indicates that this parameter represent flow package information; the partyp argument should be set as well. Multiple instances of the same layer index are allowed.}

\item{parnam}{character specifying the name of the parameter}

\item{parval}{numeric specifying the value of the parameter which is used to multiply values in the array. Defaults to 1.0}

\item{mltnam}{character vector with the same length as \code{layer}; specifying the names of multiplier arrays that are used to build the parameter. The keyword \code{"NONE"} indicates no multiplier array is present.}

\item{zonnam}{character vector with the same length as \code{layer}; specifying the names of zone arrays that are used to build the parameter. The keyword \code{"ALL"} indicates no multiplier array is present.}

\item{iz}{list with the same length as \code{layer} where each element is a numeric vector with the zone numbers for the corresponding zone array. Only read when the corresponding \code{zonnam} is not \code{"ALL"}.}

\item{partyp}{character specifying the type of flow parameter. Allowed values are \code{HK, HANI, VK, VANI, SS, SY}, \code{VKCB} for lpf and \code{SYTP} for huf. Not used when \code{layer} is \code{NULL}.}

\item{mlt}{\code{RMODFLOW} mlt object which holds the multiplier arrays specified in \code{mltnam}}

\item{zon}{\code{RMODFLOW} zon object which holds the zone arrays specified in \code{zonnam}}

\item{instnam}{optional character specying the instance name of the parameter is to be time-varying; defaults to NULL}

\item{hgunam}{character vector specifying the name(s) of the hydrogeological unit(s) if the parameter represents a huf parameter; defaults to NULL. See details.}
}
\value{
an \code{rmf_2d_array} object of class \code{rmf_parameter}
}
\description{
Create a parameter from multiplier and/or zone arrays used in MODFLOW boundary condition packages and flow packages
}
\details{
A boundary-condition parameter is created by setting the kper argument. A flow parameter is created by setting the layer and partyp arguments.
         If the boundary-condition parameter is to be time-varying, a separate parameter should be created for each instance with a unique \code{instnam} but with the same \code{name} 
         Typically, an array parameter is build from a single multiplier and/or zone array combination. However, multiple combinations can be used.
         If \code{mltnam} is \code{"NONE"} and \code{zonnam} is \code{"ALL"}, \code{parval} applies to all cells in the array and \code{mlt} and \code{zon} are not read.
         If \code{hgunam} is specified, \code{layer} can not be specified.
}
\seealso{
\code{\link{rmf_create_array}}
}
