% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-create-array-parameter.R
\name{rmf_create_parameter.rmf_2d_array}
\alias{rmf_create_parameter.rmf_2d_array}
\title{Create an input array parameter}
\usage{
rmf_create_parameter.rmf_2d_array(array, kper = attr(array, "kper"),
  parnam, parval = 1, mlt_names = "NONE", zon_names = "ALL",
  iz = NULL, mlt = NULL, zon = NULL, instnam = NULL)
}
\arguments{
\item{array}{\code{rmf_array} containing the variable}

\item{kper}{integer vector with the stress period numbers during which the parameter is active}

\item{parnam}{character specifying the name of the parameter}

\item{parval}{numeric specifying the value of the parameter which is used to multiply values in the array. Defaults to 1.0}

\item{mlt_names}{character vector specifying the names of multiplier arrays that are used to build the parameter. The keyword \code{"NONE"} indicates no multiplier array is present.}

\item{zon_names}{character vector specifying the names of zone arrays that are used to build the parameter. The keyword \code{"ALL"} indicates no multiplier array is present.}

\item{iz}{only read when the corresponding \code{zon_names} is not \code{"ALL"}. A list where each element is a numeric vector with the zone numbers for the corresponding zone array.}

\item{mlt}{\code{RMODFLOW} mlt object which holds the multiplier arrays specified in \code{mlt_names}}

\item{zon}{\code{RMODFLOW} zon object which holds the zone arrays specified in \code{zon_names}}

\item{instnam}{optional character specying the instance name of the parameter is to be time-varying; defaults to NULL}
}
\value{
an object of class \code{rmf_parm} and \code{rmf_2d_array}
}
\description{
Create a parameter for array input used in MODFLOW boundary condition packages from multiplier and/or zone arrays
}
\details{
if the parameter is to be time-varying, a separate parameter should be created for each instance with a unique \code{instnam} but with the same \code{name} 
         Typically, an array parameter is build from a single multiplier and/or zone array combination. However, multiple combinations can be used.
         If \code{mlt_names} is \code{"NONE"} and \code{zon_names} is \code{"ALL"}, \code{value} applies to all cells in the array and \code{mlt} and \code{zon} are not read.
}
\seealso{
\code{\link{rmf_create_array}}, \code{\link{rmf_create_flow_parameter}}, \code{\link{rmf_create_list_parameter}}
}
