% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_create_parameter.rmf_3d_array}
\alias{rmf_create_parameter.rmf_3d_array}
\title{Concise function for creating a 3D-array parameter.}
\usage{
\method{rmf_create_parameter}{rmf_3d_array}(
  array,
  parnam,
  parval = 1,
  layer = 1:dim(array)[3],
  partyp,
  mltnam = parnam,
  instnam = NULL
)
}
\arguments{
\item{array}{a \code{rmf_3d_array} object}

\item{parnam}{character specifying the name of the parameter}

\item{parval}{numeric specifying the value of the parameter which is used to multiply values in the array. Defaults to 1.0}

\item{layer}{integer vector denoting the layer indices represented by the parameter. Defaults to \code{seq(1, dim(array)[3])}}

\item{partyp}{character specifying the type of flow parameter. Allowed values are \code{HK, HANI, VK, VANI, SS, SY and VKCB}.}

\item{mltnam}{character vector specifying the name of the resulting multiplier array.}

\item{instnam}{optional character specying the instance name of the parameter is to be time-varying; defaults to NULL}
}
\value{
a \code{rmf_3d_array} objects of class \code{rmf_parameter}
}
\description{
Create a list of MODFLOW flow parameters from a 3D-array.
}
\details{
This function will only set the multiplier arrays for all layers of the 3D array. The user must make sure that this multiplier array is written to a separate MLT file when running a MODFLOW simulation.
This function is intended to be used when no multiplier and/or zone arrays are specified for a parameter.
Only flow parameters can be created with this function. A singly partyp is appointed to all layers.
}
