% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_create_parameter.rmf_list}
\alias{rmf_create_parameter.rmf_list}
\title{Create a List Data input parameter}
\usage{
\method{rmf_create_parameter}{rmf_list}(
  rmf_list,
  parnam,
  parval = 1,
  instnam = NULL,
  kper = attr(rmf_list, "kper")
)
}
\arguments{
\item{rmf_list}{a rmf_list object}

\item{parnam}{character specifying the name of the parameter}

\item{parval}{numeric specifying the value of the parameter which is used to multiply the flux controlling variable in the data.frame. Defaults to 1.0}

\item{instnam}{optional character specying the instance name of the parameter is to be time-varying; defaults to NULL}
}
\value{
an object of class \code{rmf_parameter} and \code{rmf_list}
}
\description{
Create a parameter for List Data input used in MODFLOW boundary condition packages.
}
\details{
the variable in the data.frame which is multiplied differs between boundary condition packages.
if the parameter is to be time-varying, a separate parameter should be created for each instance with a unique \code{instnam} but with the same \code{name}
}
\seealso{
\code{\link{rmf_create_list}}
}
