% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-pcg.R
\name{rmf_create_pcg}
\alias{rmf_create_pcg}
\title{Create an \code{RMODFLOW} pcg object}
\usage{
rmf_create_pcg(
  mxiter = 20,
  iter1 = 30,
  npcond = 1,
  ihcofadd = 0,
  hclose = 0.001,
  rclose = 0.001,
  relax = 1,
  nbpol = 1,
  iprpcg = 1,
  mutpcg = 0,
  damppcg = 1,
  damppcgt = 1
)
}
\arguments{
\item{mxiter}{maximum number of outer iterations; defaults to 20}

\item{iter1}{number of inner iterations; defaults to 30}

\item{npcond}{flag to select matrix conditioning method (1: Modified Incomplete Cholesky (default); 2: Polynomial)}

\item{ihcofadd}{flag that determines what happens to an active cell that is surrounded by dry cells; defaults to 0}

\item{hclose}{head change criterion for convergence (length); defaults to 0.001}

\item{rclose}{residual criterion for convergence (cubic length per time); defaults to 0.001}

\item{relax}{relaxation parameter used with npcond == 1; defaults to 1}

\item{nbpol}{used when npcond == 2; should equal 2 for maximum eigenvalue of 2; else the estimate is calculated; defaults to 1}

\item{iprpcg}{printout interval for pcg; defaults to 1}

\item{mutpcg}{flag controlling printing information from the solver; defaults to 0}

\item{damppcg}{damping factor}

\item{damppcgt}{damping factor for transient stress periods; optional; only read when damppcg is negative}
}
\value{
Object of class pcg
}
\description{
\code{rmf_create_pcg} creates an \code{RMODFLOW} pcg object.
}
\seealso{
\code{\link{rmf_read_pcg}}, \code{\link{rmf_write_pcg}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?pcg.htm}
}
