% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rmf_create_prj}
\alias{rmf_create_prj}
\title{Create a RMODFLOW projection object}
\usage{
rmf_create_prj(
  origin = c(0, 0, 0),
  rotation = 0,
  crs = NA,
  ulcoordinate = FALSE,
  nodecoordinate = FALSE,
  dis = NULL
)
}
\arguments{
\item{origin}{numeric vector with the x & y (and optionally z) coordinates of the origin which by default is the lowerleft corner of the model. Defaults to \code{c(0, 0, 0)}.}

\item{rotation}{numeric; counterclockwise rotation angle (in degrees). Rotation is around the lowerleft corner of the model. Defaults to 0.}

\item{crs}{coordinate reference system of the model. Any values accepted by \code{sf::st_crs} may be defined. Defaults to \code{NA}.}

\item{ulcoordinate}{logical; if \code{TRUE}, \code{origin} refers to the upperleft cell instead of the lowerleft.}

\item{nodecoordinate}{logical; if \code{TRUE}, \code{origin} refers to the cell center instead of the cell corner.}

\item{dis}{\code{RMODFLOW} dis object. Only required when \code{uulcoordinate} or \code{nodecoordinate} is TRUE.}
}
\value{
an object of class \code{prj} which is a list with the (1) origin vector containing x, y and z coordinates of the lowerleft corner,
(2) the rotation angle in degrees counterclockwise and (3) the coordinate reference system as an \code{sf crs} object
}
\description{
Create a RMODFLOW projection object
}
\details{
\code{origin} should be specified in the length units defined by \code{crs}.
If no z coordinate is specified in \code{origin}, it is set to zero.
\code{crs} is set by a call to \code{sf::st_crs}.
Rotation is around the lowerleft corner of the model.
}
\examples{
rmf_create_prj(origin = c(152082, 168000.2), rotation = -12, crs = 31370)

dis <- rmf_create_dis()
rmf_create_prj(origin = c(120, 300, 13), ulcoordinate = TRUE, dis = dis)

}
