% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-create-riv.R
\name{rmf_create_riv}
\alias{rmf_create_riv}
\title{Create an \code{RMODFLOW} riv object.}
\usage{
rmf_create_riv(npriv = 1, mxl = 10, mxactr = 10, irivcb = 0,
  option = NULL, parnam = "STREAM", parval = 1, nlst = 10,
  instances = NULL, numinst = NULL, instnam = NULL,
  layer_parm = list(array(rep(1, 10), dim = c(1, 10))),
  row_parm = list(array(rep(1, 10), dim = c(1, 10))),
  column_parm = list(array(1:10, dim = c(1, 10))),
  stage_parm = list(array(-2, dim = c(1, 10))),
  condfact_parm = list(array(5e-07, dim = c(1, 10))),
  rbot_parm = list(array(-4, dim = c(1, 10))), xyz_parm = NULL, itmp = 0,
  np = 1, layer_sp = NULL, row_sp = NULL, column_sp = NULL,
  stage_sp = NULL, cond_sp = NULL, rbot_sp = NULL, xyz_sp = NULL,
  pname = list("STREAM"), iname = NULL)
}
\arguments{
\item{npriv}{number of riv parameters; defaults to 1}

\item{mxl}{maximum number of river reaches that will be defined using parameters; defaults to 10}

\item{mxactr}{maximum number of river reaches in use during any stress period defined with and without parameters; defaults to 10}

\item{irivcb}{flag and unit number for writing cell-by-cell flow terms; defaults to 0}

\item{option}{optional character vector specifying the auxiliary variables and/or \code{NOPRINT}. Do not include the keywords \code{AUXILIARY} or \code{AUX}; defaults to NULL}

\item{parnam}{vector of length \code{npriv} specifying the parameter names; defaults to 'STREAM'}

\item{parval}{vector of length \code{npriv} specifying the parameter values; defaults to 1}

\item{nlst}{vector of length \code{npriv} specifying the number of river reaches that are included in a non-time-varying parameter or in each instance of a time-varying parameter; defaults to 10}

\item{instances}{logical vector of length \code{npriv} indicating which parameters are time-varying; defaults to NULL}

\item{numinst}{vector of length \code{npriv} indicating the number instances that are included in the time-varying parameter; defaults to NULL}

\item{instnam}{list with \code{npriv} elements where each element \code{i} is a character vector of length \code{numinst} for parameter \code{i} specifying the names of the parameter instances. If not time-varying, set numinst dimension to 1. Defaults to NULL}

\item{layer_parm}{list with \code{npriv} elements where each element \code{i} is a numeric 2D array with dimensions \code{numinst x nlst} specifying the layer numbers of the model cells containing the river reaches specified by parameter \code{i}. If not time-varying, set numinst dimension to 1. Defaults to 1 for all river reaches for the parameter.}

\item{row_parm}{list with \code{npriv} elements where each element \code{i} is a numeric 2D array with dimensions \code{numinst x nlst} specifying the row numbers of the model cells containing the river reaches specified by parameter \code{i}. If not time-varying, set numinst dimension to 1. Defaults to 1 for all river reaches for the parameter}

\item{column_parm}{list with \code{npriv} elements where each element \code{i} is a numeric 2D array with dimensions \code{numinst x nlst} specifying the column numbers of the model cells containing the river reaches specified by parameter \code{i}. If not time-varying, set numinst dimension to 1. Defaults to 1:10 for the parameter}

\item{stage_parm}{list with \code{npriv} elements where each element \code{i} is a numeric 2D array with dimensions \code{numinst x nlst} specifying the head in the river for parameter \code{i}. If not time-varying, set numinst dimension to 1. Defaults to -2 for all river reaches for the parameter}

\item{condfact_parm}{list with \code{npriv} elements where each element \code{i} is a numeric 2D array with dimensions \code{numinst x nlst} specifying the river conductance factor for parameter \code{i}. The river conductance is the product of condfact and the parameter value. If not time-varying, set numinst dimension to 1. Defaults to 5e-7 for all river reaches for the parameter}

\item{rbot_parm}{list with \code{npriv} elements where each element \code{i} is a numeric 2D array with dimensions \code{numinst x nlst} specifying the elevation of the riverbed bottom for parameter \code{i}. If not time-varying, set numinst dimension to 1. Defaults to -4 for all river reaches for the parameter}

\item{xyz_parm}{optional list with \code{npriv} elements where each element \code{i} is a character 2D array with dimensions \code{numinst x nlst} specifying the auxiliary variables for parameter \code{i}. If not time-varying, set numinst to 1. Defaults to NULL}

\item{itmp}{vector of length \code{dis$nper} specifying non-parameter river data being read for each stress period; defaults to 0}

\item{np}{vector of length \code{dis$nper} specifying the number of river parameters in use for each stress period; defaults to 1}

\item{layer_sp}{list with \code{dis$nper} elements where each element \code{i} is a numeric vector of length \code{itmp} for stress period \code{i} specifying the layer numbers of the model cells containing the river reaches; defaults to NULL}

\item{row_sp}{list with \code{dis$nper} elements where each element \code{i} is a numeric vector of length \code{itmp} for stress period \code{i} specifying the row numbers of the model cells containing the river reaches; defaults to NULL}

\item{column_sp}{list with \code{dis$nper} elements where each element \code{i} is a numeric vector of length \code{itmp} for stress period \code{i} specifying the column numbers of the model cells containing the river reaches; defaults to NULL}

\item{stage_sp}{list with \code{dis$nper} elements where each element \code{i} is a numeric vector of length \code{itmp} for stress period \code{i} specifying the river stages of the model cells containing the river reaches; defaults to NULL}

\item{cond_sp}{list with \code{dis$nper} elements where each element \code{i} is a numeric vector of length \code{itmp} for stress period \code{i} specifying the riverbed conductance of the model cells containing the river reaches; defaults to NULL}

\item{rbot_sp}{list with \code{dis$nper} elements where each element \code{i} is a numeric vector of length \code{itmp} for stress period \code{i} specifying the bottom elevations of the riverbed in the model cells containing the river reaches; defaults to NULL}

\item{xyz_sp}{optional list with \code{dis$nper} elements where each element \code{i} is a character vector of length \code{itmp} for stress period \code{i} specifying the auxiliary variables of the model cells containing the river reaches; defaults to NULL}

\item{pname}{list with \code{dis$nper} elements where each element \code{i} is a character vector of length \code{np} for stress period \code{i} specifying the names of the parameters being used; defaults to 'STREAM'}

\item{iname}{list with \code{dis$nper} elements where each element \code{i} is a character vector of length \code{np} for stress period \code{i} specifying the names of the parameter instances being used; defaults to NULL}
}
\value{
\code{RMODFLOW} riv object
}
\description{
\code{rmf_create_riv} creates an \code{RMODFLOW} riv object
}
\seealso{
\code{\link{rmf_read_riv}}, \code{\link{rmf_write_riv}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?riv.htm}
}
