% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-sip.R
\name{rmf_create_sip}
\alias{rmf_create_sip}
\title{Create an \code{RMODFLOW} sip object.}
\usage{
rmf_create_sip(
  mxiter = 50,
  nparm = 5,
  accl = 1,
  hclose = 0.01,
  ipcalc = 1,
  wseed = 0,
  iprsip = 0
)
}
\arguments{
\item{mxiter}{maximum number of iterations in one time step attempting to solve the equations; defaults to 50}

\item{nparm}{number of iteration variables; defaults to 5}

\item{accl}{acceleration variable; defaults to 1}

\item{hclose}{head change criterion for convergence in units of length; defaults to 0.01}

\item{ipcalc}{flag indicating where the seed for calculating iteration variables comes from; defaults to 1}

\item{wseed}{user-specified seed for calculating iteration variables; defaults to 0}

\item{iprsip}{printout interval for maximum head change; defaults to 0}
}
\value{
\code{RMODFLOW} sip object
}
\description{
\code{rmf_create_sip} creates an \code{RMODFLOW} sip object
}
\seealso{
\code{\link{rmf_read_sip}}, \code{\link{rmf_write_sip}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?sip.htm}
}
