% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-create-wel.R
\name{rmf_create_wel}
\alias{rmf_create_wel}
\title{Create an \code{RMODFLOW} wel object.}
\usage{
rmf_create_wel(..., dis, iwelcb = 0, noprint = FALSE, aux = NULL)
}
\arguments{
\item{...}{\code{rmf_list} (possibly of class \code{rmf_parm}) objects or a single \code{list} with \code{rmf_list} objects (possibly of class \code{rmf_parm}) elements; defines the wells.}

\item{dis}{dis object}

\item{iwelcb}{flag and unit number for writing cell-by-cell flow terms; defaults to 0 (cell-by-cell flow terms will not be written)}

\item{noprint}{logical, should the printing of WEL cells to the listing file be suppressed ? Defaults to \code{FALSE}}

\item{aux}{optional character vector specifying the names of the auxiliary variables. These variables should also be included in the \code{rmf_list} objects that are supplied; defaults to \code{NULL}}
}
\value{
\code{RMODFLOW} wel object
}
\description{
\code{rmf_create_wel} creates an \code{RMODFLOW} wel object
}
\seealso{
\code{\link{rmf_read_wel}}, \code{\link{rmf_write_wel}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?wel.htm}
}
