% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-zon.R
\name{rmf_create_zon}
\alias{rmf_create_zon}
\title{Create an \code{RMODFLOW} zon object}
\usage{
rmf_create_zon(
  nzn = length(zonnam),
  zonnam = "ZONE",
  izon = rmf_create_array(1L, dim = c(10, 10))
)
}
\arguments{
\item{nzn}{number of zone arrays to be defined; defaults to the length of zonnam}

\item{zonnam}{character vector of length \code{nzn} specifying the names of zone arrays; defaults to 'ZONE'}

\item{izon}{either a single 2d array or list with \code{nzn} 2d_arrays specifying the zone arrays; defaults to a \code{rmf_2d_array} with 1 for all cells}
}
\value{
an \code{RMODFLOW} zon object
}
\description{
\code{rmf_create_zon} creates an \code{RMODFLOW} zon object
}
\seealso{
\code{\link{rmf_read_zon}}, \code{\link{rmf_write_zon}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?zone.htm}
}
