% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_darcy_flux}
\alias{rmf_darcy_flux}
\alias{rmf_darcy_flux.rmf_2d_array}
\alias{rmf_darcy_flux.rmf_3d_array}
\alias{rmf_darcy_flux.rmf_4d_array}
\title{Calculate the Darcy flux}
\usage{
rmf_darcy_flux(...)

\method{rmf_darcy_flux}{rmf_2d_array}(hed, dis, hk, hani = rep(1,
  dis$nlay), porosity = NULL, mask = NULL)

\method{rmf_darcy_flux}{rmf_3d_array}(hed, dis, flow = NULL, hk,
  hani = rep(1, dis$nlay), vka, vani = rep(FALSE, dis$nlay),
  porosity = NULL, mask = NULL)

\method{rmf_darcy_flux}{rmf_4d_array}(hed, dis, l, ...)
}
\arguments{
\item{...}{additional arguments passed to \code{\link{rmf_darcy_flux.rmf_3d_array}}}

\item{hed}{2d array with the hydraulic heads}

\item{dis}{\code{RMODFLOW} dis object}

\item{hk}{2d array with the hydraulic conductivities along rows}

\item{hani}{either a 2d array or a single value specifying the horizontal anisotropy. See details. Defaults to isotropic conditions.}

\item{porosity}{optional 2d array with the porosity values as fractions. See details. Defaults to NULL}

\item{mask}{optional logical 2d array used to set the active cells; defaults to NULL}

\item{flow}{optional \code{RMODFLOW} flow package, i.e. one of lpf, bcf, huf or upw. Overwrites the use of hk, hani, vka & vani. See details.}

\item{vka}{3d array with either vertical hydraulic conductivities or vertical anisotropies depending on the value of vani. See details.}

\item{vani}{a single logical value for each layer specifying whether vka represents vertical hydraulic conductivity (FALSE) or vertical anisotropy for the layer (TRUE). Defaults to FALSE for all layers.}

\item{l}{integer index used to subset the 4th dimension of hed}

\item{hed}{3d array with the hydraulic heads}

\item{dis}{\code{RMODFLOW} dis object}

\item{hk}{3d array with the hydraulic conductivities along rows}

\item{hani}{either a 3d array or a single value for each layer specifying the horizontal anisotropy. See details. Defaults to isotropic conditions.}

\item{porosity}{optional 3d array with the porosity values as fractions. See details. Defaults to NULL}

\item{mask}{optional logical 3d array used to set the active cells; defaults to NULL}

\item{hed}{4d array with the hydraulic heads}

\item{dis}{\code{RMODFLOW} dis object}
}
\value{
a list with the magnitude, x and y components of the Darcy flux (or fluid velocity if porosity is specified) as 2d arrays

a list with the magnitude, x, y and z components of the Darcy flux (or fluid velocity if porosity is specified) as 3d arrays
}
\description{
Calculate the Darcy flux

Calculate the Darcy flux from a 2d array

Calculate the Darcy flux from a 3d array

Calculate the Darcy flux from a 4d array
}
\details{
Horizontal anisotropy is the ratio of hydraulic conductivity along columns (Y direction) to the hydraulic conductivity along rows (X direction).
         Darcy flux (or Darcy velocity) is not the actual velocity at which the fluid travels. In order to obtain this fluid velocity, a porosity argument should be specified.
         The flux components are positive in the direction of increasing Cartesian axes.

When flow is specified, all flow parameters are obtained from the flow package.
         Horizontal anisotropy is the ratio of hydraulic conductivity along columns (Y direction) to the hydraulic conductivity along rows (X direction).
         Vertical anisotropy is the ratio of horizontal hydraulic conductivity along rows (X direction) to the vertical hydraulic conductivity.
         Darcy flux (or Darcy velocity) is not the actual velocity at which the fluid travels. In order to obtain this fluid velocity, a porosity argument should be specified.
         The flux components are positive in the direction of increasing Cartesian axes.

hed is subsetted to a 3d array and passed to \code{\link{rmf_darcy_flux.rmf_3d_array}}.
}
