% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{rmf_example}
\alias{rmf_example}
\title{Access example MODFLOW files}
\usage{
rmf_example(name = NULL, code = NULL)
}
\arguments{
\item{name}{filename or model name; if NULL provides a list of possibilities}

\item{code}{identification of the MODFLOW variant, only required for the
examples that come with the code archives available online}
}
\value{
character vector of model or file names/paths
}
\description{
Access example MODFLOW files
}
