% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{rmf_example_file}
\alias{rmf_example_file}
\title{Path to example file}
\usage{
rmf_example_file(filename = NULL)
}
\arguments{
\item{filename}{filename of the example file}
}
\value{
path to example file
}
\description{
Path to example file
}
