% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{rmf_example_files}
\alias{rmf_example_files}
\title{Path to example files}
\usage{
rmf_example_files(pattern = NULL)
}
\arguments{
\item{pattern}{pattern to match example file filenames. If NULL (default), returns vector of all example file filenames.}
}
\value{
path to example files
}
\description{
Path to example files
}
