% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-example-model.R
\name{rmf_example_model}
\alias{rmf_example_model}
\title{List example model files}
\usage{
rmf_example_model(model)
}
\arguments{
\item{model}{}
}
\value{
example model files
}
\description{
List example model files
}
