% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_export_raster}
\alias{rmf_export_raster}
\title{Generic function to export GIS raster layers from RMODFLOW arrays}
\usage{
rmf_export_raster(...)
}
\description{
Generic function to export GIS raster layers from RMODFLOW arrays
}
