% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-export-table.R,
%   R/rmf-export-table.rmf-4d-array.R
\name{rmf_export_table}
\alias{rmf_export_table}
\alias{rmf_export_table.rmf_4d_array}
\title{Generic function to export tables from RMODFLOW arrays}
\usage{
rmf_export_table(...)

\method{rmf_export_table}{rmf_4d_array}(array, k, l, dis, bas = NULL,
  mask = rmfi_ifelse0(is.null(bas), array * 0 + 1, bas$ibound[, , 1]),
  prj = NULL, crs = NULL, file = "rmf_export.csv", type = "csv")
}
\description{
Generic function to export tables from RMODFLOW arrays

Generic function to export tables from RMODFLOW arrays
}
