% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-export-vector.R,
%   R/rmf-export-vector.rmf-2d-array.R
\name{rmf_export_vector}
\alias{rmf_export_vector}
\alias{rmf_export_vector.rmf_2d_array}
\title{Generic function to export GIS vector layers from RMODFLOW arrays}
\usage{
rmf_export_vector(...)

\method{rmf_export_vector}{rmf_2d_array}(array, dis, bas = NULL,
  mask = rmfi_ifelse0(is.null(bas), array * 0 + 1, bas$ibound[, , 1]),
  prj = NULL, crs = NULL, file = "rmf_vector",
  type = "ESRI Shapefile", include_ijk = FALSE)
}
\description{
Generic function to export GIS vector layers from RMODFLOW arrays

Generic function to export vectors
}
