% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rmf_extent}
\alias{rmf_extent}
\title{Obtain the corners and bounding box of the MODFLOW grid}
\usage{
rmf_extent(dis, prj = rmf_get_prj(dis))
}
\arguments{
\item{dis}{\code{RMODFLOW} dis object}

\item{prj}{optional \code{RMODFLOW} prj object}
}
\value{
a list with (1) the coordinates of the grid corners as a data.frame and (2) the bounding box of the grid as
a \code{sf bbox} object
}
\description{
Obtain the corners and bounding box of the MODFLOW grid
}
\examples{
dis <- rmf_create_dis()
prj <- rmf_create_prj(origin = c(152082, 168000.2), rotation = -12, crs = 31370)
rmf_extent(dis, prj)

}
