% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-find.R
\name{rmf_find}
\alias{rmf_find}
\title{Find paths to executables}
\usage{
rmf_find(name = "MODFLOW-2005", precision = "single")
}
\arguments{
\item{name}{Character. Name of the software. Currently supported values are \code{"MODFLOW-2005"}, \code{"MODFLOW-OWHM"}, \code{"MODFLOW-NWT"}, \code{"MODFLOW-LGR"} and \code{"MODFLOW-CFP"}.}

\item{precision}{Character. Can be \code{"single"} or \code{"double"}. Only relevant for MODFLOW-2005.}
}
\value{
Path to the executable.
}
\description{
\code{rmf_find} finds the path to external software executables.
}
\details{
The function first looks for the executable in the current working
directory. If not there, it looks in \file{C:/WRDAPP/}, where the
software might have been installed by \code{\link{rmf_install}}. If the
executable cannot be found, the system path variable is used in an attempt
to locate it. If it still cannot be located, you should first install the
software or add the folder with the executable to the system path.
}
\examples{
rmf_find("MODFLOW-2005")
}
