% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rmf_get_ibound}
\alias{rmf_get_ibound}
\title{Title}
\usage{
rmf_get_ibound(obj, dis, prj = NULL, op = sf::st_intersects,
  sparse = FALSE)
}
\arguments{
\item{obj}{}

\item{dis}{}

\item{prj}{}

\item{op}{}

\item{sparse}{}
}
\description{
Title
}
