% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_gradient}
\alias{rmf_gradient}
\alias{rmf_gradient.rmf_2d_array}
\alias{rmf_gradient.rmf_3d_array}
\alias{rmf_gradient.rmf_4d_array}
\title{Calculate a gradient field}
\usage{
rmf_gradient(...)

\method{rmf_gradient}{rmf_2d_array}(obj, dis, na_value = NULL,
  mask = obj * 0 + 1)

\method{rmf_gradient}{rmf_3d_array}(obj, dis, na_value = NULL,
  mask = obj * 0 + 1)

\method{rmf_gradient}{rmf_4d_array}(obj, dis, l, ...)
}
\arguments{
\item{...}{additional arguments passed to \code{\link{rmf_gradient.rmf_3d_array}}}

\item{obj}{2d array with the scalars}

\item{dis}{\code{RMODFLOW} dis object}

\item{na_value}{optional; sets these values in obj to 0; defaults to NULL}

\item{mask}{logical 2d array indicating which cells to include in the gradient calculation; defaults to all cells active}

\item{l}{integer index used to subset the 4th dimension of the 4d array}

\item{obj}{3d array with the scalars}

\item{dis}{\code{RMODFLOW} dis object}

\item{na_value}{optional; sets these values in obj to 0; defaults to NULL}

\item{mask}{logical 3d array indicating which cells to include in the gradient calculation; defaults to all cells active}

\item{obj}{4d array with the scalars}

\item{dis}{\code{RMODFLOW} dis object}
}
\value{
a list with the x and y components of the gradient field as 2d arrays

a list with the x, y and z components of the gradient field as 3d arrays

a list with the x, y and z components of the gradient field as 3d arrays
}
\description{
\code{rmf_gradient.rmf_2d_array} calculates the x and y components of the gradient from a 2d scalar field

\code{rmf_gradient.rmf_3d_array} calculates the x, y and z components of the gradient from a 3d scalar field

\code{rmf_gradient.rmf_4d_array} calculates the x, y and z components of the gradient from a 4d scalar field where the 4th dimension is time
}
\details{
The gradient is evaluated in the direction of increasing x & y values.

The gradient is evaluated in the direction of increasing x, y & z values.

the 4d array is subsetted on the 4th dimension to a 3d array from which the gradient is calculated
}
