% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{rmf_guide}
\alias{rmf_guide}
\title{Launch the online guide pages}
\usage{
rmf_guide(name = NULL)
}
\arguments{
\item{name}{Name of the \code{.htm} file linked to in the online guide contents or
index menu. If \code{NULL} (default), the introduction page is launched.}
}
\description{
This function can be used to launch the \href{https://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/}{Online Guide to MODFLOW} or
specific pages thereof. Other online guides are available as well, but this
one should be the most complete version with information on all different
MODFLOW 2005 variants that are part of the RMODFLOW scope. Some shortcuts are
available as well, to more easily access some of the pages through the
MODFLOW file type or ModelMuse file extension conventions.
}
\examples{
rmf_guide() # Launches the introduction page.
rmf_guide("name_file") # Launches the "name_file.htm" page.
rmf_guide("nam") # Launches the same page using the "nam" shortcut.
}
\seealso{
\code{\link[=rmf_browse]{rmf_browse()}} for launching the \href{https://rogiersbart.github.io/RMODFLOW/}{RMODFLOW pkgdown website}.
}
