% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-install.R
\name{rmf_install}
\alias{rmf_install}
\title{Install external software}
\usage{
rmf_install(software = "MODFLOW-2005")
}
\arguments{
\item{software}{character}
}
\description{
Install external software
}
\details{
Currently supported software: MODFLOW-2005.
}
\examples{
rmf_install("MODFLOW-2005")
}
