% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{rmf_optimize}
\alias{rmf_optimize}
\title{Optimize a MODFLOW model}
\usage{
rmf_optimize(
  path,
  code = "2005",
  start = NULL,
  lower = -Inf,
  upper = Inf,
  include = NULL,
  transform = NULL,
  cost = "ssq",
  backup = TRUE,
  restore = FALSE,
  visualize = interactive(),
  export = NULL,
  iterate = 50,
  tolerate = 1e-04,
  ...
)
}
\arguments{
\item{path}{Path to the NAM file. Typically with extension \code{.nam}.}

\item{code}{Name of the MODFLOW variant to use, or path to the executable.}

\item{start, lower, upper}{Vectors of PVAL file parameter values to start the
optimization, and corresponding lower and upper limits. These should be
named vectors if not all parameters are provided in their order of
occurrence, where the names (can be regular expressions to) match the
parameter names. Parameters that are not mentioned take the starting value
from the PVAL file, and/or have no constraints. If NULL (default), no
values are changed in the original PVAL file, and/or no constraints are
imposed. \code{lower} and \code{upper} also take named lists of functions and/or
numeric values, where the functions are then applied to the \code{start} or
original PVAL values. Note support for lower and upper limits is achieved
with the Nelder and Mead method by pretending the model cannot be evaluated
when violating them.}

\item{include}{Character vector indicating which PVAL file parameters should
be included. Regular expressions can be used to include multiple parameters
at once. Parameters that are not mentioned are not included. If \code{NULL}
(default), all parameters are included.}

\item{transform}{Character vector of transformations. This should be a named
vector if not all parameters are transformed, and listed in their order of
occurrence, where the names (can be regular expressions to) match the
parameter names. Parameters that are not mentioned are not transformed. If
\code{NULL} (default), no transformations are done. Currently only \code{"log"} is
supported.}

\item{cost}{Character. The performance measure that should be used as the
cost function. Possible values are those supported by \code{\link[=rmf_performance]{rmf_performance()}}.
Defaults to \code{"ssq"}.}

\item{backup}{Logical. Should a backup (with \code{.old} suffix) of the original
PVAL file be created? Defaults to \code{FALSE}.}

\item{restore}{Logical. Should the original PVAL file be restored? Defaults
to \code{FALSE}.}

\item{visualize}{Logical. Should the results be visualized during the
analysis? Defaults to \code{TRUE} in an interactive session; \code{FALSE} otherwise.}

\item{export}{Optional file path to export intermediate results to after each
iteration.}

\item{iterate}{Integer. Maximum number of iterations.}

\item{tolerate}{Double. Relative convergence tolerance.}

\item{...}{Optional arguments passed to \code{\link[=optim]{optim()}}.}
}
\value{
Invisible list with \code{\link[=optim]{optim()}} results and the full parameter list.
}
\description{
This function performs local optimization of a MODFLOW model.
}
\details{
Only works with models using MODFLOW parameters and having a head
predictions output file as defined in the HOB object The only method
currently available is "Nelder-Mead". See \code{\link{optim}}
}
\seealso{
\code{\link[=rmf_execute]{rmf_execute()}} for executing a MODFLOW model.\cr
\code{\link[=rmf_analyze]{rmf_analyze()}} for analyzing a MODFLOW model.
}
