% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_performance.hpr}
\alias{rmf_performance.hpr}
\alias{rmf_performance}
\title{Get model performance measures from a hpr object}
\usage{
\method{rmf_performance}{hpr}(hpr)

rmf_performance(...)
}
\arguments{
\item{hpr}{head predictions file object}
}
\value{
performance measures
}
\description{
Get model performance measures from a hpr object

Generic function to get model performance measures
}
