% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmf_plot.cbc}
\alias{rmf_plot.cbc}
\title{Plot a flux component of a cell-by-cell budget object}
\usage{
\method{rmf_plot}{cbc}(cbc, dis, i = NULL, j = NULL, k = NULL,
  l = NULL, flux = NULL, type = "fill", kper = NULL, kstp = NULL,
  active_only = FALSE, hed = NULL, porosity = NULL, ...)
}
\arguments{
\item{cbc}{a \code{RMODFLOW} cell-by-cell budget object}

\item{dis}{a \code{RMODFLOW} dis object}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{l}{time step number to plot; defaults to plotting the final time step.}

\item{flux}{character denoting which flux to plot. See details.}

\item{type}{plot type: 'fill' (default), 'factor', 'grid', 'contour' or 'vector'}

\item{kper}{integer specifying the stress-period. Use in conjunction with kstp. See details.}

\item{kstp}{integer specifying the time step of kper. Use in conjunction with kper. See details.}

\item{active_only}{logical; indicating if only the active cells should be plotted for list-directed components of the cbc object. Non-active cells are set to NA. Defaults to FALSE.}

\item{hed}{optional hed object for only plotting the saturated part of the grid; possibly subsetted with time step number. Also used in \code{\link{rmf_convert_cbc_to_darcy}} when flux = 'darcy' and type = 'vector'.}

\item{porosity}{optional 3d array with porosity values passed to \code{\link{rmf_convert_cbc_to_darcy}} when flux = 'darcy' and type = 'vector'.}

\item{...}{additional parameters passed to \code{\link{rmf_plot.rmf_4d_array}} or \code{\link{rmf_plot.rmf_list}}}
}
\value{
ggplot2 object or layer
}
\description{
Plot a flux component of a cell-by-cell budget object
}
\details{
Flux can be \code{'constant_head'}, \code{'storage'}, \code{'flow_right_face'}, \code{'flow_front_face'}, \code{'flow_lower_face'}, \code{'wells'},
\code{'river_leakage'}, \code{'recharge'}, \code{'drains'}, \code{'head_dep_bounds'} or any other description as written by MODFLOW.
 Additionally, flux can be \code{'darcy'}, in which case Darcy fluxes are computed by \code{\link{rmf_convert_cbc_to_darcy}} and the magnitude is plotted. If
 type = 'vector', the x, y and z components are used to determine the direction of the arrows.
 
 There are two ways to specify which time step to plot. The \code{l} argument can be specified which represents the total time step number. 
 The other option is to specify both \code{kper} & \code{kstp} which specify the stress-period and corresponding time step in that stress-period.
 A negative \code{kstp} will plot the final time step of the stress-period.
 
 If no output is written for the specified time step, as controlled by the Output Control file in MODFLOW, an error is thrown.
}
