% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmf_plot.chd}
\alias{rmf_plot.chd}
\title{Plot a RMODFLOW chd object}
\usage{
\method{rmf_plot}{chd}(chd, dis, kper = NULL, variable = "id",
  i = NULL, j = NULL, k = NULL, active_only = TRUE, fun = mean,
  ...)
}
\arguments{
\item{chd}{an \code{RMODFLOW} chd object}

\item{dis}{a \code{RMODFLOW} dis object}

\item{kper}{integer specifying the stress-period to plot}

\item{variable}{single character or numeric indicating which column of \code{chd$data} to plot. Defaults to 'id', which plots the locations of the cells.}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{active_only}{logical; indicating if only the active cells should be plotted. Non-active cells are set to NA. Defaults to TRUE.}

\item{fun}{function to compute values in the case multiple values are defined for the same MODFLOW cell. Typically either \code{mean} or \code{sum}. Defaults to mean for variables 'shead' & 'ehead'.}

\item{...}{additional arguments passed to \code{\link{rmf_plot.rmf_3d_array}}}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
Plot a RMODFLOW chd object
}
