% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmf_plot.ddn}
\alias{rmf_plot.ddn}
\title{Plot a MODFLOW drawdown file object}
\usage{
\method{rmf_plot}{ddn}(
  ddn,
  dis,
  i = NULL,
  j = NULL,
  k = NULL,
  l = NULL,
  kper = NULL,
  kstp = NULL,
  ...
)
}
\arguments{
\item{ddn}{\code{RMODFLOW} ddn object}

\item{dis}{\code{RMODFLOW} dis object}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{l}{time step number to plot; defaults to plotting the final time step.}

\item{kper}{integer specifying the stress-period. Use in conjunction with kstp. See details.}

\item{kstp}{integer specifying the time step of kper. Use in conjunction with kper. See details.}

\item{...}{additional parameters passed to \code{\link{rmf_plot.rmf_4d_array}}}
}
\value{
ggplot2 object or layer
}
\description{
Plot a MODFLOW drawdown file object
}
\details{
There are two ways to specify which time step to plot. The \code{l} argument can be specified which represents the total time step number.
The other option is to specify both \code{kper} & \code{kstp} which specify the stress-period and corresponding time step in that stress-period.
A negative \code{kstp} will plot the final time step of the stress-period.

If no output is written for the specified time step, as controlled by the Output Control file in MODFLOW, an error is thrown.
}
