% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-plot.evt.R
\name{rmf_plot.evt}
\alias{rmf_plot.evt}
\title{Plot a 2D section of an RMODFLOW evt object}
\usage{
\method{rmf_plot}{evt}(evt, dis, mlt = NULL, zon = NULL, all_parm = F,
  variable = "evt", instnum = NULL, l = NULL, sp = NULL, ...)
}
\arguments{
\item{evt}{an \code{RMODFLOW} evt object}

\item{dis}{an \code{RMODFLOW} dis object}

\item{mlt}{optional; an \code{RMODFLOW} mlt object; used for plotting parameter data}

\item{zon}{optional; an \code{RMODFLOW} zon object; used for plotting parameter data}

\item{all_parm}{logical, should all parameters defined by plotted (i.e. indepedent of stress periods); defaults to FALSE}

\item{variable}{character, what data should be plotted. Possible values are: "identity" (plots the evapotranspiration cells' locations), "evt" (default) which defines the values of maximum ET, "surf" (for stress period data), "exdp" (for stress period data), "mltnam" (for parameter data), "zonnam" (for parameter data), "iz" (for parameter data), "parnam" (for parameter data), "instnam" (for time-varying parameter data) and "parval" (for parameter data); defaults to "evt"}

\item{instnum}{numeric vector of length \code{npevt} holding the instance numbers for each time-varying parameter which need to be plotted. Only one instance per parameter is allowed. If a certain parameter \code{i} is not time-varying, specify instnum[i] as '1'; defaults to NULL}

\item{l}{time step number for selecting which stress period to plot; defaults to NULL (last stress period)}

\item{sp}{optional stress period number to plot; will override the stress period calculated from \code{l}; defaults to NULL}

\item{...}{additional arguments passed to \code{rmf_plot.rmf-3d-array}}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
\code{rmf_plot.evt} plots a 2D section of an \code{RMODFLOW} evt object using \code{rmf_plot.rmf-3d-array}
}
