% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmf_plot.ghb}
\alias{rmf_plot.ghb}
\title{Plot a RMODFLOW ghb object}
\usage{
\method{rmf_plot}{ghb}(
  ghb,
  dis,
  kper = NULL,
  variable = "id",
  i = NULL,
  j = NULL,
  k = NULL,
  active_only = TRUE,
  fun = ifelse(variable == "bhead", mean, sum),
  add = FALSE,
  ...
)
}
\arguments{
\item{ghb}{an \code{RMODFLOW} ghb object}

\item{dis}{a \code{RMODFLOW} dis object}

\item{kper}{integer specifying the stress-period to plot}

\item{variable}{single character or numeric indicating which column of \code{ghb$data} to plot. Defaults to 'id', which plots the locations of the cells.}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{active_only}{logical; indicating if only the active cells should be plotted. Non-active cells are set to NA. Defaults to TRUE.}

\item{fun}{function to compute values in the case multiple values are defined for the same MODFLOW cell. Typically either \code{mean} or \code{sum}. Defaults to mean for variable 'bhead' and sum for variable 'cond'.}

\item{add}{logical; if TRUE, provide ggplot2 layers instead of object, or add 3D plot to existing rgl device; defaults to FALSE}

\item{...}{additional arguments passed to \code{\link{rmf_plot.rmf_3d_array}}}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
Plot a RMODFLOW ghb object
}
