% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmf_plot.hed}
\alias{rmf_plot.hed}
\title{Plot a MODFLOW head file object}
\usage{
\method{rmf_plot}{hed}(
  hed,
  dis,
  i = NULL,
  j = NULL,
  k = NULL,
  l = NULL,
  nstp = NULL,
  kper = NULL,
  kstp = NULL,
  saturated = FALSE,
  type = "fill",
  gridlines = FALSE,
  add = FALSE,
  ...
)
}
\arguments{
\item{hed}{\code{RMODFLOW} hed object}

\item{dis}{\code{RMODFLOW} dis object}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{l}{time step number to plot (index for 4th dimension); defaults to plotting the final time step. See details.}

\item{nstp}{integer specifying the flow time step to plot. See details.}

\item{kper}{integer specifying the stress-period. Use in conjunction with kstp. See details.}

\item{kstp}{integer specifying the time step of kper. Use in conjunction with kper. See details.}

\item{saturated}{logical indicating if the saturated grid should be used. Defaults to FALSE. See details.}

\item{...}{additional parameters passed to \code{\link{rmf_plot.rmf_4d_array}}}
}
\value{
ggplot2 object or layer
}
\description{
Plot a MODFLOW head file object
}
\details{
There are three ways to specify which time step to plot.
The \code{l} argument can be specified which represents the dimension index. This can be smaller than the total number of time steps since output may not be written for all steps.
The second option is to specify \code{nstp} which is the total time step number. Since output at every time step is not always written, a specified nstp value might not be available in the array.
The third option is to specify both \code{kper} & \code{kstp} which specify the stress-period and corresponding time step in that stress-period.
A negative \code{kstp} will plot the final time step of the stress-period.

If \code{saturated} is TRUE, the saturated grid is plotted as given by \code{\link{rmf_convert_dis_to_saturated_dis}}, which might be useful
for cross-sections with \code{grid = TRUE}.

If no output is written for the specified time step, as controlled by the Output Control file in MODFLOW, an error is thrown.
}
