% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmf_plot.hfb}
\alias{rmf_plot.hfb}
\title{Plot a RMODFLOW hfb object}
\usage{
\method{rmf_plot}{hfb}(
  hfb,
  dis,
  variable = "id",
  all_hfb = FALSE,
  i = NULL,
  j = NULL,
  k = NULL,
  type = "fill",
  active_only = TRUE,
  fun = ifelse(variable == "hydchr", sum, mean),
  prj = NULL,
  crs = NULL,
  size = 1,
  colour = "black",
  crop = FALSE,
  add = FALSE,
  ...
)
}
\arguments{
\item{hfb}{an \code{RMODFLOW} hfb object}

\item{dis}{a \code{RMODFLOW} dis object}

\item{variable}{single character or numeric indicating which column of \code{hfb$data} to plot. Defaults to 'id', which plots the locations of the cells.}

\item{all_hfb}{logical; should all horizontal-flow barriers be plotted (TRUE) or only the ones that are active in the simulation (FALSE; default)}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{active_only}{logical; indicating if only the active cells should be plotted. Non-active cells are set to NA. Defaults to TRUE.}

\item{fun}{function to compute values in the case multiple values are defined for the same MODFLOW cell. Typically either \code{mean} or \code{sum}. Defaults to sum for variable 'hydchr'}

\item{...}{additional arguments passed to \code{\link{rmf_plot.rmf_3d_array}}}
}
\value{
ggplot2 object or layer
}
\description{
Plot a RMODFLOW hfb object
}
