% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmf_plot.huf}
\alias{rmf_plot.huf}
\title{Plot a 2D section through a MODFLOW 3D array}
\usage{
\method{rmf_plot}{huf}(
  huf,
  dis,
  i = NULL,
  j = NULL,
  k = NULL,
  hgu = NULL,
  bas = NULL,
  mask = rmf_convert_huf_to_mask(huf = huf, dis = dis, bas = bas),
  colour_palette = rmfi_rev_rainbow,
  nlevels = 7,
  levels = setNames(huf$hgunam, 1:huf$nhuf),
  type = "factor",
  gridlines = FALSE,
  ...
)
}
\arguments{
\item{huf}{an object of class huf}

\item{dis}{discretization file object}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{hgu}{character or integer of hgu to plot}

\item{bas}{basic file object; optional}

\item{mask}{a 3D array with 0 or F indicating inactive cells optional;}

\item{colour_palette}{a colour palette for imaging continuous array values.}

\item{nlevels}{number of levels for the colour scale; defaults to 7}

\item{levels}{labels that should be used on the factor legend; huf$hgunam is used by default}

\item{type}{plot type: 'fill', 'factor' (default) or 'grid'}

\item{gridlines}{logical; should grid lines be plotted? Alternatively, provide colour of the grid lines or set to 'huf' which plots the outline of the hgu's}

\item{...}{parameters provided to plot.rmf_2d_array}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
\code{rmf_plot.huf} plots a 2D section through a MODFLOW 3D array.
}
