% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmf_plot.modflow}
\alias{rmf_plot.modflow}
\title{Plot a 2D section through a MODFLOW model}
\usage{
\method{rmf_plot}{modflow}(
  modflow,
  i = NULL,
  j = NULL,
  k = NULL,
  omit = NULL,
  to_k = FALSE,
  gridlines = TRUE,
  prj = rmf_get_prj(modflow),
  legend = "",
  ...
)
}
\arguments{
\item{modflow}{\code{RMODFLOW} modflow model}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{omit}{character vector with MODFLOW ftypes specifying which packages to omit from the plot. Defaults to plotting all packages.}

\item{to_k}{logical; if a layer section is plotted (\code{k} is given), reproject all boundary conditions (not IBOUND) to the layer? Defaults to FALSE.}

\item{gridlines}{logical; should grid lines be plotted? alternatively, provide colour of the grid lines. Defaults to TRUE.}

\item{prj}{projection file object}

\item{legend}{character denoting the legend of the plot. Defaults to \code{''}.}

\item{...}{parameters provided to plot.rmf_3d_array}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
\code{rmf_plot.modflow} plots a 2D section through a MODFLOW model showing the IBOUND array and discrete boundary conditions.
}
\details{
Only the IBOUND and the locations of the discrete boundary condition packages are plotted.
\code{to_k} only reprojects boundary condition cells, not IBOUND cells, to layer k. It is not used when \code{k} is not defined. This is useful
to plot all boundary condition cells even if they are not present in the current layer k, e.g. wells in deeper layers.
}
\examples{
m <- rmf_example_file('water-supply-problem.nam') \%>\% rmf_read(verbose = FALSE)
rmf_plot(m, k = 1)
rmf_plot(m, j = 33, gridlines = FALSE)

m2 <- rmf_example_file('rocky-mountain-arsenal.nam') \%>\% rmf_read(verbose = FALSE)
rmf_plot(m2, k = 1, legend = 'Type')
rmf_plot(m2, k = 1, legend = 'Type', omit = 'WEL')

m3 <- rmf_example_file('example-model.nam') \%>\% rmf_read(verbose = FALSE)
rmf_plot(m3, k = 1)
rmf_plot(m3, k = 2)

# reproject boundary conditions (all in layer 1) to layer 2
rmf_plot(m3, k = 2, to_k = TRUE)
}
