% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmf_plot.rch}
\alias{rmf_plot.rch}
\title{Plot a RMODFLOW rch object}
\usage{
\method{rmf_plot}{rch}(rch, dis, kper = NULL, variable = "recharge",
  ...)
}
\arguments{
\item{rch}{an \code{RMODFLOW} rch object}

\item{dis}{a \code{RMODFLOW} dis object}

\item{kper}{integer specifying the stress-period to plot}

\item{variable}{character specifying which variable to plot. Possible values are 'recharge' (default) and 'irch' if defined.}

\item{...}{additional arguments passed to \code{\link{rmf_plot.rmf_2d_array}}}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
Plot a RMODFLOW rch object
}
