% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmf_plot.rmf_analyze}
\alias{rmf_plot.rmf_analyze}
\title{Plot a MODFLOW sensitivity analysis object}
\usage{
\method{rmf_plot}{rmf_analyze}(analysis, type = "css")
}
\arguments{
\item{analysis}{sensitivity analysis object from \link{rmf_analyze}}

\item{type}{plot type: 'css' or 'dss'}
}
\description{
Plot a MODFLOW sensitivity analysis object
}
