% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-plot.sen.R
\name{rmf_plot.sen}
\alias{rmf_plot.sen}
\title{Plot a MODFLOW sensitivity analysis object}
\usage{
\method{rmf_plot}{sen}(sen, type = "css")
}
\arguments{
\item{sen}{sensitivity analysis object}

\item{plot}{type: 'css' or 'dss'}
}
\description{
Plot a MODFLOW sensitivity analysis object
}
