% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflow.R
\name{rmf_read}
\alias{rmf_read}
\title{Read a MODFLOW model}
\usage{
rmf_read(
  file = {
     cat("Please select nam file ...\\n")
     file.choose()
 },
  output = FALSE,
  precision = "single",
  verbose = TRUE
)
}
\arguments{
\item{file}{NAME file; typically '*.nam'}

\item{output}{logical; should output also be read. Defaults to FALSE.}

\item{precision}{either \code{'single'} or \code{'double'}. Specifies the precision of binary files. Defaults to \code{'single'}.}

\item{verbose}{logical; should information on reading files be printed to the console ? Defaults to TRUE.}
}
\value{
a \code{modflow} object which is a list containing all MODFLOW packages and optionally, model output
}
\description{
\code{rmf_read} reads in a MODFLOW model and returns it as a \code{modflow} object
}
\seealso{
\code{\link{rmf_create}}, \code{\link{rmf_write}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html}
}
