% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_read_array}
\alias{rmf_read_array}
\title{Read a MODFLOW array from a separate file.}
\usage{
rmf_read_array(
  file,
  nrow = NULL,
  ncol = NULL,
  nlay = 1,
  nstp = 1,
  binary = FALSE,
  kper = NULL,
  integer = FALSE,
  header = binary,
  precision = "single"
)
}
\arguments{
\item{file}{filename to read the array from}

\item{nrow}{number of rows in the array}

\item{ncol}{number of columns in the array}

\item{nlay}{number of layers in the array that should be read (3th dimension); defaults to 1}

\item{nstp}{number of timesteps in the array that should be read (4th dimension); defaults to 1}

\item{binary}{logical; is the array read from a binary file.}

\item{kper}{integer vector specifying the stress periods in which the array is active. Defaults to \code{NULL}}

\item{integer}{logical; does the array hold integer values. Only used for binary files. Might not work optimally.}

\item{header}{logical; should a MODFLOW style header be read for the array (see 'Details'). Defaults to TRUE if \code{binary = TRUE} and FALSE otherwise.}

\item{precision}{character: either \code{'single'} (default) or \code{'double'}. Denotes the precision of the binary file.}
}
\value{
a rmf_array with optional attributes if a header was read.
}
\description{
\code{rmf_read_array} reads a MODFLOW array from a separate file. Binary and ASCII formats are supported
}
\details{
\code{nrow}, \code{ncol}, \code{nlay}, \code{nstp} have to be specified if header is FALSE. They are used to dimension the array.

The \code{integer} flag is only used when reading binary files.

The header file consists of the following elements:
\code{KSTP}, \code{KPER},\code{PERTIM},\code{DESC},\code{NCOL}, \code{NROW}, \code{ILAY} and \code{FMTIN}
(see \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?frequently_asked_questions.htm} for their respective meaning.)
The \code{FMTIN} element is not read for binary files.
If a header is read, the values are set as attributes to the array.
}
\seealso{
\code{\link{rmf_write_array}}
}
