% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-read-balance.R
\name{rmf_read_balance}
\alias{rmf_read_balance}
\title{Reads the volumetric budget from a MODFLOW listing file}
\usage{
rmf_read_balance(list_path = paste(getwd(), "output.lst", sep = "/"))
}
\arguments{
\item{list_path}{path to the listing file; typically '*.lst'; defaults to 'output.lst' in the working directory}
}
\value{
an object of class balance which is a list with the data frame elements holding the volumetric balance values
}
\description{
\code{rmf_read_balance} reads in a volumetric budget from a MODFLOW listing file and returns it as a list with data frame elements
}
