% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-bcf.R
\name{rmf_read_bcf}
\alias{rmf_read_bcf}
\title{Read a MODFLOW bcf file}
\usage{
rmf_read_bcf(
  file = {     cat("Please select bcf file ...\\n")     file.choose() },
  dis = {     cat("Please select corresponding dis file ...\\n")    
    rmf_read_dis(file.choose()) },
  ...
)
}
\arguments{
\item{file}{filename; typically '*_bcf'}

\item{dis}{an \code{RMODFLOW} dis object}

\item{...}{arguments passed to \code{rmfi_parse_array} and \code{rmfi_parse_variables}. Can be ignored when input is 'free' format and arrays are INTERNAL or CONSTANT.}
}
\value{
an \code{RMODFLOW} bcf object
}
\description{
\code{rmf_read_bcf} reads in a MODFLOW block-centered flow file and returns it as an \code{RMODFLOW} bcf object
}
\seealso{
\code{\link{rmf_write_bcf}}, \code{\link{rmf_create_bcf}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?bcf.htm}
}
