% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-oc.R
\name{rmf_read_cbc}
\alias{rmf_read_cbc}
\title{Read a MODFLOW cell-by-cell budget file}
\usage{
rmf_read_cbc(file = {    
  cat("Please select cell-by-cell budget file ...\\n")     file.choose() },
  dis = {     cat("Please select corresponding dis file ...\\n")    
  rmf_read_dis(file.choose()) }, huf = NULL, oc = NULL,
  precision = "single", fluxes = "all")
}
\arguments{
\item{file}{filename; typically '*.cbc'}

\item{dis}{dis object.}

\item{huf}{huf object; optional. Provide only if huf heads are being read and \code{dis} is not NULL. See details.}

\item{oc}{oc object; optional. See details.}

\item{precision}{either \code{'single'} or \code{'double'}. Specifies the precision of the binary file.}

\item{fluxes}{character; denotes which fluxes to read. Defaults to reading all fluxes. See details.}
}
\value{
object of class \code{cbc} which is a list consisting of named rmf_arrays and/or data.frames. The names of the elements correspond to the fluxes.
}
\description{
\code{rmf_read_cbc} reads in a MODFLOW cell-by-cell budget file
}
\details{
Fluxes include \code{'constant_head'}, \code{'storage'}, \code{'flow_right_face'}, \code{'flow_front_face'}, \code{'flow_lower_face'}, \code{'wells'},
\code{'river_leakage'}, \code{'recharge'}, \code{'drains'}, \code{'head_dep_bounds'} or any other description as written by MODFLOW.
 
If a \code{oc} object is supplied, a rmf_array of dimensions NROW x NCOL x NLAY x sum(NSTP) is created and filled. Time steps for which no output is given are filled with \code{NA}.
If no \code{oc} object is supplied, a rmf_array of dimensions NROW x NCOL x NLAY is read and binded at each time step for which output is written. 
The resulting dimensions of the final array are NROW x NCOL x NLAY x STPS where STPS are timesteps for which output is saved. 

If flows are interpolated to huf units, a \code{huf} object is to be supplied as well to dimension the array. This will only affect the constant-head and cell flow terms.
The final array will have NHUF layers instead of NLAY.
}
