% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-read-de4.R
\name{rmf_read_de4}
\alias{rmf_read_de4}
\title{Read a MODFLOW direct solver file}
\usage{
rmf_read_de4(file = {     cat("Please select direct solver file ...\\n") 
     file.choose() })
}
\arguments{
\item{file}{filename; typically '*_de4'}
}
\value{
\code{RMODFLOW} de4 object
}
\description{
\code{rmf_read_de4} reads in a MODFLOW direct solver file and returns it as an \code{RMODFLOW} de4 object
}
\seealso{
\code{\link{rmf_write_de4}}, \code{\link{rmf_create_de4}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?de4.htm}
}
