% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-drn.R
\name{rmf_read_drn}
\alias{rmf_read_drn}
\title{Read a MODFLOW drain file}
\usage{
rmf_read_drn(file = {     cat("Please select drain file ...\\n")    
  file.choose() }, dis = {    
  cat("Please select corresponding dis file ...\\n")    
  rmf_read_dis(file.choose()) }, ...)
}
\arguments{
\item{file}{filename; typically '*.drn'}

\item{dis}{an \code{RMODFLOW} dis object}

\item{...}{arguments passed to \code{rmfi_parse_variables} and \code{rmfi_parse_list}.}
}
\value{
\code{RMODFLOW} drn object
}
\description{
\code{rmf_read_drn} reads in a MODFLOW drain file and returns it as an \code{RMODFLOW} drn object.
}
\seealso{
\code{\link{rmf_write_drn}}, \code{\link{rmf_create_drn}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?drn.htm}
}
