% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-read-evt.R
\name{rmf_read_evt}
\alias{rmf_read_evt}
\title{Read a MODFLOW evapotranspiration file}
\usage{
rmf_read_evt(file = {    
  cat("Please select evapotranspiration file ...\\n")     file.choose() },
  dis = {     cat("Please select corresponding dis file ...\\n")    
  rmf_read_dis(file.choose()) }, ...)
}
\arguments{
\item{file}{filename; typically '*.evt'}

\item{dis}{an \code{RMODFLOW} dis object}

\item{...}{arguments passed to \code{rmfi_parse_array}. Can be ignored when input arrays are free-format and INTERNAL or CONSTANT.}
}
\value{
\code{RMODFLOW} evt object
}
\description{
\code{rmf_read_evt} reads in a MODFLOW evapotranspiration file and returns it as an \code{RMODFLOW} evt object.
}
\seealso{
\code{\link{rmf_write_evt}}, \code{\link{rmf_create_evt}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?evt.htm}
}
