% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-ghb.R
\name{rmf_read_ghb}
\alias{rmf_read_ghb}
\title{Read a MODFLOW general-head boundary file}
\usage{
rmf_read_ghb(
  file = {     cat("Please select general-head boundary file ...\\n")     file.choose()
    },
  dis = {     cat("Please select corresponding dis file ...\\n")    
    rmf_read_dis(file.choose()) },
  ...
)
}
\arguments{
\item{file}{filename; typically '*.ghb'}

\item{dis}{an \code{RMODFLOW} dis object}

\item{...}{arguments passed to \code{rmfi_parse_variables} and \code{rmfi_parse_list}.}
}
\value{
\code{RMODFLOW} ghb object
}
\description{
\code{rmf_read_ghb} reads in a MODFLOW general-head boundary file and returns it as an \code{RMODFLOW} ghb object.
}
\seealso{
\code{\link{rmf_write_ghb}}, \code{\link{rmf_create_ghb}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?ghb.htm}
}
