% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-gmg.R
\name{rmf_read_gmg}
\alias{rmf_read_gmg}
\title{Read a MODFLOW Geometric Multigrid Solver file}
\usage{
rmf_read_gmg(
  file = {     cat("Please select gmg file ...\\n")     file.choose() },
  ...
)
}
\arguments{
\item{file}{filename; typically '*.gmg'}

\item{...}{ignored}
}
\value{
object of class gmg
}
\description{
\code{rmf_read_gmg} reads in a MODFLOW Geometric Multigrid Solver file and returns it as an \code{\link{RMODFLOW}} gmg object.
}
\seealso{
\code{\link{rmf_write_gmg}}, \code{\link{rmf_create_gmg}} and \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?gmg.htm}
}
