% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-read-gms-2d-grid.R
\name{rmf_read_gms_2d_grid}
\alias{rmf_read_gms_2d_grid}
\alias{rmf_read_gms_2d_grid}
\title{Read a GMS 2D grid file}
\usage{
rmf_read_gms_2d_grid(...)

rmf_read_gms_2d_grid(...)
}
\arguments{
\item{file}{filename}
}
\value{
object of class gms2dgrid
}
\description{
\code{read_gms_2d_grid} reads in a GMS 2D grid file and returns it as an \code{\link{RMODFLOW}} gms2dgrid object.
}
\section{Functions}{
\itemize{
\item \code{rmf_read_gms_2d_grid}: Deprecated function name
}}

