% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-read-hfb.R
\name{rmf_read_hfb}
\alias{rmf_read_hfb}
\title{Read a MODFLOW horizontal flow barrier file}
\usage{
rmf_read_hfb(file = {    
  cat("Please select horizontal flow barrier file ...\\n")     file.choose() })
}
\arguments{
\item{file}{filename; typically '*.hfb'}
}
\value{
\code{RMODFLOW} hfb object
}
\description{
\code{rmf_read_hfb} reads in a MODFLOW horizontal flow barrier file and returns it as an \code{RMODFLOW} hfb object.
}
\seealso{
\code{\link{rmf_write_hfb}}, \code{\link{rmf_create_hfb}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?hfb6.htm}
}
