% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-lmt.R
\name{rmf_read_lmt}
\alias{rmf_read_lmt}
\title{Read a MODFLOW Link-MT3DMS Package file}
\usage{
rmf_read_lmt(
  file = {     cat("Please select lmt file ...\\n")     file.choose() },
  ...
)
}
\arguments{
\item{file}{filename; typically '*.lmt'}

\item{...}{ignored}
}
\value{
object of class lmt
}
\description{
Read a MODFLOW Link-MT3DMS Package file
}
\examples{
file <- rmf_example_file('rocky-mountain-arsenal.lmt')
rmf_read_lmt(file)

}
\seealso{
\code{\link{rmf_create_lmt}}, \code{\link{rmf_write_lmt}} and \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?lmt6.htm}
}
