% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-read-lpf.R
\name{rmf_read_lpf}
\alias{rmf_read_lpf}
\alias{read_lpf}
\title{Read a MODFLOW layer-property flow file}
\usage{
rmf_read_lpf(file = {     cat("Please select lpf file ...\\n")    
  file.choose() })

read_lpf(...)
}
\arguments{
\item{file}{filename; typically '*.lpf'}
}
\value{
object of class lpf
}
\description{
\code{read_lpf} reads in a MODFLOW layer property file and returns it as an \code{\link{RMODFLOW}} lpf object.
}
\section{Functions}{
\itemize{
\item \code{read_lpf}: Deprecated function name
}}

\seealso{
\code{\link{write_lpf}}, \code{\link{create_lpf}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?lpf.htm}
}
