% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-nam.R
\name{rmf_read_nam}
\alias{rmf_read_nam}
\alias{read_nam}
\title{Read a MODFLOW name file}
\usage{
rmf_read_nam(file = {     cat("Please select nam file ...\\n")    
  file.choose() })

read_nam(...)
}
\arguments{
\item{file}{filename; typically '*.nam'}
}
\value{
object of class nam
}
\description{
\code{rmf_read_nam} reads in a MODFLOW name file and returns it as an \code{\link{RMODFLOW}} nam object.
}
\section{Functions}{
\itemize{
\item \code{read_nam}: Deprecated function name
}}

