% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-oc.R
\name{rmf_read_oc}
\alias{rmf_read_oc}
\title{Read a MODFLOW output control option file}
\usage{
rmf_read_oc(
  file = {
     cat("Please select oc file ...\\n")
     file.choose()
 },
  dis = {
     cat("Please select corresponding dis file ...\\n")
    
    rmf_read_dis(file.choose())
 },
  ...
)
}
\arguments{
\item{file}{filename; typically '*.oc'}

\item{dis}{an \code{RMODFLOW} dis object. Used when reading OC specified using numeric codes.}

\item{...}{arguments passed to \code{rmfi_parse_variables}. Can be ignored when input is 'free' format.}
}
\value{
object of class oc
}
\description{
\code{read_oc} reads in a MODFLOW output control option file and returns it as an \code{\link{RMODFLOW}} oc object.
}
\seealso{
\code{\link{rmf_write_oc}}, \code{\link{rmf_create_oc}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?oc.htm}
}
