% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-pval.R
\name{rmf_read_pvl}
\alias{rmf_read_pvl}
\title{Read a MODFLOW parameter value file}
\usage{
rmf_read_pvl(
  file = {     cat("Please select pvl file ...\\n")     file.choose() },
  read_all = FALSE
)
}
\arguments{
\item{file}{filename; typically '*.pvl'}

\item{read_all}{logical, indicating if \code{np} parameters should be read, or the full parameter table (only relevant if external codes use the pvl file for storing additional parameters).}
}
\value{
object of class pvl
}
\description{
\code{rmf_read_pvl} reads in a MODFLOW parameter value file and returns it as an \code{\link{RMODFLOW}} pvl object.
}
\seealso{
\code{\link{rmf_create_pvl}}, \code{\link{rmf_write_pvl}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?pvl.htm}
}
