% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-rch.R
\name{rmf_read_rch}
\alias{rmf_read_rch}
\title{Read a MODFLOW recharge file}
\usage{
rmf_read_rch(file = {     cat("Please select rch file ...\\n")    
  file.choose() }, dis = {    
  cat("Please select corresponding dis file ...\\n")    
  rmf_read_dis(file.choose()) }, ...)
}
\arguments{
\item{file}{filename; typically '*.rch'}

\item{dis}{an \code{RMODFLOW} dis object}

\item{...}{arguments passed to \code{rmfi_parse_array}. Can be ignored when input arrays are free-format and INTERNAL or CONSTANT.}
}
\value{
\code{RMODFLOW} rch object
}
\description{
\code{rmf_read_rch} reads in a MODFLOW recharge file and returns it as an \code{RMODFLOW} rch object.
}
\seealso{
\code{\link{rmf_write_rch}}, \code{\link{rmf_create_rch}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?rch.htm}
}
