% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-read-riv.R
\name{rmf_read_riv}
\alias{rmf_read_riv}
\title{Read a MODFLOW river file}
\usage{
rmf_read_riv(file = {     cat("Please select river file ...\\n")    
  file.choose() }, dis = {    
  cat("Please select corresponding dis file ...\\n")    
  rmf_read_dis(file.choose()) })
}
\arguments{
\item{file}{filename; typically '*.riv'}

\item{dis}{an \code{RMODFLOW} dis object}
}
\value{
\code{RMODFLOW} riv object
}
\description{
\code{rmf_read_riv} reads in a MODFLOW river file and returns it as an \code{RMODFLOW} riv object.
}
\seealso{
\code{\link{rmf_write_riv}}, \code{\link{rmf_create_riv}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?riv.htm}
}
