% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rmf_read_usgs_model_reference}
\alias{rmf_read_usgs_model_reference}
\title{Read RMODFLOW projection information from a USGS model reference file}
\usage{
rmf_read_usgs_model_reference(
  file = {     cat("Please select usgs.model.reference file ...\\n")     file.choose()
    },
  dis = {     cat("Please select corresponding dis file ...\\n")    
    rmf_read_dis(file.choose()) }
)
}
\arguments{
\item{file}{path to the USGS model reference file; typically "usgs.model.reference"}

\item{dis}{\code{RMODFLOW} dis object}
}
\value{
a \code{prj} object
}
\description{
Read RMODFLOW projection information from a USGS model reference file
}
