% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-wel.R
\name{rmf_read_wel}
\alias{rmf_read_wel}
\title{Read a MODFLOW well file}
\usage{
rmf_read_wel(file = {     cat("Please select well file ...\\n")    
  file.choose() }, dis = {    
  cat("Please select corresponding dis file ...\\n")    
  rmf_read_dis(file.choose()) }, ...)
}
\arguments{
\item{file}{filename; typically '*.wel'}

\item{dis}{an \code{RMODFLOW} dis object}

\item{...}{arguments passed to \code{rmfi_parse_variables} and \code{rmfi_parse_list}.}
}
\value{
\code{RMODFLOW} wel object
}
\description{
\code{rmf_read_wel} reads in a MODFLOW well file and returns it as an \code{RMODFLOW} wel object.
}
\seealso{
\code{\link{rmf_write_wel}}, \code{\link{rmf_create_wel}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?wel.htm}
}
