% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-zon.R
\name{rmf_read_zon}
\alias{rmf_read_zon}
\title{Read a MODFLOW zone file}
\usage{
rmf_read_zon(file = {     cat("Please select zon file ...\\n")    
  file.choose() }, dis = {     cat("Please select dis file ...\\n")    
  rmf_read_dis(file.choose()) }, ...)
}
\arguments{
\item{file}{filename; typically '*.zon'}

\item{dis}{discretization file object; defaults to that with the same filename but with extension '.dis'}

\item{...}{arguments passed to \code{rmfi_parse_array}. Can be ignored when input arrays are free-format and INTERNAL or CONSTANT.}
}
\value{
\code{RMODFLOW} zon object
}
\description{
\code{rmf_read_zon} reads in a MODFLOW zone file and returns it as an \code{RMODFLOW} zon object
}
\seealso{
\code{\link{rmf_write_zon}}, \code{\link{rmf_create_zon}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?zone.htm}
}
