% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{rmf_run_modflow.character}
\alias{rmf_run_modflow.character}
\alias{rmf_run_modflow.modflow}
\alias{rmf_run_modflow}
\title{Run a MODFLOW model}
\usage{
\method{rmf_run_modflow}{character}(file, version = 2005,
  executable = NULL, par = NULL)

\method{rmf_run_modflow}{modflow}(modflow, version = 2005,
  executable = NULL, par = NULL)

rmf_run_modflow(...)
}
\arguments{
\item{file}{path to name file; typically '*.nam'}

\item{version}{MODFLOW version to use; 2005 (default) is currently the only option}

\item{executable}{name of the MODFLOW executable to use; if not provided, the executable distributed with RMODFLOW is used, corresponding to version, machine and sysname}

\item{par}{vector of parameter value file parameter values to run the model with}

\item{modflow}{modflow object}

\item{version}{MODFLOW version to use; 2005 (default) is currently the only option}

\item{executable}{name of the MODFLOW executable to use; if not provided, the executable distributed with RMODFLOW is used, corresponding to version, machine and sysname}

\item{par}{vector of parameter value file parameter values to run the model with}
}
\description{
\code{run_modflow} runs a MODFLOW model.

\code{run_modflow} runs a MODFLOW model.
}
