% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-run-sen.R
\name{rmf_run_sen}
\alias{rmf_run_sen}
\alias{run_sen}
\title{Run a MODFLOW model sensitivity analysis, based on the parameter value file}
\usage{
rmf_run_sen(file, executable = "mf2005", par = NULL, include = NULL)

run_sen(...)
}
\arguments{
\item{file}{path to name file; typically '*.nam'}

\item{executable}{name of the MODFLOW executable to use}

\item{par}{central parameter values (for all or only included parameters); parameter value file values are used if par is not provided}

\item{include}{logical vector indicating which parameters in the parameter value file to include in the sensitivity analysis}
}
\value{
sensitivity analysis results
}
\description{
\code{run_modflow_sen} performs a MODFLOW model sensitivity analysis.
}
\section{Functions}{
\itemize{
\item \code{run_sen}: Deprecated function name
}}

