% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_saturated_thickness}
\alias{rmf_saturated_thickness}
\title{Calculate saturated thicknesses}
\usage{
rmf_saturated_thickness(hed, dis, l = NULL, na_values = NULL)
}
\arguments{
\item{hed}{3d or 4d array object containing hydraulic head values for each cell}

\item{dis}{\code{RMODFLOW} dis object}

\item{l}{integer used to subset the 4th dimension of \code{hed}. If not supplied, the final time step is used.}

\item{na_values}{optional; specifies which \code{hed} values should be set to \code{NA}.}
}
\value{
a \code{rmf_3d_array} with saturated thicknesses for each cell.
}
\description{
Calculate saturated thicknesses
}
