% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-time-steps.R
\name{rmf_time_steps}
\alias{rmf_time_steps}
\title{Calculate the internal time step sequence of a transient MODFLOW model}
\usage{
rmf_time_steps(dis = NULL, perlen = NULL, tsmult = NULL,
  nstp = NULL, incl_ss = T)
}
\arguments{
\item{dis}{optional, an \code{RMODFLOW} dis object}

\item{perlen}{optional, only read when a \code{dis} object is not supplied; numeric vector of length \code{nper} specifying the stress period lengths}

\item{tsmult}{optional, only read when a \code{dis} object is not supplied; numeric vector of length \code{nper} specifying the time step multipliers}

\item{nstp}{optional, only read when a \code{dis} object is not supplied; numeric vector of length \code{nper} specifying the number of time steps in each stress period}

\item{incl_ss}{logical, only read when a \code{dis} object is supplied; should the lengths of steady-state stress periods in the \code{dis} object be incorporated in the calculation; defaults to TRUE}
}
\value{
a list holding the numeric vectors of the computed sequence of time step lengths and its cumulative sum
}
\description{
\code{rmf_time_steps} calculates the internal sequence of time steps of a transient MODFLOW model from either an \code{RMODFLOW} dis object or separate parameters
}
\seealso{
\url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?dis.htm}
}
