% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflow.R
\name{rmf_write}
\alias{rmf_write}
\title{Write a MODFLOW model}
\usage{
rmf_write(
  modflow,
  file = {
    
    cat("Please select nam file to overwrite or provide new filename ...\\n")
    
    file.choose()
 },
  exclude = NULL,
  suppress = FALSE,
  verbose = TRUE,
  iprn = -1
)
}
\arguments{
\item{modflow}{\code{RMODFLOW} modflow object}

\item{file}{filename of the name file to write}

\item{exclude}{character vector with packages names to exclude from the simulation. Defaults to NULL}

\item{suppress}{logical; remove non-supported (and thus not written) packages in the NAME file ? Defaults to FALSE}

\item{verbose}{logical; should information on writing files be printed to the console ? Defaults to TRUE.}

\item{iprn}{format code for printing arrays in the listing file; defaults to -1 (no printing)}
}
\value{
\code{NULL}
}
\description{
\code{rmf_write} writes all input packages in a \code{RMODFLOW} modflow object to a directory
}
\details{
All arrays use free-format headers INTERNAL or CONSTANT
All packages will be written according to the filenames (fname) defined in the nam object.
To prevent any files being overwritten, it is best to write to an empty directory.
}
\seealso{
\code{\link{rmf_create}}, \code{\link{rmf_write}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html}
}
