% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_write_array}
\alias{rmf_write_array}
\title{Write a MODFLOW array to a separate file.}
\usage{
rmf_write_array(array, file, append = FALSE, binary = FALSE,
  header = ifelse(binary, TRUE, FALSE), dis = NULL, desc = "HEAD",
  precision = "single", xsection = FALSE)
}
\arguments{
\item{array}{a 2D, 3D or 4D \code{rmf_array} to write.}

\item{file}{filename to write to}

\item{append}{logical; should the array be appended to the file}

\item{binary}{logical; should the array be written to a binary file}

\item{header}{logical; should a MODFLOW style header be written for the array (see 'Details'). Defaults to TRUE if binary is TRUE and FALSE otherwise.}

\item{dis}{optional \code{RMODFLOW} dis object. Used when \code{KPER}, \code{PERTIM} and \code{TOTIM} in the header should be exact.}

\item{desc}{character of maximum 16 characters. Used to set the \code{desc} element in the header. Default to \code{'HEAD'}}

\item{precision}{character; either \code{'single'} or \code{'double'}. Denotes the precision of the binary file.}

\item{xsection}{logical; does the array represent a NLAY x NCOL cross-section. See 'Details'.}
}
\value{
\code{NULL}
}
\description{
\code{rmf_write_array} writes a MODFLOW array to an output file. Binary and ASCII formats are supported
}
\details{
the header file consists of the following elements:
 \code{KSTP}, \code{KPER},\code{PERTIM},\code{DESC},\code{NCOL}, \code{NROW}, \code{ILAY} and \code{FMTIN}
 (see \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?frequently_asked_questions.htm} for their respective meaning.)
 If the array is 2D or 3D, \code{KSTP}, \code{KPER},\code{PERTIM} are always set to 1; otherwise they are all equal to the index of the 4th dimension.
 
 The \code{DESC} element must be read but it not used by MODFLOW itself. Users may want to specify a different \code{DESC} value when using the array with postprocessing software.
 
 The \code{FMTIN} element is not written for binary files. For ASCII files, \code{RMODFLOW} sets it as \code{NCOL * F}. Note that the ASCII format
 is irrelevant when using free-format reading and writing.
 
 \code{xsection} can be set to TRUE if the array represents a cross-section, i.e. the ibound or strt array in the
  \code{bas} file. The user must make sure the array is of dimension NLAY * NCOL. The sole function of \code{xsection} is to 
  set the \desc(ILAY) argument to -1 which promts MODFLOW to write slightly different information to the listing file. 
  \code{xsection} does not affect simulation results (assuming the array dimensions are correct)
}
\seealso{
\code{\link{rmf_read_array}}
}
