% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-bas.R
\name{rmf_write_bas}
\alias{rmf_write_bas}
\title{Write a MODFLOW basic file}
\usage{
rmf_write_bas(
  bas,
  file = {    
    cat("Please select bas file to overwrite or provide new filename ...\\n")    
    file.choose() },
  iprn = -1,
  ...
)
}
\arguments{
\item{bas}{an \code{\link{RMODFLOW}} bas object}

\item{file}{filename to write to; typically '*.bas'}

\item{iprn}{format code for printing arrays in the listing file; defaults to -1 (no printing)}

\item{...}{arguments passed to \code{rmfi_write_array}. Can be ignored when arrays are INTERNAL or CONSTANT.}
}
\value{
\code{NULL}
}
\description{
\code{rmf_write_bas} writes a MODFLOW basic file based on an \code{\link{RMODFLOW}} bas object.
}
\seealso{
\code{\link{rmf_read_bas}}, \code{\link{rmf_create_bas}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?bas6.htm}
}
