% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-write-bcf.R
\name{rmf_write_bcf}
\alias{rmf_write_bcf}
\title{Write a MODFLOW block-centered flow file}
\usage{
rmf_write_bcf(bcf, dis = rmf_read_dis(), file = {    
  cat("Please select bcf file to overwrite or provide new filename ...\\n")    
  file.choose() })
}
\arguments{
\item{bcf}{an \code{RMODFLOW} bcf object}

\item{dis}{an \code{RMODFLOW} dis object}

\item{file}{filename to write to; typically '*.bcf'}
}
\value{
\code{NULL}
}
\description{
\code{rmf_write_bcf} writes a MODFLOW block-centered flow file based on an \code{RMODFLOW} bcf object
}
\seealso{
\code{\link{rmf_read_bcf}}, \code{\link{rmf_create_bcf}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?bcf.htm}
}
