% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-chd.R
\name{rmf_write_chd}
\alias{rmf_write_chd}
\title{Write a MODFLOW time-variant specified-head file}
\usage{
rmf_write_chd(chd, dis = rmf_read_dis(), file = {    
  cat("Please choose chd file to overwrite or provide new filename ...\\n")
      file.choose() }, ...)
}
\arguments{
\item{chd}{an \code{RMODFLOW} chd object}

\item{dis}{an \code{RMODFLOW} dis object}

\item{file}{filename to write to; typically '*.chd'}

\item{...}{arguments passed to \code{rmfi_write_variables} when writing a fixed format file.}
}
\value{
\code{NULL}
}
\description{
\code{rmf_write_chd} writes a MODFLOW time-variant specified-head file based on an \code{RMODFLOW} chd object
}
\seealso{
\code{\link{rmf_read_chd}}, \code{\link{rmf_create_chd}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?chd.htm}
}
