% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-de4.R
\name{rmf_write_de4}
\alias{rmf_write_de4}
\title{Write a MODFLOW direct solver package}
\usage{
rmf_write_de4(de4, file = {    
  cat("Please choose de4 file to overwrite or provide new filename ...\\n")
      file.choose() }, ...)
}
\arguments{
\item{de4}{an \code{RMODFLOW} de4 object}

\item{file}{filename to write to; typically '*.de4'}

\item{...}{ignored}
}
\value{
\code{NULL}
}
\description{
\code{rmf_write_de4} writes a MODFLOW direct solver file based on an \code{RMODFLOW} de4 object
}
\seealso{
\code{\link{rmf_read_de4}}, \code{\link{rmf_create_de4}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?de4.htm}
}
