% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-write-drn.R
\name{rmf_write_drn}
\alias{rmf_write_drn}
\title{Write a MODFLOW drain file}
\usage{
rmf_write_drn(drn, dis = read_dis(), file = {    
  cat("Please choose drn file to overwrite or provide new filename ...\\n")    
  file.choose() })
}
\arguments{
\item{drn}{an \code{RMODFLOW} drn object}

\item{dis}{an \code{RMODFLOW} dis object}

\item{file}{filename to write to; typically '*.drn'}
}
\value{
\code{NULL}
}
\description{
\code{rmf_write_drn} writes a MODFLOW drain file based on an \code{RMODFLOW} drn object
}
\seealso{
\code{\link{rmf_read_drn}}, \code{\link{rmf_create_drn}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?drn.htm}
}
