% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-write-evt.R
\name{rmf_write_evt}
\alias{rmf_write_evt}
\title{Write a MODFLOW evapotranspiration file}
\usage{
rmf_write_evt(evt, dis = rmf_read_dis(), file = {    
  cat("Please select evt file to overwrite or provide new filename ...\\n")    
  file.choose() })
}
\arguments{
\item{evt}{an \code{RMODFLOW} evt object}

\item{dis}{an \code{RMODFLOW} dis object}

\item{file}{filename to write to; typically '*.evt'}
}
\value{
\code{NULL}
}
\description{
\code{rmf_write_evt} writes a MODFLOW evapotranspiration file based on an \code{RMODFLOW} evt object
}
\seealso{
\code{\link{rmf_read_evt}}, \code{\link{rmf_create_evt}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?evt.htm}
}
