% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-ghb.R
\name{rmf_write_ghb}
\alias{rmf_write_ghb}
\title{Write a MODFLOW general-head boundary file}
\usage{
rmf_write_ghb(
  ghb,
  dis = rmf_read_dis(),
  file = {
    
    cat("Please choose ghb file to overwrite or provide new filename ...\\n")
    
    file.choose()
 },
  ...
)
}
\arguments{
\item{ghb}{an \code{RMODFLOW} ghb object}

\item{dis}{an \code{RMODFLOW} dis object}

\item{file}{filename to write to; typically '*.ghb'}

\item{...}{arguments passed to \code{rmfi_write_variables} and \code{rmfi_write_list} when writing a fixed format file.}
}
\value{
\code{NULL}
}
\description{
\code{rmf_write_ghb} writes a MODFLOW general-head boundary file based on an \code{RMODFLOW} ghb object
}
\seealso{
\code{\link{rmf_read_ghb}}, \code{\link{rmf_create_ghb}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?ghb.htm}
}
