% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-gmg.R
\name{rmf_write_gmg}
\alias{rmf_write_gmg}
\title{Write a MODFLOW Geometric Multigrid Solver file}
\usage{
rmf_write_gmg(
  gmg,
  file = {    
    cat("Please select gmg file to overwrite or provide new filename ...\\n")    
    file.choose() },
  ...
)
}
\arguments{
\item{gmg}{an \code{\link{RMODFLOW}} gmg object}

\item{file}{filename to write to; typically '*.gmg'}

\item{...}{ignored}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW Geometric Multigrid Solver file
}
\seealso{
\code{\link{rmf_read_gmg}}, \code{\link{rmf_create_gmg}} and \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?gmg.htm}
}
