% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-write-hob.R
\name{rmf_write_hob}
\alias{rmf_write_hob}
\alias{write_hob}
\title{Write a MODFLOW head observations file}
\usage{
rmf_write_hob(hob, file = {    
  cat("Please select hob file to overwrite or provide new filename ...\\n")    
  file.choose() })

write_hob(...)
}
\arguments{
\item{hob}{an \code{\link{RMODFLOW}} hob object}

\item{file}{filename to write to; typically '*.hob'}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW head observations file
}
\section{Functions}{
\itemize{
\item \code{write_hob}: Deprecated function name
}}

