% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-huf.R
\name{rmf_write_kdep}
\alias{rmf_write_kdep}
\title{Write a MODFLOW hydraulic conductivity depth-dependence capability file}
\usage{
rmf_write_kdep(
  kdep,
  file = {
    
    cat("Please select kdep file to overwrite or provide new filename ...\\n")
    
    file.choose()
 },
  iprn = -1,
  ...
)
}
\arguments{
\item{kdep}{an \code{RMODFLOW} kdep object}

\item{file}{filename to write to; typically '*.kdep'}

\item{iprn}{format code for printing arrays in the listing file; defaults to -1 (no printing)}

\item{...}{arguments passed to \code{rmfi_write_array}. Can be ignored when arrays are INTERNAL or CONSTANT.}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW hydraulic conductivity depth-dependence capability file
}
\seealso{
\code{\link{rmf_create_kdep}}, \code{\link{rmf_read_kdep}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?kdep.htm}
}
