% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-lmt.R
\name{rmf_write_lmt}
\alias{rmf_write_lmt}
\title{Write a MODFLOW Link-MT3DMS Package file}
\usage{
rmf_write_lmt(
  lmt,
  file = {
    
    cat("Please choose lmt file to overwrite or provide new filename ...\\n")
    
    file.choose()
 },
  ...
)
}
\arguments{
\item{lmt}{\code{RMODFLOW} lmt object}

\item{file}{filename to write to; typically '*.lmt'}

\item{...}{ignored}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW Link-MT3DMS Package file
}
\examples{
lmt <- rmf_create_lmt()
rmf_write_lmt(lmt, file = file.path(tempdir(), 'input.lmt'))

}
\seealso{
\code{\link{rmf_create_lmt}}, \code{\link{rmf_read_lmt}} and \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?lmt6.htm}
}
