% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-nwt.R
\name{rmf_write_nwt}
\alias{rmf_write_nwt}
\title{Write a MODFLOW Newton solver package file}
\usage{
rmf_write_nwt(
  nwt,
  file = {
    
    cat("Please select nwt file to overwrite or provide new filename ...\\n")
    
    file.choose()
 },
  ...
)
}
\arguments{
\item{nwt}{an \code{\link{RMODFLOW}} nwt object}

\item{file}{filename to write to; typically '*.nwt'}

\item{...}{arguments passed to \code{rmfi_write_variables} when writing a fixed format file.}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW Newton solver package file
}
\seealso{
\code{\link{rmf_read_nwt}}, \code{\link{rmf_create_nwt}} and \url{https://water.usgs.gov/ogw/modflow-nwt/MODFLOW-NWT-Guide/}
}
