% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-pcg.R
\name{rmf_write_pcg}
\alias{rmf_write_pcg}
\title{Write a MODFLOW preconditioned conjugate-gradient package file}
\usage{
rmf_write_pcg(
  pcg,
  file = {
    
    cat("Please select pcg file to overwrite or provide new filename ...\\n")
    
    file.choose()
 },
  ...
)
}
\arguments{
\item{pcg}{an \code{\link{RMODFLOW}} pcg object}

\item{file}{filename to write to; typically '*.pcg'}

\item{...}{arguments passed to \code{rmfi_write_variables} when writing a fixed format file.}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW preconditioned conjugate-gradient package file
}
\seealso{
\code{\link{rmf_read_pcg}}, \code{\link{rmf_create_pcg}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?pcg.htm}
}
