% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_write_prj}
\alias{rmf_write_prj}
\title{Write an RMODFLOW projection file}
\usage{
rmf_write_prj(
  prj,
  file = {    
    cat("Please select prj file to overwrite or provide new filename ...\\n")    
    file.choose() }
)
}
\arguments{
\item{prj}{an \code{\link{RMODFLOW}} prj object}

\item{file}{filename to write to; typically '*.prj'}
}
\description{
\code{write.prj} writes a projection file
}
