% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-pval.R
\name{rmf_write_pvl}
\alias{rmf_write_pvl}
\title{Write a MODFLOW parameter value file}
\usage{
rmf_write_pvl(
  pvl,
  file = {    
    cat("Please select pvl file to overwrite or provide new filename ...\\n")    
    file.choose() }
)
}
\arguments{
\item{pvl}{an \code{\link{RMODFLOW}} pvl object}

\item{file}{filename to write to; typically '*.pvl'}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW parameter value file
}
\seealso{
\code{\link{rmf_create_pvl}}, \code{\link{rmf_read_pvl}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?pvl.htm}
}
