% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-rch.R
\name{rmf_write_rch}
\alias{rmf_write_rch}
\title{Write a MODFLOW recharge file}
\usage{
rmf_write_rch(rch, dis = rmf_read_dis(), file = {    
  cat("Please select rch file to overwrite or provide new filename ...\\n")
      file.choose() })
}
\arguments{
\item{rch}{an \code{RMODFLOW} rch object}

\item{dis}{an \code{RMODFLOW} dis object}

\item{file}{filename to write to; typically '*.rch'}
}
\value{
\code{NULL}
}
\description{
\code{rmf_write_rch} writes a MODFLOW recharge file based on an \code{RMODFLOW} rch object
}
\seealso{
\code{\link{rmf_read_rch}}, \code{\link{rmf_create_rch}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?rch.htm}
}
