% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-write-riv.R
\name{rmf_write_riv}
\alias{rmf_write_riv}
\title{Write a MODFLOW river file}
\usage{
rmf_write_riv(riv, dis = read_dis(), file = {    
  cat("Please choose riv file to overwrite or provide new filename ...\\n")    
  file.choose() })
}
\arguments{
\item{riv}{an \code{RMODFLOW} riv object}

\item{dis}{an \code{RMODFLOW} dis object}

\item{file}{filename to write to; typically '*.riv'}
}
\value{
\code{NULL}
}
\description{
\code{rmf_write_riv} writes a MODFLOW river file based on an \code{RMODFLOW} riv object
}
\seealso{
\code{\link{rmf_read_riv}}, \code{\link{rmf_create_riv}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?riv.htm}
}
