% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-write-sip.R
\name{rmf_write_sip}
\alias{rmf_write_sip}
\title{Write a MODFLOW strongly implicit procedure package}
\usage{
rmf_write_sip(sip, file = {    
  cat("Please choose sip file to overwrite or provide new filename ...\\n")
      file.choose() }, ...)
}
\arguments{
\item{sip}{an \code{RMODFLOW} sip object}

\item{file}{filename to write to; typically '*.sip'}

\item{...}{arguments passed to \code{rmfi_write_variables} when writing a fixed format file.}
}
\value{
\code{NULL}
}
\description{
\code{rmf_write_sip} writes a MODFLOW strongly implicit procedure file based on an \code{RMODFLOW} sip object
}
\seealso{
\code{\link{rmf_read_sip}}, \code{\link{rmf_create_sip}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?sip.htm}
}
