% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-upw.R
\name{rmf_write_upw}
\alias{rmf_write_upw}
\title{Write a MODFLOW-NWT upstream weighting file}
\usage{
rmf_write_upw(upw, file = {    
  cat("Please select upw file to overwrite or provide new filename ...\\n")
      file.choose() }, dis = {    
  cat("Please select corresponding dis file ...\\n")    
  rmf_read_dis(file.choose()) }, iprn = -1, ...)
}
\arguments{
\item{upw}{an \code{\link{RMODFLOW}} upw object}

\item{file}{filename to write to; typically '*.upw'}

\item{iprn}{format code for printing arrays in the listing file; defaults to -1 (no printing)}

\item{...}{arguments passed to \code{rmfi_write_array}. Can be ignored when arrays are INTERNAL or CONSTANT.}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW-NWT upstream weighting file
}
\note{
upw input structure is nearly identical to lpf but calculations are done differently. Differences include the addition of the iphdry value and the ommision of optional keywords. Layer wetting capabilities are also not supported by upw.

upw must be used with the Newton solver. See also \code{\link{rmf_create_nwt}}.
}
