% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-write-wel.R
\name{rmf_write_wel}
\alias{rmf_write_wel}
\title{Write a MODFLOW well file}
\usage{
rmf_write_wel(wel, dis = read_dis(), file = {    
  cat("Please choose wel file to overwrite or provide new filename ...\\n")    
  file.choose() })
}
\arguments{
\item{wel}{an \code{RMODFLOW} wel object}

\item{dis}{an \code{RMODFLOW} dis object}

\item{file}{filename to write to; typically '*.wel'}
}
\value{
\code{NULL}
}
\description{
\code{rmf_write_wel} writes a MODFLOW well file based on an \code{RMODFLOW} wel object
}
\seealso{
\code{\link{rmf_read_wel}}, \code{\link{rmf_create_wel}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?wel.htm}
}
