% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-write-zon.R
\name{rmf_write_zon}
\alias{rmf_write_zon}
\title{Write a MODFLOW zone file}
\usage{
rmf_write_zon(zon, file = {    
  cat("Please choose zon file to overwrite or provide new filename ...\\n")
      file.choose() }, ...)
}
\arguments{
\item{zon}{an \code{RMODFLOW} zon object}

\item{file}{filename to write to; typically '*.zon'}

\item{...}{arguments passed to \code{rmfi_write_array}. Can be ignored when arrays are INTERNAL or CONSTANT.}
}
\value{
\code{NULL}
}
\description{
\code{rmf_write_zon} writes a MODFLOW zone file based on a \code{RMODFLOW} zon object
}
\seealso{
\code{\link{rmf_read_zon}}, \code{\link{rmf_create_zon}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?zone.htm}
}
