% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{rmfi_backup_pval}
\alias{rmfi_backup_pval}
\title{Backup a PVAL file}
\usage{
rmfi_backup_pval(dir, nam)
}
\arguments{
\item{dir}{Character. Path to directory containing the NAM file.}

\item{nam}{Character. File name of the NAM file.}
}
\description{
Backup a PVAL file
}
