% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmfi_bilinear_intp}
\alias{rmfi_bilinear_intp}
\title{Bilinear interpolation on a rectilinear grid}
\usage{
rmfi_bilinear_intp(x, y, f, xout, yout)
}
\arguments{
\item{x}{x coordinates of known points}

\item{y}{y coordinates of known points}

\item{f}{values at known poins (length 4). Order: bottom-left, bottom-right, top-left, top-right; using R's column-major ordering}

\item{xout}{x coordinate of point to interpolate}

\item{yout}{y coordinate of point to interpolate}
}
\value{
single bilinear interpolated value
}
\description{
Bilinear interpolation on a rectilinear grid
}
\keyword{internal}
