% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmfi_convert_coordinates}
\alias{rmfi_convert_coordinates}
\title{Convert data frame coordinates to another coordinate reference system}
\usage{
rmfi_convert_coordinates(dat, from, to)
}
\arguments{
\item{dat}{data frame with x and y (and/or z) coordinates}

\item{from}{coordinate reference system of the data}

\item{to}{target coordinate reference system}
}
\value{
data frame with converted coordinates
}
\description{
Convert data frame coordinates to another coordinate reference system
}
\keyword{internal}
