% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmfi_convert_huf_to_nlay}
\alias{rmfi_convert_huf_to_nlay}
\title{Convert a huf object to an rmf_3d_array with the number of numerical layers per hydrogeological unit}
\usage{
rmfi_convert_huf_to_nlay(huf, dis, bas = NULL)
}
\arguments{
\item{huf}{huf object}

\item{dis}{dis object, corresponding to the huf object}

\item{bas}{bas object, corresponding to the huf object; defaults to NULL}
}
\value{
nlay rmf_3d_array
}
\description{
Convert a huf object to an rmf_3d_array with the number of numerical layers per hydrogeological unit
}
\keyword{internal}
