% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmfi-create-bc-array.R
\name{rmfi_create_bc_array}
\alias{rmfi_create_bc_array}
\title{Set array input for a MODFLOW boundary condition package}
\usage{
rmfi_create_bc_array(arg, dis)
}
\arguments{
\item{arg}{list of (1) \code{rmf_2d_array's} and/or \code{rmf_parm array} objects or (2) a single nested \code{list} with \code{rmf_2d_array's} and/or rmf_parm elements or (3) a single \code{rmf_2d_array}; defines the boundary condition input.}

\item{dis}{dis object. If not explicitely suplied, the function will look in the arg argument for an object of class 'dis'.}
}
\value{
list with the parameters, input arrays and the kper argument
}
\description{
Set array input for a MODFLOW boundary condition package
}
\details{
typically, \code{arg} is \code{list(...)} where the ellipsis contains all the input \code{rmf_arrays} for the \code{rmf_create_*} function.
}
\seealso{
\code{\link{rmfi_bc_array}}
}
\keyword{internal}
