% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{rmfi_find}
\alias{rmfi_find}
\title{Find paths to executables}
\usage{
rmfi_find(code = "2005", precision = "single")
}
\arguments{
\item{code}{Name of the MODFLOW variant to use, or path to the executable.}

\item{precision}{Character. Can be \code{"single"} or \code{"double"}. Only
relevant for MODFLOW-2005.}
}
\value{
Path to the executable.
}
\description{
This function tries to locate external code executables.
}
\details{
It first looks for the executable in the current working directory. If not
there, it looks in the bin subfolder of \code{getOption("RMODFLOW.path")}, where
the software might have been installed by \code{\link[=rmf_install]{rmf_install()}}. If the executable
cannot be found, a final attempt is made by checking the system path
variable. If it still cannot be located, an error is thrown.
}
