% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmfi_fortran_format}
\alias{rmfi_fortran_format}
\title{Returns character lengths per value from a FORTRAN format}
\usage{
rmfi_fortran_format(format)
}
\arguments{
\item{format}{character; FORTRAN format within parentheses as read from a MODFLOW array header}
}
\value{
integer vector with the number of characters per value as specified by the FORTRAN format
}
\description{
Returns character lengths per value from a FORTRAN format
}
\keyword{internal}
