% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmfi_geomean}
\alias{rmfi_geomean}
\title{Calculate a geometric mean}
\usage{
rmfi_geomean(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{further arguments passed to \code{\link{prod}}}
}
\description{
Calculate a geometric mean
}
\seealso{
\code{\link{rmfi_harmean}} and \code{\link{mean}}
}
\keyword{internal}
