% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmfi-get-balance.R
\name{rmfi_get_balance}
\alias{rmfi_get_balance}
\title{Converts MODLFOW budget lines from the listing file into a data frame}
\usage{
rmfi_get_balance(balance_lines)
}
\arguments{
\item{balance_lines}{character vector of length n where each element is a line of a MODFLOW listing file (txt) with n lines. The first line should be "VOLUMETRIC BUDGET OF THE ENTIRE MODEL AT TIME STEP K, STRESS PERIOD M"; the last line should be "PERCENT DISCREPANCY = ..."}
}
\value{
a data frame with the mass balance
}
\description{
Converts MODLFOW budget lines from the listing file into a data frame
}
\keyword{internal}
