% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{rmfi_guide_url}
\alias{rmfi_guide_url}
\title{Complete guide url from page name or shortcut}
\usage{
rmfi_guide_url(name)
}
\arguments{
\item{name}{Name of the online guide page to launch, or a shortcut from
\code{rmfd_guide_shortcuts} that can be translated to one.}
}
\value{
Complete URL of the online guide page.
}
\description{
Complete guide url from page name or shortcut
}
