% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmfi-harmean.R
\name{rmfi_harmean}
\alias{rmfi_harmean}
\title{Calculate a harmonic mean}
\usage{
rmfi_harmean(x, ...)
}
\arguments{
\item{x}{An invertable R object.}

\item{...}{further arguments passed to \code{\link{mean}}}
}
\description{
Calculate a harmonic mean
}
\seealso{
\code{\link{rmfi_geomean}} and \code{\link{mean}}
}
\keyword{internal}
