% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmfi-ifelse0.R
\name{rmfi_ifelse0}
\alias{rmfi_ifelse0}
\title{Conditional return}
\usage{
rmfi_ifelse0(test, yes, no)
}
\arguments{
\item{test}{an object which can be coerced to logical mode.}

\item{yes}{return value for \code{test==TRUE}}

\item{no}{return value for \code{test==FALSE}}
}
\description{
\code{rmfi_ifelse0} returns \code{yes} if \code{test} is \code{TRUE}. If \code{test} is \code{FALSE}, it returns \code{no}.
}
