% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{rmfi_install_code}
\alias{rmfi_install_code}
\title{Install codes}
\usage{
rmfi_install_code(code, overwrite)
}
\arguments{
\item{code}{Character vector with the codes to install, or \code{"all"} (default).}

\item{overwrite}{Logical. Overwrite when the code is already installed? If
\code{NULL} (default), the user is asked what to do in an interactive session.
An error message is issued otherwise.}
}
\description{
Install codes
}
