% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmfi_list_packages}
\alias{rmfi_list_packages}
\title{List supported MODFLOW packages}
\usage{
rmfi_list_packages(type = "all")
}
\arguments{
\item{type}{character denoting type of packages to list; possible values are \code{'all' (default), 'basic', 'flow', 'boundary', 'solver', 'oc', 'sub', 'obs', 'swr', 'cfp', 'farm', 'cbc', 'output'}}
}
\value{
data.frame with ftype and rmf columns denoting the MODFLOW and \code{RMODFLOW} abbreviations for the requested packages
}
\description{
List supported MODFLOW packages
}
\details{
'cbc' returns all packages which allow a i*cbc flag to be set which is a flag and unit number for writing cell-by-cell flow data. 'output' lists all supported output types.
}
\note{
this function should be updated every time a new MODFLOW package is supported in \code{RMODFLOW}
}
\keyword{internal}
