% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{rmfi_look_for_path}
\alias{rmfi_look_for_path}
\title{Look for a file path in a NAM file}
\usage{
rmfi_look_for_path(dir, nam, type = NULL, unit = NULL)
}
\arguments{
\item{dir}{Character. Path to directory containing the NAM file.}

\item{nam}{Character. File name of the NAM file.}

\item{type}{Character. File type (ftype) of the entry to look for.}

\item{unit}{Integer. Unit number of the entry to look for.}
}
\description{
Look for a file path in a NAM file
}
