% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmfi_parse_array_parameters}
\alias{rmfi_parse_array_parameters}
\title{Read MODFLOW array parameters for boundary-condition packages}
\usage{
rmfi_parse_array_parameters(lines, dis, np, mlt = NULL, zon = NULL)
}
\arguments{
\item{lines}{lines to read the parameter arrays from}

\item{dis}{\code{RMODFLOW} dis object}

\item{np}{numeric; number of parameters to read}

\item{mlt}{a \code{RMODFLOW} mlt object. Only needed when reading parameter arrays defined by multiplier arrays}

\item{zon}{a \code{RMODFLOW} zon object. Only needed when reading parameter arrays defined by zone arrays}
}
\value{
A list containing the parameter arrays and the remaining text of the MODFLOW input file
}
\description{
Read MODFLOW array parameters for boundary-condition packages
}
\seealso{
\code{\link{rmfi_write_array_parameters}}
}
\keyword{internal}
