% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmfi_parse_bc_list}
\alias{rmfi_parse_bc_list}
\title{Read input for a MODFLOW boundary condition package which uses list-directed input}
\usage{
rmfi_parse_bc_list(lines, dis, varnames, option, scalevar, ...)
}
\arguments{
\item{lines}{lines as returned from readr::read_lines}

\item{dis}{an \code{RMODFLOW} dis object}

\item{varnames}{character vector with the names of the variables starting from the 4th column (so after ijk)}

\item{option}{optional named logical vector with the names of the options besides aux}

\item{scalevar}{integer, indicating which column is (possibly) scaled by SFAC}

\item{...}{arguments passed to \code{rmfi_parse_variables} and \code{rmfi_parse_list}.}
}
\value{
list with (optional) comments, icb, option, aux and rmf_lists
}
\description{
Read input for a MODFLOW boundary condition package which uses list-directed input
}
\seealso{
\code{\link{rmfi_create_bc_list}}, \code{\link{rmfi_write_bc_list}}
}
\keyword{internal}
