% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmfi-parse-comments.R
\name{rmfi_parse_comments}
\alias{rmfi_parse_comments}
\title{Read comments
Internal function used in the read_* functions to read comments}
\usage{
rmfi_parse_comments(remaining_lines)
}
\description{
Read comments
Internal function used in the read_* functions to read comments
}
\details{
prevents copying of RMODFLOW header comment
}
\keyword{internal}
