% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmfi_parse_list}
\alias{rmfi_parse_list}
\title{Reads a MODFLOW list}
\usage{
rmfi_parse_list(
  remaining_lines,
  nlst,
  l = NULL,
  varnames,
  scalevar = 4,
  file,
  naux = 0,
  format = "free",
  precision = "single",
  ...
)
}
\arguments{
\item{nlst}{number of list rows to read}

\item{l}{stress period number}

\item{varnames}{character vector; names of the variables starting from the 4th column (so after ijk). Length of varnames is used to dimension the dataframe}

\item{scalevar}{column name or integer; this column will be scaled}

\item{file}{the file that is being read; needed if list is specified through an OPEN/CLOSE statement}

\item{naux}{integer; number of auxiliary variables to read (which are always free format). Defaults to 0.}

\item{format}{either 'fixed' or 'free'}

\item{...}{ignored}
}
\description{
Reads a MODFLOW list
}
\keyword{internal}
