% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmfi_plot_bc}
\alias{rmfi_plot_bc}
\title{Plot a RMODFLOW list-directed boundary condition object}
\usage{
rmfi_plot_bc(obj, dis, kper = NULL, variable = "id", i = NULL,
  j = NULL, k = NULL, active_only = FALSE, fun = sum, ...)
}
\arguments{
\item{obj}{a \code{RMODFLOW} list-directed boundary condition object, i.e. riv, chd, wel, drn, ghb, and derivatives}

\item{dis}{a \code{RMODFLOW} dis object}

\item{kper}{integer specifying the stress-period to plot}

\item{variable}{single character or numeric indicating which column in the \code{rmf_list} object to plot. Defaults to 'id', which plots the locations of the cells.}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{active_only}{logical; indicating if only the active cells should be plotted. Non-active cells are set to NA. Defaults to FALSE.}

\item{fun}{function to compute values in the case multiple values are defined for the same MODFLOW cell. Typically either \code{mean} or \code{sum}. Defaults to sum.}

\item{...}{additional arguments passed to \code{\link{rmf_plot.rmf_3d_array}}}
}
\value{
ggplot2 object or layer
}
\description{
Plot a RMODFLOW list-directed boundary condition object
}
\keyword{internal}
