% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rmfi_prj_length_multiplier}
\alias{rmfi_prj_length_multiplier}
\title{Obtain a multiplier to convert MODFLOW length units to projection length units}
\usage{
rmfi_prj_length_multiplier(dis, prj, to)
}
\arguments{
\item{dis}{\code{RMODFLOW} dis object}

\item{prj}{\code{RMODFLOW} prj object}

\item{to}{either 'grid' or 'xyz' specifying if the multiplier should convert coordinates to the modflow grid or to real world coordinates (inverse)}
}
\value{
single numeric value which can be used to multiply MODFLOW coordinates with so to convert them to prj length units.
}
\description{
Obtain a multiplier to convert MODFLOW length units to projection length units
}
\details{
The MODFLOW length unit \code{(dis$lenuni)} can be different from the projection unit \code{(prj$crs$units)}. When converting coordinates
using \code{rmf_convert_grid_to_xyz} or \code{rmf_convert_xyz_to_grid} the difference in length unit needs to be corrected for.
}
\keyword{internal}
