% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmfi-read-bc-array.R
\name{rmfi_read_bc_array}
\alias{rmfi_read_bc_array}
\title{Read input for a MODFLOW boundary condition package which uses list-directed input}
\usage{
rmfi_read_bc_array(lines, dis, nam = NULL, ...)
}
\arguments{
\item{lines}{lines as returned from readr::read_lines}

\item{dis}{an \code{RMODFLOW} dis object}

\item{nam}{a \code{RMODFLOW} nam object. Only needed when reading parameter arrrays defined by multiplier and/or zone arrays}

\item{...}{arguments passed to \code{rmfi_parse_variables} and \code{rmfi_parse_array}.}
}
\value{
list with (optional) comments and rmf_arrays
}
\description{
Read input for a MODFLOW boundary condition package which uses list-directed input
}
\seealso{
\code{\link{rmfi_create_bc_array}}, \code{\link{rmfi_write_bc_array}}
}
\keyword{internal}
