% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmfi-remove-comments-end-of-line.R
\name{rmfi_remove_comments_end_of_line}
\alias{rmfi_remove_comments_end_of_line}
\title{Remove comments at the end of a string}
\usage{
rmfi_remove_comments_end_of_line(line)
}
\arguments{
\item{line}{A string.}
}
\value{
The string, without the commented part.
}
\description{
Remove comments at the end of a string
}
\keyword{internal}
