% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmfi_remove_empty_strings}
\alias{rmfi_remove_empty_strings}
\title{Remove empty elements from a vector of strings.}
\usage{
rmfi_remove_empty_strings(vector_of_strings)
}
\arguments{
\item{vector_of_strings}{Vector of strings.}
}
\value{
Vector of strings without the empty items.
}
\description{
Remove empty elements from a vector of strings.
}
\keyword{internal}
