% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmfi_trilinear_intp}
\alias{rmfi_trilinear_intp}
\title{Trilinear interpolation on a rectilinear grid}
\usage{
rmfi_trilinear_intp(x, y, z, f, xout, yout, zout)
}
\arguments{
\item{x}{x coordinates of known points}

\item{y}{y coordinates of known points}

\item{z}{z coordinates of known points}

\item{f}{values at known poins (length 8). Order: upper:bottom-left, bottom-right, top-left, top-right; bottom:bottom-left, bottom-right, top-left, top-right; using R's column-major ordering}

\item{xout}{x coordinate of point to interpolate}

\item{yout}{y coordinate of point to interpolate}

\item{zout}{z coordinate of points to interpolate}
}
\value{
single trilinear interpolated value
}
\description{
Trilinear interpolation on a rectilinear grid
}
\keyword{internal}
