% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmfi-weighted-geomean.R
\name{rmfi_weighted_geomean}
\alias{rmfi_weighted_geomean}
\title{Calculate a weighted geometric mean}
\usage{
rmfi_weighted_geomean(x, w, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{further arguments passed to \code{\link{prod}}}
}
\description{
Calculate a weighted geometric mean
}
\seealso{
\code{\link{rmfi_weighted_harmean}}, \code{\link{weighted.mean}}, \code{\link{rmfi_geomean}}, \code{\link{rmfi_harmean}} and \code{\link{mean}}
}
\keyword{internal}
