% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmfi-weighted-harmean.R
\name{rmfi_weighted_harmean}
\alias{rmfi_weighted_harmean}
\title{Calculate a weighted harmonic mean}
\usage{
rmfi_weighted_harmean(x, w, ...)
}
\arguments{
\item{x}{An invertable R object.}

\item{...}{further arguments passed to \code{\link{sum}}}
}
\description{
Calculate a weighted harmonic mean
}
\seealso{
\code{\link{rmfi_weighted_geomean}}, \code{\link{weighted.mean}}, \code{\link{rmfi_harmean}} \code{\link{rmfi_geomean}} and \code{\link{mean}}
}
