% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmfi-write-array.R
\name{rmfi_write_array}
\alias{rmfi_write_array}
\title{Write modflow array
Internal function used in the write_* functions for writing array datasets}
\usage{
rmfi_write_array(array, file, cnstnt = 1, iprn = -1, append = TRUE)
}
\description{
Write modflow array
Internal function used in the write_* functions for writing array datasets
}
