% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmfi_write_array_parameters}
\alias{rmfi_write_array_parameters}
\title{Write MODFLOW array parameters for boundary-condition packages}
\usage{
rmfi_write_array_parameters(obj, arrays, file, partyp, ...)
}
\arguments{
\item{obj}{\code{RMODFLOW} object to write}

\item{arrays}{list of arrays to write}

\item{file}{filename to write to}

\item{...}{additional arguments passed to \code{rmfi_write_array}}

\item{type}{character; type of array parameter. Allowed values are \code{'bc'} for boundary-condition arrays and \code{'flow'} for flow package arrays}
}
\description{
Write MODFLOW array parameters for boundary-condition packages
}
\seealso{
\code{\link{rmfi_parse_array_parameters}}
}
\keyword{internal}
