% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmfi_write_bc_list}
\alias{rmfi_write_bc_list}
\title{Write a MODFLOW boundary condition package which uses list-directed input to a file}
\usage{
rmfi_write_bc_list(file, obj, dis, varnames, header, package, partyp, ...)
}
\arguments{
\item{file}{filename to write to}

\item{obj}{an \code{RMODFLOW} boundary condition rmf_package object with list directed input}

\item{dis}{an \code{RMODFLOW} dis object}

\item{varnames}{character vector with the names of the variables starting from the 4th column (so after ijk)}

\item{header}{character; package name. Part of the header comment written to the output file}

\item{package}{character; acronym (often 3 letters) used by MODFLOW to name to package}

\item{partyp}{character; specifies the parameter type}

\item{...}{arguments passed to \code{rmfi_write_variables} and \code{rmfi_write_list} when writing a fixed format file.}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW boundary condition package which uses list-directed input to a file
}
\seealso{
\code{\link{rmfi_create_bc_list}}, \code{\link{rmfi_parse_bc_list}}
}
\keyword{internal}
