% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmfi_write_list}
\alias{rmfi_write_list}
\title{Write a RMODFLOW list}
\usage{
rmfi_write_list(
  df,
  file,
  varnames,
  aux = NULL,
  format = "free",
  append = TRUE,
  ...
)
}
\arguments{
\item{df}{\code{RMODFLOW} list}

\item{file}{filename to write to}

\item{varnames}{character vector with the names of the variables starting from the 4th column (so after ijk) including optional auxiliary variables}

\item{aux}{character vector with the names of the auxiliary variables defined in \code{varnames}}

\item{format}{either \code{"free"} (default) or \code{"fixed"}}

\item{append}{logical}

\item{...}{ignored}
}
\value{
\code{NULL}
}
\description{
Write a RMODFLOW list
}
\keyword{internal}
