% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rmfi_write_prj}
\alias{rmfi_write_prj}
\title{Write prj object information to file header}
\usage{
rmfi_write_prj(dis, prj, file)
}
\arguments{
\item{dis}{\code{RMODFLOW} dis object}

\item{prj}{\code{RMODFLOW} prj object}

\item{file}{character with path of file to write to. Typically the discretization file.}
}
\value{
\code{NULL}
}
\description{
Write prj object information to file header
}
\details{
Writes RMODFLOW projection information into the header of a file, typically the discretization file. All lines start with "#".
This information consists of a starter line, the coordinates of the 4 model corners, the grid rotation angle in degrees counterclockwise,
the z coordinate of the lower left corner and the crs description. This might be a EPSG code, a proj4string, or a wkt string.
A ending line specifies the end of the RMODFLOW projection information.
}
\keyword{internal}
