% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmfi_write_variables}
\alias{rmfi_write_variables}
\title{Write modflow variables
Internal function used in the write_* functions for writing single line datasets}
\usage{
rmfi_write_variables(..., file, append = TRUE, format = "free")
}
\arguments{
\item{format}{either \code{'fixed'} or \code{'free'}. Fixed format assumes 10 character spaces for each value}
}
\description{
Write modflow variables
Internal function used in the write_* functions for writing single line datasets
}
\keyword{internal}
