% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run_modflow_opt}
\alias{run_modflow_opt}
\title{Run a MODFLOW model optimization, based on the parameter value file}
\usage{
run_modflow_opt(file, dir = getwd(), modflow_executable = "mf2005",
  par = NULL, include = NULL, trans = NULL, method = "Nelder-Mead",
  lower = -Inf, upper = Inf, control = list(), ...)
}
\arguments{
\item{file}{Name file; typically "*.nam"}

\item{dir}{Directory of the namefile}

\item{modflow_executable}{name of the MODFLOW executable to use}

\item{par}{initial parameter values (for all or only included parameters); parameter value file values are used if par is not provided}

\item{include}{logical vector indicating which parameters in the parameter value file to include in the optimization}

\item{...}{further arguments provided to \code{optim}}
}
\value{
\code{optim} results with the full list of parameters
}
\description{
\code{run_modflow_optim} runs a MODFLOW optimization.
}

