% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run_modflow_sen}
\alias{run_modflow_sen}
\title{Run a MODFLOW model sensitivity analysis, based on the parameter value file}
\usage{
run_modflow_sen(file, dir = getwd(), modflow_executable = "mf2005",
  par = NULL, include = NULL)
}
\arguments{
\item{file}{Name file; typically "*.nam"}

\item{dir}{Directory of the namefile}

\item{modflow_executable}{name of the MODFLOW executable to use}

\item{par}{initial parameter values (for all or only included parameters); parameter value file values are used if par is not provided}

\item{include}{logical vector indicating which parameters in the parameter value file to include in the optimization}
}
\value{
sensitivity analysis results
}
\description{
\code{run_modflow_sens} performs a MODFLOW model sensitivity analysis.
}

