% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sen.R
\name{run_sen}
\alias{run_sen}
\title{Run a MODFLOW model sensitivity analysis, based on the parameter value file}
\usage{
run_sen(file, modflow_executable = "mf2005", par = NULL, include = NULL)
}
\arguments{
\item{file}{path to name file; typically '*.nam'}

\item{modflow_executable}{name of the MODFLOW executable to use}

\item{par}{central parameter values (for all or only included parameters); parameter value file values are used if par is not provided}

\item{include}{logical vector indicating which parameters in the parameter value file to include in the sensitivity analysis}
}
\value{
sensitivity analysis results
}
\description{
\code{run_modflow_sen} performs a MODFLOW model sensitivity analysis.
}

