% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{split_line_numbers}
\alias{split_line_numbers}
\title{Split a string in numbers.}
\usage{
split_line_numbers(string)
}
\arguments{
\item{string}{A string.}
}
\value{
Vector of numbers within the string, without comments at the end of the string.
}
\description{
Split a string in numbers.
}

