% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{split_line_words}
\alias{split_line_words}
\title{Split a string in words.}
\usage{
split_line_words(string)
}
\arguments{
\item{string}{A string.}
}
\value{
Vector of words within the string, without comments at the end of the string.
}
\description{
Split a string in words.
}

