% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str.modflow_package.R
\name{str.modflow_package}
\alias{str.modflow_package}
\title{Structure of modflow package object}
\usage{
\method{str}{modflow_package}(modflow_package)
}
\arguments{
\item{modflow_package}{RMODFLOW modflow package object}
}
\description{
Structure of modflow package object
}
\details{
Provides the object comment attribute additional to the standard \code{str} output.
}

