% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.modflow_package.R
\name{summary.modflow_package}
\alias{summary.modflow_package}
\title{Summary of modflow package object}
\usage{
\method{summary}{modflow_package}(modflow_package)
}
\arguments{
\item{modflow_package}{RMODFLOW modflow package object}
}
\description{
Summary of modflow package object
}
\details{
Provides the object comment attribute additional to the standard \code{summary} output.
}

