% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{weighted.geomean}
\alias{weighted.geomean}
\title{Calculate a weighted geometric mean}
\usage{
weighted.geomean(x, w, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{further arguments passed to \code{\link{prod}}}
}
\description{
Calculate a weighted geometric mean
}
\seealso{
\code{\link{weighted.harmean}}, \code{\link{weighted.mean}}, \code{\link{geomean}}, \code{\link{harmean}} and \code{\link{mean}}
}

