% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{weighted.harmean}
\alias{weighted.harmean}
\title{Calculate a weighted harmonic mean}
\usage{
weighted.harmean(x, w, ...)
}
\arguments{
\item{x}{An invertable R object.}

\item{...}{further arguments passed to \code{\link{sum}}}
}
\description{
Calculate a weighted harmonic mean
}
\seealso{
\code{\link{weighted.geomean}}, \code{\link{weighted.mean}}, \code{\link{harmean}} \code{\link{geomean}} and \code{\link{mean}}
}

