% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{write_array}
\alias{write_array}
\title{Write modflow array
Internal function used in the write_* functions for writing array datasets}
\usage{
write_array(array, file, cnstnt = 1, iprn = -1, append = TRUE)
}
\description{
Write modflow array
Internal function used in the write_* functions for writing array datasets
}

