% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write_ba6}
\alias{write_ba6}
\title{Write a MODFLOW basic file}
\usage{
write_ba6(ba6, file, IPRN = -1)
}
\arguments{
\item{ba6}{An \code{\link{RMODFLOW}} ba6 object}

\item{file}{File to write to; typically "*.ba6"}

\item{IPRN}{Needed for the format code.}

\item{info}{Additional information}
}
\value{
None
}
\description{
\code{write_ba6} writes a MODFLOW basic file based on an \code{\link{RMODFLOW}} ba6 object.
}

