% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_bas.R
\name{write_bas}
\alias{write_bas}
\title{Write a MODFLOW basic file}
\usage{
write_bas(bas, file, iprn = -1)
}
\arguments{
\item{bas}{an \code{\link{RMODFLOW}} bas object}

\item{file}{filename to write to; typically '*.bas'}

\item{iprn}{format code for printing arrays in the listing file; defaults to -1 (no printing)}
}
\value{
\code{NULL}
}
\description{
\code{write_bas} writes a MODFLOW basic file based on an \code{\link{RMODFLOW}} bas object.
}

