% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dis.R
\name{write_dis}
\alias{write_dis}
\title{Write a MODFLOW discretization file}
\usage{
write_dis(dis, file = {    
  cat("Please select dis file to overwrite or provide new filename ...\\n")    
  file.choose() }, iprn = -1)
}
\arguments{
\item{dis}{an \code{\link{RMODFLOW}} dis object}

\item{file}{filename to write to; typically '*.dis'}

\item{iprn}{format code for printing arrays in the listing file; defaults to -1 (no printing)}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW discretization file
}
\seealso{
\code{\link{read_dis}}, \code{\link{create_dis}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?dis.htm}
}

