% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hob.R
\name{write_hob}
\alias{write_hob}
\title{Write a MODFLOW head observations file}
\usage{
write_hob(hob, file = {    
  cat("Please select hob file to overwrite or provide new filename ...\\n")    
  file.choose() })
}
\arguments{
\item{hob}{an \code{\link{RMODFLOW}} hob object}

\item{file}{filename to write to; typically '*.hob'}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW head observations file
}

