% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_modflow_array}
\alias{write_modflow_array}
\title{Write modflow array}
\usage{
write_modflow_array(modflow_array, file, CNSTNT = 1, IPRN = -1,
  append = TRUE)
}
\description{
Internal function used in the write_* functions for writing array datasets
}

