% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nam.R
\name{write_nam}
\alias{write_nam}
\title{Write a MODFLOW name file}
\usage{
write_nam(nam, file = {    
  cat("Please select nam file to overwrite or provide new filename ...\\n")    
  file.choose() })
}
\arguments{
\item{nam}{an \code{\link{RMODFLOW}} nam object}

\item{file}{filename to write to; typically '*.nam'}
}
\value{
\code{NULL}
}
\description{
\code{write_nam} writes a MODFLOW name file based on an \code{\link{RMODFLOW}} nam object.
}

