% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_oc.R
\name{write_oc}
\alias{write_oc}
\title{Write a MODFLOW output control option file}
\usage{
write_oc(oc, file = {    
  cat("Please select oc file to overwrite or provide new filename ...\\n")    
  file.choose() })
}
\arguments{
\item{oc}{an \code{\link{RMODFLOW}} oc object}

\item{file}{filename to write to; typically '*.oc'}
}
\value{
\code{NULL}
}
\description{
\code{write_oc} writes a MODFLOW output control option file based on an \code{\link{RMODFLOW}} oc object.
}
\seealso{
\code{\link{read_oc}}, \code{\link{create_oc}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?oc.htm}
}

