% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write_prj}
\alias{write_prj}
\title{Write an RMODFLOW projection file}
\usage{
write_prj(prj, file)
}
\arguments{
\item{prj}{an \code{\link{RMODFLOW}} prj object}

\item{file}{filename to write to; typically '*.prj'}
}
\description{
\code{write.prj} writes a projection file
}

