% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_pvl.R
\name{write_pvl}
\alias{write_pvl}
\title{Write a MODFLOW parameter value file}
\usage{
write_pvl(pvl, file)
}
\arguments{
\item{pvl}{an \code{\link{RMODFLOW}} pvl object}

\item{file}{filename to write to; typically '*.pvl'}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW parameter value file
}

