% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{write_variables}
\alias{write_variables}
\title{Write modflow variables}
\usage{
write_variables(..., file, append = TRUE)
}
\description{
Internal function used in the write_* functions for writing single line datasets
}

