% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cell_centers.R, R/cell_centers.dis.R, R/cell_centers.huf.R
\name{cell_centers}
\alias{cell_centers}
\alias{cell_centers.dis}
\alias{cell_centers.huf}
\title{Generic function to get cell center z coordinates}
\usage{
cell_centers(...)

\method{cell_centers}{dis}(dis)

\method{cell_centers}{huf}(huf)
}
\arguments{
\item{dis}{dis object}

\item{huf}{huf object}
}
\value{
3d array with cell center z coordinates

3d array with cell center z coordinates
}
\description{
Generic function to get cell center z coordinates

Get cell center z coordinates from a dis object

Get cell center z coordinates from a huf object
}

