% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_bud_to_darcy.R
\name{convert_bud_to_darcy}
\alias{convert_bud_to_darcy}
\title{Convert bud object fluxes to darcy velocities}
\usage{
convert_bud_to_darcy(bud, dis)
}
\arguments{
\item{bud}{bud object}

\item{dis}{dis object}
}
\value{
list of four 3d arrays: right, front, lower and magnitude
}
\description{
Convert bud object fluxes to darcy velocities
}

