% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_bud_to_darcy}
\alias{convert_bud_to_darcy}
\title{Convert bud object fluxes to darcy velocities}
\usage{
convert_bud_to_darcy(bud, dis)
}
\arguments{
\item{dis}{dis object}

\item{values}{Vector of parameter values, corresponding to HGUNAM}

\item{huf}{huf object}
}
\value{
object with four 3d arrays: right, front, lower and magnitude
}
\description{
Convert bud object fluxes to darcy velocities
}

