% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_huf_to_dis}
\alias{convert_huf_to_dis}
\title{Convert a parameter defined on the HUF grid to the numerical grid}
\usage{
convert_huf_to_dis(values, huf, dis, mask = dis$TOP/dis$TOP,
  type = "arithmetic")
}
\arguments{
\item{values}{vector of parameter values, in the order of HGUNAM}

\item{huf}{huf object}

\item{dis}{dis object}

\item{mask}{masking 3d array, typically the IBOUND array, to speed up grid conversion; defaults to including all cells}

\item{type}{type of averaging that should be performed; either arithmetic (default), harmonic or geometric}
}
\value{
3d array
}
\description{
Convert a parameter defined on the HUF grid to the numerical grid
}

