% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_huf_to_grid.R
\name{convert_huf_to_grid}
\alias{convert_huf_to_grid}
\title{Convert a parameter defined on the HUF grid to the numerical grid}
\usage{
convert_huf_to_grid(values, huf, dis, mask = dis$top/dis$top,
  type = "arithmetic")
}
\arguments{
\item{values}{vector of parameter values, in the order of \code{hgunam}}

\item{huf}{huf object}

\item{dis}{dis object}

\item{mask}{masking 3d array, typically the \code{ibound} array, to speed up grid conversion; defaults to including all cells}

\item{type}{type of averaging that should be performed; either arithmetic (default), harmonic or geometric}
}
\value{
3d array
}
\description{
Convert a parameter defined on the HUF grid to the numerical grid
}

