% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_id_to_ijk}
\alias{convert_id_to_ijk}
\title{Convert id to ijk}
\usage{
convert_id_to_ijk(id, dis)
}
\arguments{
\item{id}{cell id, providing the place of the number in an input file 3d array (not like the way R uses ids for arrays or matrices; rows and columns are switched)}

\item{dis}{a discretisation file object}
}
\description{
Convert id to ijk
}

