% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dis.R
\name{create_dis}
\alias{create_dis}
\title{Create an \code{RMODFLOW} dis object}
\usage{
create_dis(nlay = 3, nrow = 10, ncol = 10, nper = 1, itmuni = 1,
  lenuni = 2, laycbd = rep(0, nlay), delr = rep(100, ncol),
  delc = rep(100, nrow), top = matrix(0, nrow = nrow, ncol = ncol),
  botm = array(rep(seq(0, -10 * nlay, length = nlay + 1)[2:(nlay + 1)], each =
  nrow * ncol), dim = c(nrow, ncol, nlay)), perlen = rep(1, nper),
  nstp = rep(1, nper), tsmult = rep(1, nper), sstr = c("SS", rep("TS",
  nper - 1)))
}
\arguments{
\item{nlay}{number of layers; defaults to 3}

\item{nrow}{number of rows; defaults to 10}

\item{ncol}{number of columns; defaults to 10}

\item{nper}{number of stress periods; defaults to 1}

\item{itmuni}{time unit; defaults to 1 (seconds)}

\item{lenuni}{length unit; defaults to 2 (metres)}

\item{laycbd}{vector of quasi-3D confining bed flags; defaults to 0 for all layers}

\item{delr}{vector of cell widths along rows; defaults to 100 for all columns}

\item{delc}{vector of cell widths along columns; defaults to 100 for all rows}

\item{top}{matrix with the top elevation of layer 1; defaults to 0 for all nrow x ncol cells}

\item{botm}{3D array with the bottom elevations of all layers; defaults to nlay layers, equally spaced between 0 (top first layer) and -100 (bottom last layer)}

\item{perlen}{vector of stress period lengths}

\item{nstp}{vector of stress period time steps}

\item{tsmult}{vector of successive time step length multipliers}

\item{sstr}{character vector with steady state ('SS') or transient ('TS') stress period indicator}
}
\value{
Object of class dis
}
\description{
\code{create_dis} creates an \code{RMODFLOW} dis object.
}
\seealso{
\code{\link{read_dis}}, \code{\link{write_dis}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?dis.htm}
}

