% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_hob.R
\name{create_hob}
\alias{create_hob}
\title{Create an \code{RMODFLOW} hob object}
\usage{
create_hob(locations, time_series, dis, hydraulic_conductivity = array(1, dim
  = c(dis$NROW, dis$NCOL, dis$NLAY)), IUHOBSV = 0, HOBDRY = -888,
  NOPRINT = FALSE, TOMULTH = 1, unique_obsnam = FALSE, prj = NULL)
}
\arguments{
\item{locations}{data.frame with the name, x and y coordinates, and top and bottom of the monitoring well filter}

\item{time_series}{data.frame with the name of the monitoring well filter and the observation time and head}

\item{dis}{RMODFLOW dis object}

\item{hydraulic_conductivity}{3d array with hydraulic conductivity values, for calculating the weights for a multilayer observation}

\item{IUHOBSV}{file unit for saving observation data in a file; the file unit should be included as DATA type in the name file; defaults to 0 (\emph{i.e.} no observation output file will be written)}

\item{HOBDRY}{value of the simulated equivalent for a dry cell used in the observation output file}

\item{NOPRINT}{logical; if TRUE, the input and output are not printed in the listing file}

\item{unique_obsnam}{logical; should an ID number be added to OBSNAM for filters with observations at different times? Defaults to FALSE.}

\item{prj}{RMODFLOW prj object; provide if coordinates in locations are real world coordinates}

\item{TOMULT}{time offset multiplier, to convert the time unit in time_series to the MODFLOW time unit}
}
\value{
RMODFLOW hob object
}
\description{
\code{create_hob} creates an RMODFLOW hob object
}
\seealso{
\code{\link{read.hob}} and \code{\link{write.hob}}
}

