% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflowDataFuncs.R
\name{exit}
\alias{exit}
\title{Exit Function}
\usage{
exit(msg)
}
\value{
\code{msg} Displays message on exit
}
\description{
\code{exit} Provides Function to exit a little more nicely
}
\examples{
     exit("abnormal termination")
}
