% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_vector.rmodflow_2d_array.R
\name{export_vector.rmodflow_2d_array}
\alias{export_vector.rmodflow_2d_array}
\title{Generic function to export vectors}
\usage{
\method{export_vector}{rmodflow_2d_array}(array, dis, bas = NULL,
  mask = ifelse0(is.null(bas), array * 0 + 1, bas$ibound[, , 1]),
  prj = NULL, crs = NULL, file = "rmodflow_vector",
  type = "ESRI Shapefile", include_ijk = FALSE)
}
\description{
Generic function to export vectors
}

