% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_comments_from_lines}
\alias{get_comments_from_lines}
\title{Get comments at the beginning of a vector of strings}
\usage{
get_comments_from_lines(lines)
}
\arguments{
\item{lines}{Vector of strings, typically obtained from using \code{scan} with \code{sep='\n'}.}
}
\value{
Vector of strings with the commented lines.
}
\description{
Get comments at the beginning of a vector of strings
}

