% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R, R/performance.hpr.R
\name{performance}
\alias{performance}
\alias{performance.hpr}
\title{Generic function to get model performance measures}
\usage{
performance(...)

\method{performance}{hpr}(hpr)
}
\arguments{
\item{hpr}{head predictions file object}
}
\value{
performance measures
}
\description{
Generic function to get model performance measures

Get model performance measures from a hpr object
}

