% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.modflow_3d_array}
\alias{plot.modflow_3d_array}
\title{Plot a 2D section through a MODFLOW 3D array}
\usage{
\method{plot}{modflow_3d_array}(modflow_3d_array, i = NULL, j = NULL,
  k = NULL, dis, mask = modflow_3d_array * 0 + 1,
  zlim = range(modflow_3d_array[ifelse0(is.null(i),
  c(1:dim(modflow_3d_array)[1]), i), ifelse0(is.null(j),
  c(1:dim(modflow_3d_array)[2]), j), ifelse0(is.null(k),
  c(1:dim(modflow_3d_array)[3]), k)], finite = TRUE),
  color.palette = rev_rainbow, nlevels = 7, ...)
}
\arguments{
\item{modflow_3d_array}{An object of class modflow_3d_array}

\item{k}{The number of the layer to plot}

\item{mask}{A 3D ibound array with 1 or TRUE indicating active cells, and 0 or F indicating inactive cells}

\item{zlim}{}

\item{color.palette}{A color palette for imaging the parameter values}

\item{nlevels}{}

\item{levels}{}

\item{main}{}
}
\value{
None
}
\description{
\code{plot.modflow_3d_array} plots a 2D section through a MODFLOW 3D array.
}

