% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_ba6.R
\name{read_ba6}
\alias{read_ba6}
\title{Read a MODFLOW basic file}
\usage{
read_ba6(file, dis = read_dis(paste(substring(file, 1, nchar(file) - 4),
  ".dis", sep = "")))
}
\arguments{
\item{file}{filename; typically '*.ba6'}

\item{dis}{discretization file object; defaults to that with the same filename but with extension '.dis'}
}
\value{
object of class ba6
}
\description{
\code{read_ba6} reads in a MODFLOW basic file and returns it as an \code{\link{RMODFLOW}} ba6 object.
}

