% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bud.R
\name{read_bud}
\alias{read_bud}
\title{Read a MODFLOW budget file}
\usage{
read_bud(file = {     cat("Please select bud file...\\n")     file.choose() },
  binary = TRUE)
}
\arguments{
\item{file}{filename; typically 'Budget.out'}

\item{binary}{logical; is source file binary?}
}
\value{
object of class bud
}
\description{
\code{read_bud} reads in a MODFLOW budget file
}

