% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_dis.R
\name{read_dis}
\alias{read_dis}
\title{Read a MODFLOW discretization file}
\usage{
read_dis(file = {     cat("Please select dis file...\\n")     file.choose() })
}
\arguments{
\item{file}{filename; typically '*.dis'}
}
\value{
object of class dis
}
\description{
\code{read_dis} reads in a MODFLOW discretization file and returns it as an \code{\link{RMODFLOW}} dis object.
}
\seealso{
\code{\link{write_dis}}, \code{\link{create_dis}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?dis.htm}
}

