% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_hed.R
\name{read_hed}
\alias{read_hed}
\title{Read a MODFLOW head file}
\usage{
read_hed(file, dis = read_dis(paste(substring(file, 1, nchar(file) - 4),
  ".dis", sep = "")), ba6 = read_ba6(paste(substring(file, 1, nchar(file) -
  4), ".ba6", sep = "")), convert_HNOFLO_to_NA = TRUE)
}
\arguments{
\item{file}{filename; typically '*.hed'}

\item{dis}{discretization file object; defaults to that with the same filename but with extension '.dis'}

\item{ba6}{basic file object; defaults to that with the same filename but with extension '.ba6'}

\item{convert_HNOFLO_to_NA}{logical; should HNOFLO values be converted to NA?}
}
\value{
object of class hed
}
\description{
\code{read_hed} reads in a MODFLOW head file and returns it as an \code{\link{RMODFLOW}} hed object.
}

