% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_hed.R
\name{read_hed}
\alias{read_hed}
\title{Read a MODFLOW head file}
\usage{
read_hed(file = {     cat("Please select hed file...\\n")     file.choose() },
  dis = {     cat("Please select dis file...\\n")     read_dis(file.choose())
  }, bas = {     cat("Please select bas file...\\n")    
  read_bas(file.choose(), dis = dis) }, convert_hnoflo_to_NA = TRUE,
  binary = TRUE)
}
\arguments{
\item{file}{filename; typically '*.hed'}

\item{dis}{discretization file object; defaults to that with the same filename but with extension '.dis'}

\item{bas}{basic file object; defaults to that with the same filename but with extension '.bas'}

\item{convert_hnoflo_to_NA}{logical; should hnoflo values be converted to NA?}
}
\value{
object of class hed
}
\description{
\code{read_hed} reads in a MODFLOW head file and returns it as an \code{\link{RMODFLOW}} hed object.
}

