% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_hob}
\alias{read_hob}
\title{Read a MODFLOW head observations file}
\usage{
read_hob(file)
}
\arguments{
\item{file}{filename; typically '*.hob'}
}
\value{
object of class hob
}
\description{
\code{read_hob} reads in a MODFLOW head observations file and returns it as an \code{\link{RMODFLOW}} hob object.
}

