% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_hpr}
\alias{read_hpr}
\title{Read a MODFLOW head predictions file}
\usage{
read_hpr(file)
}
\arguments{
\item{file}{filename; typically '*.hpr'}
}
\value{
object of class hpr
}
\description{
\code{read_hpr} reads in a MODFLOW head predictions file and returns it as an \code{\link{RMODFLOW}} hpr object.
}

