% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_huf.R
\name{read_huf}
\alias{read_huf}
\title{Read a MODFLOW hydrogeologic unit flow file}
\usage{
read_huf(file, dis = read_dis(paste(substring(file, 1, nchar(file) - 4),
  ".dis", sep = "")))
}
\arguments{
\item{file}{filename; typically '*.huf'}

\item{dis}{discretization file object; defaults to that with the same filename but with extension '.dis'}
}
\value{
object of class huf
}
\description{
\code{read_huf} reads in a MODFLOW hydrogeologic unit flow file and returns it as an \code{\link{RMODFLOW}} huf object.
}

