% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_huf.R
\name{read_huf}
\alias{read_huf}
\title{Read a MODFLOW hydrogeologic unit flow file}
\usage{
read_huf(file = {     cat("Please select huf file ...\\n")     file.choose()
  }, dis = read_dis())
}
\arguments{
\item{file}{filename; typically '*.huf'}

\item{dis}{discretization file object; defaults to that with the same filename but with extension '.dis'}
}
\value{
object of class huf
}
\description{
\code{read_huf} reads in a MODFLOW hydrogeologic unit flow file and returns it as an \code{\link{RMODFLOW}} huf object.
}

