% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mlt.R
\name{read_mlt}
\alias{read_mlt}
\title{Read a MODFLOW multiplier file}
\usage{
read_mlt(file = {     cat("Please select mlt file ...\\n")     file.choose()
  }, dis = {     cat("Please select dis file ...\\n")    
  read_dis(file.choose()) })
}
\arguments{
\item{file}{filename; typically '*.mlt'}

\item{dis}{discretization file object; defaults to that with the same filename but with extension '.dis'}
}
\value{
object of class mlt
}
\description{
\code{read_mlt} reads in a MODFLOW multiplier file and returns it as an \code{\link{RMODFLOW}} mlt object.
}

