% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_nam.R
\name{read_nam}
\alias{read_nam}
\title{Read a MODFLOW name file}
\usage{
read_nam(file = {     cat("Please select nam file...\\n")     file.choose() })
}
\arguments{
\item{file}{filename; typically '*.nam'}
}
\value{
object of class nam
}
\description{
\code{read_nam} reads in a MODFLOW name file and returns it as an \code{\link{RMODFLOW}} nam object.
}

