% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_prj}
\alias{read_prj}
\title{Read a projection file}
\usage{
read_prj(file)
}
\arguments{
\item{file}{filename; typically '*.prj'}
}
\value{
object of class prj
}
\description{
\code{read_prj} reads in projection file and returns it as a prj object.
}

