% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_prj.R
\name{read_prj}
\alias{read_prj}
\title{Read a projection file}
\usage{
read_prj(file = {     cat("Please select prj file...\\n")     file.choose() })
}
\arguments{
\item{file}{filename; typically '*.prj'}
}
\value{
object of class prj
}
\description{
\code{read_prj} reads in projection file and returns it as a prj object.
}

