% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_pvl}
\alias{read_pvl}
\title{Read a MODFLOW parameter value file}
\usage{
read_pvl(file, read_all = F)
}
\arguments{
\item{file}{filename; typically '*.pvl'}

\item{read_all}{logical, indicating if \code{NP} parameters should be read, or the full parameter table (only relevant if external codes use the pvl file for storing additional parameters).}
}
\value{
object of class pvl
}
\description{
\code{read_pvl} reads in a MODFLOW parameter value file and returns it as an \code{\link{RMODFLOW}} pvl object.
}

