% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflowDataFuncs.R
\name{readCBCbinByTerm}
\alias{readCBCbinByTerm}
\title{Read budget term by layer}
\usage{
readCBCbinByTerm(filPtr, term, SP_rng, lay)
}
\arguments{
\item{filPtr, }{term, SP_rng, lay}
}
\value{
bigVector
}
\description{
\code{readCBCbinByTerm} search for a defined budget term
     and returns a vector of values by stress periods identified in range of values
}
