% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflowDataFuncs.R
\name{readCBCbinByTerm}
\alias{readCBCbinByTerm}
\title{Read budget term by layer}
\usage{
readCBCbinByTerm(filPtr, term, SP_rng, lay)
}
\arguments{
\item{term}{Modflow budget term}

\item{SP_rng}{Stress Period range}

\item{lay}{layer to read}

\item{cbbFile}{filename to be referenced by file pointer filPtr}
}
\value{
bigVector
}
\description{
\code{readCBCbinByTerm} search for a defined budget term
     and returns a vector of values by stress periods identified in range of values
}
