% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflowDataFuncs.R
\name{readgridPoints}
\alias{readgridPoints}
\title{Read Model Mesh Points}
\usage{
readgridPoints(Modelgrd.Path, Model.Shape)
}
\arguments{
\item{Modelgrd.Path}{Path to model mesh shapefile}

\item{Model.Shape}{Name of model mesh shapefile without .shp extension}
}
\value{
ModelGridCoords Data.frame providing model row, column and coordinates
}
\description{
\code{readgridPoints} Reads arcGIS shape file of model mesh
   -Converts mesh polygons to points referencing the center of the model cells
   -Changes spatial reference if necessary
   -Converts point coordinates into a data frame
}
\examples{
     Modelgrd.Path <- '//ad.sfwmd.gov/dfsroot/data/wsd/GIS/GISP_2012/DataLib/ModelData/LWCSIM'
     Model.Shape <- 'LWCSIM_mesh'
\dontrun{
     ModelGridCoords \%<-\% readgridPoints(Modelgrd.Path,Model.Shape)
     }
}
