% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflowDataFuncs.R
\name{readHeadsbin}
\alias{readHeadsbin}
\title{Read Heads by Layer}
\usage{
readHeadsbin(filPtr, SP_rng)
}
\arguments{
\item{filPtr, }{SP_rng [stress period range]}
}
\value{
bigVector
}
\description{
\code{readHeadsbin} searchs for a Modflow binary Heads Layer
     and returns a vector of values by stress periods
     identified in range of values
}
