% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rmf_as_array.sf}
\alias{rmf_as_array.sf}
\title{Convert a simple features object to a rmf_array (rasterize)}
\usage{
\method{rmf_as_array}{sf}(
  obj,
  dis,
  select,
  na_value = 0,
  prj = rmf_get_prj(dis),
  sparse = TRUE,
  kper = attr(obj, "kper"),
  ...
)
}
\arguments{
\item{obj}{\code{sf} object}

\item{dis}{\code{RMODFLOW} dis object}

\item{select}{integer or character specifying which column from \code{obj} to rasterize}

\item{na_value}{value of the cells in the array which are not specified in obj; defaults to 0}

\item{prj}{\code{RMODFLOw} prj object}

\item{sparse}{logical; should a 2d (TRUE; default) or 3d (FALSE) array be returned}

\item{kper}{optional integers specifying the stress-periods during which this array is active}

\item{...}{additional arguments passed to \code{stars::st_rasterize}}
}
\value{
a \code{rmf_2d_array} if \code{sparse = TRUE}; a \code{rmf_3d_array} if \code{sparse = FALSE}
}
\description{
Convert a simple features object to a rmf_array (rasterize)
}
\details{
\code{stars::st_rasterize} is used to rasterize \code{obj} to the MODFLOW grid which calls GDALRasterize.
 Alternatively, the user can call \code{rmf_as_list} on \code{obj} and \code{rmf_as_array} on the resulting \code{rmf_list}. Results may differ.
}
\examples{
sfc <- sf::st_sfc(list(sf::st_point(c(100,200)), sf::st_point(c(750, 800)), sf::st_point(c(700, 850))))
obj <- sf::st_sf(q = c(-500, -400, -300), geom = sfc)
dis <- rmf_create_dis()
rmf_as_array(obj, dis = dis, select = 'q')
rmf_as_array(obj, dis = dis, select = 'q', options = c('MERGE_ALG=ADD'))

# alternative
rmf_as_list(obj, dis = dis, select = 'q') \%>\%
  rmf_as_array(dis = dis)
}
