% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rmf_as_list.sf}
\alias{rmf_as_list.sf}
\title{Title}
\usage{
\method{rmf_as_list}{sf}(obj, dis,
  select = colnames(sf::st_set_geometry(obj, NULL)), prj = NULL,
  k = NULL, kper = attr(obj, "kper"), op = sf::st_intersects)
}
\arguments{
\item{obj}{}

\item{dis}{}

\item{select}{}

\item{prj}{}

\item{kper}{}

\item{op}{}
}
\description{
Title
}
