% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rmf_as_list.stars}
\alias{rmf_as_list.stars}
\title{Title}
\usage{
\method{rmf_as_list}{stars}(obj, dis, select = names(obj), k = NULL,
  prj = NULL, kper = attr(obj, "kper"), op = sf::st_intersects)
}
\arguments{
\item{obj}{}

\item{dis}{}

\item{k}{}

\item{prj}{}

\item{kper}{}

\item{op}{}
}
\description{
Title
}
