% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_as_sf}
\alias{rmf_as_sf}
\title{Generic function to convert rmf_array objects to simple features}
\usage{
rmf_as_sf(...)
}
\description{
Generic function to convert rmf_array objects to simple features
}
