% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-as-tibble.R
\name{rmf_as_tibble}
\alias{rmf_as_tibble}
\title{Generic function to convert rmf_array objects to tibbles}
\usage{
rmf_as_tibble(...)
}
\description{
Generic function to convert rmf_array objects to tibbles
}
