% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-cell-coordinates.R,
%   R/rmf-cell-coordinates.dis.R, R/rmf-cell-coordinates.huf.R
\name{rmf_cell_coordinates}
\alias{rmf_cell_coordinates}
\alias{rmf_cell_coordinates.dis}
\alias{rmf_cell_coordinates.huf}
\title{Generic function to get cell coordinates}
\usage{
rmf_cell_coordinates(...)

\method{rmf_cell_coordinates}{dis}(dis, include_faces = FALSE)

\method{rmf_cell_coordinates}{huf}(huf, dis = NULL, include_faces = FALSE)
}
\arguments{
\item{dis}{dis object}

\item{include_faces}{logical; should face coordinates be included?}

\item{huf}{huf object}

\item{dis}{dis object, corresponding to the huf object}

\item{include_faces}{logical; should face coordinates be included?}
}
\value{
list with with cell coordinate 3d arrays

3d array with cell coordinates
}
\description{
Generic function to get cell coordinates

Get cell x, y and z coordinates from a dis object

Get cell coordinates from a huf object
}
