% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_cell_coordinates.dis}
\alias{rmf_cell_coordinates.dis}
\alias{rmf_cell_coordinates.huf}
\alias{rmf_cell_coordinates}
\title{Get cell x, y and z coordinates from a dis object}
\usage{
\method{rmf_cell_coordinates}{dis}(dis, prj = NULL,
  include_faces = FALSE)

\method{rmf_cell_coordinates}{huf}(huf, dis = NULL, prj = NULL,
  include_faces = FALSE)

rmf_cell_coordinates(...)
}
\arguments{
\item{dis}{dis object}

\item{prj}{projection file object}

\item{include_faces}{logical; should face coordinates be included?}

\item{huf}{huf object}

\item{dis}{dis object, corresponding to the huf object}

\item{prj}{projection file object}

\item{include_faces}{logical; should face coordinates be included?}
}
\value{
list with with cell coordinate 3d arrays

3d array with cell coordinates
}
\description{
Get cell x, y and z coordinates from a dis object

Get cell coordinates from a huf object

Generic function to get cell coordinates
}
