% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_cell_dimensions}
\alias{rmf_cell_dimensions}
\alias{rmf_cell_dimensions.dis}
\alias{rmf_cell_dimensions.huf}
\title{Generic function to get cell dimensions}
\usage{
rmf_cell_dimensions(...)

\method{rmf_cell_dimensions}{dis}(
  dis,
  hed = NULL,
  include_volume = FALSE,
  include_faces = FALSE
)

\method{rmf_cell_dimensions}{huf}(
  huf,
  dis = NULL,
  hed = NULL,
  include_volume = FALSE,
  include_faces = FALSE
)
}
\arguments{
\item{dis}{dis object}

\item{hed}{hed object, used for calculating the saturated thickness; if not specified, the regular cell thickness is returned}

\item{include_volume}{logical; should the cell volumes be included?}

\item{include_faces}{logical; should face areas be included?}

\item{huf}{huf object}
}
\value{
list with x y and z cell dimension 3d arrays and optionally, cell volume
}
\description{
Generic function to get cell dimensions
}
