% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_cell_info}
\alias{rmf_cell_info}
\alias{rmf_cell_info.dis}
\alias{rmf_cell_info.huf}
\title{Generic function to print information at a certain grid cell}
\usage{
rmf_cell_info(...)

\method{rmf_cell_info}{dis}(dis, i, j)

\method{rmf_cell_info}{huf}(huf, i, j)
}
\arguments{
\item{dis}{a discretization file object}

\item{i}{row number}

\item{j}{column number}

\item{huf}{a hydrogeologic unit file object}
}
\value{
\code{NULL}
}
\description{
Generic function to print information at a certain grid cell
}
