% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_cell_info.dis}
\alias{rmf_cell_info.dis}
\alias{rmf_cell_info.huf}
\alias{rmf_cell_info}
\title{Get information from a dis object at a certain grid cell}
\usage{
\method{rmf_cell_info}{dis}(dis, i, j)

\method{rmf_cell_info}{huf}(huf, i, j)

rmf_cell_info(...)
}
\arguments{
\item{dis}{a discretization file object}

\item{i}{row number}

\item{j}{column number}

\item{huf}{a hydrogeologic unit file object}

\item{i}{row number}

\item{j}{column number}
}
\value{
\code{NULL}

\code{NULL}
}
\description{
Get information from a dis object at a certain grid cell

Get information from a huf object at a certain grid cell

Generic function to get information at a certain grid cell
}
