% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_convert_dis_to_saturated_dis}
\alias{rmf_convert_dis_to_saturated_dis}
\alias{convert_dis_to_saturated_dis}
\title{Convert a dis object to correspond to the saturated volume}
\usage{
rmf_convert_dis_to_saturated_dis(dis, hed, l = NULL)

convert_dis_to_saturated_dis(...)
}
\arguments{
\item{dis}{dis object}

\item{hed}{hed object}
}
\value{
dis object
}
\description{
Convert a dis object to correspond to the saturated volume
}
\section{Functions}{
\itemize{
\item \code{convert_dis_to_saturated_dis}: Deprecated function name
}}

