% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_convert_hob_to_time_series}
\alias{rmf_convert_hob_to_time_series}
\alias{convert_hob_to_time_series}
\title{Convert a hob object to a time series data frame}
\usage{
rmf_convert_hob_to_time_series(hob, dis, prj)

convert_hob_to_time_series(...)
}
\arguments{
\item{hob}{hob object}

\item{dis}{dis object}

\item{prj}{prj object}
}
\value{
time series data frame containing name, time and head columns
}
\description{
Convert a hob object to a time series data frame
}
\section{Functions}{
\itemize{
\item \code{convert_hob_to_time_series}: Deprecated function name
}}

