% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_convert_huf_to_dis}
\alias{rmf_convert_huf_to_dis}
\title{Convert a huf to a dis object (for plotting)}
\usage{
rmf_convert_huf_to_dis(huf, dis)
}
\arguments{
\item{huf}{huf object}

\item{dis}{dis object, corresponding to the huf object}
}
\value{
dis object containing the layers of the huf object
}
\description{
Convert a huf to a dis object (for plotting)
}
