% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_convert_huf_to_mask}
\alias{rmf_convert_huf_to_mask}
\title{Convert a huf to a mask object}
\usage{
rmf_convert_huf_to_mask(huf, dis, bas = NULL)
}
\arguments{
\item{huf}{huf object}

\item{dis}{dis object, corresponding to the huf object}

\item{bas}{bas object, corresponding to the huf object; defaults to NULL}
}
\value{
mask rmf_3d_array
}
\description{
Convert a huf to a mask object
}
