% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modflow.R
\name{rmf_create}
\alias{rmf_create}
\alias{create_modflow}
\title{Create an \code{RMODFLOW} modflow object}
\usage{
rmf_create(...)

create_modflow(...)
}
\arguments{
\item{...}{RMODFLOW objects to be included in the modflow model; if a nam object is not provided, it as added automatically}
}
\value{
Object of class modflow
}
\description{
\code{rmf_create} creates an \code{RMODFLOW} modflow object.
}
\section{Functions}{
\itemize{
\item \code{create_modflow}: Deprecated function name
}}

\seealso{
\code{\link{rmf_read}}, \code{\link{rmf_write}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html}
}
