% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-create-array.R
\name{rmf_create_array}
\alias{rmf_create_array}
\alias{create_rmodflow_array}
\title{Add rmf array class to object based on object dimensions}
\usage{
rmf_create_array(obj = NA, dim = NULL)

create_rmodflow_array(...)
}
\arguments{
\item{obj}{object to add class to}

\item{dim}{the dim attribute for the array to be created; by default, dim(obj) is used}
}
\description{
Add rmf array class to object based on object dimensions
}
\section{Functions}{
\itemize{
\item \code{create_rmodflow_array}: Deprecated function name
}}

