% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-chd.R
\name{rmf_create_chd}
\alias{rmf_create_chd}
\title{Create an \code{RMODFLOW} chd object.}
\usage{
rmf_create_chd(..., dis, noprint = FALSE, aux = NULL)
}
\arguments{
\item{...}{\code{rmf_list} (possibly of class \code{rmf_parameter}) objects or a single \code{list} with \code{rmf_list} objects (possibly of class \code{rmf_parameter}) elements; defines the constant head cells}

\item{dis}{dis object}

\item{noprint}{logical, should the printing of CHD cells to the listing file be suppressed ? Defaults to \code{FALSE}}

\item{aux}{optional character vector specifying the names of the auxiliary variables. These variables should also be included in the \code{rmf_list} objects that are supplied; defaults to \code{NULL}}
}
\value{
\code{RMODFLOW} chd object
}
\description{
\code{rmf_create_chd} creates an \code{RMODFLOW} chd object
}
\seealso{
\code{\link{rmf_read_chd}}, \code{\link{rmf_write_chd}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?chd.htm}
}
