% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-create-hfb.R
\name{rmf_create_hfb}
\alias{rmf_create_hfb}
\title{Create an \code{RMODFLOW} hfb object.}
\usage{
rmf_create_hfb(nphfb = 0, mxfb = 0, nhfbnp = 1, option = NULL,
  parnam = NULL, parval = NULL, nlst = NULL, layer_parm = NULL,
  irow1_parm = NULL, icol1_parm = NULL, irow2_parm = NULL,
  icol2_parm = NULL, factor_parm = NULL, layer_noparm = 2,
  irow1_noparm = 3, icol1_noparm = 7, irow2_noparm = 4,
  icol2_noparm = 7, hydchr_noparm = -1.5, nacthfb = 0,
  pname = NULL)
}
\arguments{
\item{nphfb}{number of hfb parameters; defaults to 0}

\item{mxfb}{maximum number of hfb barriers that will be defined using parmeters; defaults to 0}

\item{nhfbnp}{number of hfb barriers not defined by parameters; defaults to 1}

\item{option}{optional character vector specifying \code{NOPRINT}; defaults to NULL}

\item{parnam}{vector of length \code{nphfb} specifying the parameter names; defaults to NULL}

\item{parval}{vector of length \code{nphfb} specifying the parameter values; defaults to NULL}

\item{nlst}{vector of length \code{nphfb} specifying the number of horizontal-flow barrier cells included in the parameter; defaults to NULL}

\item{layer_parm}{list with \code{nphfb} elements where each element \code{i} is a numeric vector of length \code{nlst} for parameter \code{i} specifying the layer numbers of the model cells containing the horizontal-flow barriers for parameter \code{i}; defaults to NULL}

\item{irow1_parm}{list with \code{nphfb} elements where each element \code{i} is a numeric vector of length \code{nlst} for parameter \code{i} specifying the row numbers of the model cells on one side of the cells containing the horizontal-flow barriers for parameter \code{i}; defaults to NULL}

\item{icol1_parm}{list with \code{nphfb} elements where each element \code{i} is a numeric vector of length \code{nlst} for parameter \code{i} specifying the column numbers of the model cells on one side of the cells containing the horizontal-flow barriers for parameter \code{i}; defaults to NULL}

\item{irow2_parm}{list with \code{nphfb} elements where each element \code{i} is a numeric vector of length \code{nlst} for parameter \code{i} specifying the row numbers of the model cells on the other side of the cells containing the horizontal-flow barriers for parameter \code{i}; defaults to NULL}

\item{icol2_parm}{list with \code{nphfb} elements where each element \code{i} is a numeric vector of length \code{nlst} for parameter \code{i} specifying the column numbers of the model cells on the other side of the cells containing the horizontal-flow barriers for parameter \code{i}; defaults to NULL}

\item{factor_parm}{list with \code{nphfb} elements where each element \code{i} is a numeric vector of length \code{nlst} for parameter \code{i} specifying the factor used to calculate the hydraulic characteristic from the parameter value of the model cells containing the horizontal-flow barriers for parameter \code{i}. The hydraulic characteristic is the product of factor and the parameter value; defaults to NULL}

\item{layer_noparm}{numeric vector of length \code{nhfbnp} specifying the layer numbers of the model cells containing the horizontal-flow barriers; defaults to 2 for all horizontal-flow barriers}

\item{irow1_noparm}{numeric vector of length \code{nhfbnp} specifying the row numbers of the model cells on one side of the cells containing the horizontal-flow barriers; defaults to 3 for all horizontal-flow barriers}

\item{icol1_noparm}{numeric vector of length \code{nhfbnp} specifying the column numbers of the model cells on one side of the cells containing the horizontal-flow barriers; defaults to 7 for all horizontal-flow barriers}

\item{irow2_noparm}{numeric vector of length \code{nhfbnp} specifying the row numbers of the model cells on the other side of the cells containing the horizontal-flow barriers; defaults to 4 for all horizontal-flow barriers}

\item{icol2_noparm}{numeric vector of length \code{nhfbnp} specifying the column numbers of the model cells on the other side of the cells containing the horizontal-flow barriers; defaults to 7 for all horizontal-flow barriers}

\item{hydchr_noparm}{numeric vector of length \code{nhfbnp} specifying the hydraulic characteristic of the model cells containing the horizontal-flow barriers. The hydraulic characteristic is the product of factor and the parameter value; defaults to -0.01}

\item{nacthfb}{number of active hfb parameters; defaults to 0}

\item{pname}{character vector of length \code{nacthfb} specifying the names of the active hfb parameters; defaults to NULL}
}
\value{
\code{RMODFLOW} hfb object
}
\description{
\code{rmf_create_hfb} creates an \code{RMODFLOW} hfb object
}
\seealso{
\code{\link{rmf_read_hfb}}, \code{\link{rmf_write_hfb}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?hfb6.htm}
}
