% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-create-lpf.R
\name{rmf_create_lpf}
\alias{rmf_create_lpf}
\alias{create_lpf}
\title{Create an \code{RMODFLOW} lpf object}
\usage{
rmf_create_lpf(dis = create_dis(), ilpfcb = 0, hdry = -888,
  nplpf = 0, storagecoefficient = FALSE, constantcv = FALSE,
  thickstrt = FALSE, nocvcorrection = FALSE, novfc = FALSE,
  noparcheck = FALSE, laytyp = ifelse(dis$nlay == 1, list(1), list(c(1,
  rep(0, dis$nlay - 1))))[[1]], layavg = laytyp * 0, chani = rep(1,
  dis$nlay), layvka = rep(0, dis$nlay), laywet = rep(0, dis$nlay),
  wetfct = 0.1, iwetit = 1, ihdwet = 0, parnam = NULL,
  partyp = NULL, parval = NULL, nclu = NULL, mltarr = NULL,
  zonarr = NULL, iz = NULL, hk = rmf_create_array(1e-04, dim =
  c(dis$nrow, dis$ncol, dis$nlay)), hani = rmf_create_array(1, dim =
  c(dis$nrow, dis$ncol, dis$nlay)), vka = hk,
  ss = rmf_create_array(1e-05, dim = c(dis$nrow, dis$ncol, dis$nlay)),
  sy = rmf_create_array(0.15, dim = c(dis$nrow, dis$ncol, dis$nlay)),
  vkcb = rmf_create_array(0, dim = c(dis$nrow, dis$ncol, dis$nlay)),
  wetdry = rmf_create_array(-0.01, dim = c(dis$nrow, dis$ncol,
  dis$nlay)))

create_lpf(...)
}
\arguments{
\item{dis}{RMODFLOW dis object}

\item{ilpfcb}{flag and unit number for writing cell-by-cell flow terms; defaults to 0}

\item{hdry}{head assigned to cells that are converted to dry cells; defaults to -888}

\item{nplpf}{number of lpf parameters; defaults to 0}

\item{storagecoefficient}{logical; should STORAGECOEFFICIENT keyword be included?; defaults to FALSE}

\item{constantcv}{logical; should CONSTANTCV keyword be included?; defaults to FALSE}

\item{thickstrt}{logical; should THICKSTRT keyword be included?; defaults to FALSE}

\item{nocvcorrection}{logical; should NOCVCORRECTION keyword be included?; defaults to FALSE}

\item{novfc}{logical; should NOVFC keyword be included?; defaults to FALSE}

\item{noparcheck}{logical; should NOPARCHECK keyword be included?; defaults to FALSE}

\item{laytyp}{vector of flags for each layer, specifying layer type; defaults to all confined (0) except the first layer (1)}

\item{layavg}{vector of flags for each layer, specifying interblock transmissivity calculation method; defaults to 0 for each layer}

\item{chani}{vector of flags or horizontal anisotropies for each layer; defaults to 1 for each layer}

\item{layvka}{vector of flags for each layer, indicating whether vka is the vertical hydraulic conductivity or the ratio of horizontal to vertical; defaults to 0 for each layer}

\item{laywet}{vector of flags for each layer, indicating if wetting is active; defaults to 0 for each layer}

\item{wetfct}{is a factor that is included in the calculation of the head that is initially established at a cell when it is converted from dry to wet; defaults to 0.1}

\item{iwetit}{is the iteration interval for attempting to wet cells; defaults to 1}

\item{ihdwet}{is a flag that determines which equation is used to define the initial head at cells that become wet; defaults to 0}

\item{parnam}{vector of parameter names; names should not be more than 10 characters, are not case sensitive, and should be unique}

\item{partyp}{vector of parameter types; the lpf parameter types are HK, HANI, VK, VANI, SS, SY, or VKCB}

\item{parval}{vector of parameter values}

\item{nclu}{vector with the number of clusters required for each parameter}

\item{mltarr}{matrix of multiplier array names, with dis$nlay rows and lpf$nplpf columns; cells with non-occurring layer-parameter combinations should be NA}

\item{zonarr}{matrix of zone array names, with dis$nlay rows and lpf$nplpf columns; cells with non-occurring layer-parameter combinations should be NA}

\item{iz}{matrix of zone number combinations separated by spaces, with dis$nlay rows and lpf$nplpf columns; cells with non-occurring layer-parameter combinations should be NA; if zonarr is "ALL", iz should be ""}

\item{hk}{3d array with hydraulic conductivity along rows; defaults to 1}

\item{hani}{3d array with the ratio of hydraulic conductivity along columns to that along rows; defaults to 1}

\item{vka}{3d array with vertical hydraulic conductivity or the ratio of horizontal to vertical; defaults to hk}

\item{ss}{3d array with specific storage; only required when there are transient stress periods; defaults to 1E-5}

\item{sy}{3d array with specific yield; only required when there are transient stress periods; defaults to 0.15}

\item{vkcb}{3d array with vertical hydraulic conductivity of quasi-three-dimensional confining beds; defaults to 0}

\item{wetdry}{3d array with a wetting threshold and flag indicating which neighboring cells can cause a cell to become wet; defaults to -0.01}
}
\value{
Object of class lpf
}
\description{
\code{rmf_create_lpf} creates an \code{RMODFLOW} lpf object.
}
\section{Functions}{
\itemize{
\item \code{create_lpf}: Deprecated function name
}}

\seealso{
\code{\link{read_lpf}}, \code{\link{write_lpf}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?lpf.htm}
}
