% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-create-mlt.R
\name{rmf_create_mlt}
\alias{rmf_create_mlt}
\title{Create an \code{RMODFLOW} mlt object}
\usage{
rmf_create_mlt(nml = 1, mltnam = "MULT", functn = NULL, rmlt = array(1,
  dim = c(10, 10, 1)), operators = NULL, iprn = NULL)
}
\arguments{
\item{nml}{number of multiplier arrays to be defined; defaults to 1}

\item{mltnam}{character vector of length \code{nml} specifying the names of the multiplier arrays; defaults to 'MULT'}

\item{functn}{optional logical vector of length \code{nml} indicating if the multiplier array will be constructed from other multiplier arrays previously defined; defaults to NULL}

\item{rmlt}{numeric 3D array of dimensions \code{dis$nrow x dis$ncol x nml} specifying the mutliplier arrays; defaults to 1 for all cells}

\item{operators}{list with \code{nml} elements where each element is a character vector with the correct function which will be printed for that multiplier array. If no function is to be specifyied for an array, set to NULL; defaults to NULL}

\item{iprn}{numeric vector of length \code{nml} indicating the printing format and whether the multiplier array constructed in data set 4 will be printed to the listing file; defaults to NULL}
}
\value{
an \code{RMODFLOW} mlt object
}
\description{
\code{rmf_create_mlt} creates an \code{RMODFLOW} mlt object
}
\seealso{
\code{\link{rmf_read_mlt}}, \code{\link{rmf_write_mlt}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?mult.htm}
}
