% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-oc.R
\name{rmf_create_oc}
\alias{rmf_create_oc}
\alias{create_oc}
\title{Create an \code{RMODFLOW} oc object}
\usage{
rmf_create_oc(dis = create_dis(), ihedfm = NA, chedfm = NA,
  ihedun = 666, iddnfm = NA, cddnfm = NA, iddnun = NA,
  cboufm = NA, ibouun = NA, compact_budget = TRUE, aux = FALSE,
  head_label = TRUE, drawdown_label = TRUE, ibound_label = TRUE,
  iperoc = rep(1:dis$nper, dis$nstp),
  itsoc = unlist(apply(data.frame(1, dis$nstp), 1, FUN = function(x)
  x[1]:x[2])), print_head = rep(FALSE, length(iperoc)),
  print_drawdown = rep(FALSE, length(iperoc)),
  print_budget = rep(FALSE, length(iperoc)), save_head = rep(TRUE,
  length(iperoc)), save_drawdown = rep(FALSE, length(iperoc)),
  save_ibound = rep(FALSE, length(iperoc)), save_budget = rep(TRUE,
  length(iperoc)), incode = NULL, ihddfl = NULL, ibudfl = NULL,
  icbcfl = NULL, hdpr = NULL, ddpr = NULL, hdsv = NULL,
  ddsv = NULL)

create_oc(...)
}
\arguments{
\item{dis}{dis object}

\item{ihedfm}{code for the format in which heads will be printed; defaults to NA (heads are not printed)}

\item{chedfm}{code for the format in which heads will be saved; defaults to NA (heads are saved in binary file)}

\item{ihedun}{unit number on which heads will be saved; defaults to 666}

\item{iddnfm}{code for the format in which drawdowns will be printed; defaults to NA (drawdowns are not printed)}

\item{cddnfm}{code for the format in which drawdowns will be saved; defaults to NA (drawdowns are saved in binary file)}

\item{iddnun}{unit number on which drawdowns will be saved; defaults to NA (drawdowns are not saved)}

\item{cboufm}{code for the format in which ibound will be saved; defaults to NA (ibound is saved in binary file)}

\item{ibouun}{unit number on which ibound will be saved; defaults to NA (ibound is not saved)}

\item{compact_budget}{logical; should the COMPACT BUDGET option be used}

\item{aux}{logical; should the auxiliary data be included in the budget file}

\item{head_label}{logical; should labels be included in the unformatted head file?}

\item{drawdown_label}{logical; should labels be included in the unformatted drawdown file?}

\item{ibound_label}{logical; should labels be included in the unformatted ibound file?}

\item{iperoc}{vector of stress period numbers}

\item{itsoc}{vector of time step numbers (within the corresponding stress period)}

\item{print_head}{logical vector or logical matrix of dimensions NLAY x length(iperoc); should heads be printed for the time step corresponding to iperoc and itsoc? If not a matrix, all layers are treated the same.}

\item{print_drawdown}{logical vector or logical matrix of dimensions NLAY x length(iperoc); should drawdowns be printed for the time step corresponding to iperoc and itsoc? If not a matrix, all layers are treated the same.}

\item{print_budget}{logical vector; should budget be printed for the time step corresponding to iperoc and itsoc?}

\item{save_head}{logical vector or logical matrix of dimensions NLAY x length(iperoc); should heads be saved for the time step corresponding to iperoc and itsoc? If not a matrix, all layers are treated the same.}

\item{save_drawdown}{logical vector or logical matrix of dimensions NLAY x length(iperoc); should drawdowns be saved for the time step corresponding to iperoc and itsoc? If not a matrix, all layers are treated the same.}

\item{save_ibound}{logical vector or logical matrix of dimensions NLAY x length(iperoc); should ibound be saved for the time step corresponding to iperoc and itsoc? If not a matrix, all layers are treated the same.}

\item{save_budget}{logical vector; should budget be saved for the time step corresponding to iperoc and itsoc?}

\item{incode}{vector of length \code{sum(dis$nstp)}; used when OC is specified with numeric codes; defaults to NULL (OC is specified using words)}

\item{ihddfl}{vector of length \code{sum(dis$nstp)}; used when OC is specified with numeric codes; defaults to NULL (OC is specified using words)}

\item{ibudfl}{vector of length \code{sum(dis$nstp)}; used when OC is specified with numeric codes; defaults to NULL (OC is specified using words)}

\item{icbcfl}{vector of length \code{sum(dis$nstp)}; used when OC is specified with numeric codes; defaults to NULL (OC is specified using words)}

\item{hdpr}{matrix of dimensions \code{sum(dis$nstp), dis$nlay}; used when OC is specified with numeric codes; defaults to NULL (OC is specified using words)}

\item{ddpr}{matrix of dimensions \code{sum(dis$nstp), dis$nlay}; used when OC is specified with numeric codes; defaults to NULL (OC is specified using words)}

\item{hdsv}{matrix of dimensions \code{sum(dis$nstp), dis$nlay}; used when OC is specified with numeric codes; defaults to NULL (OC is specified using words)}

\item{ddsv}{matrix of dimensions \code{sum(dis$nstp), dis$nlay}; used when OC is specified with numeric codes; defaults to NULL (OC is specified using words)}
}
\value{
Object of class oc
}
\description{
\code{rmf_create_oc} creates an \code{RMODFLOW} oc object.
}
\section{Functions}{
\itemize{
\item \code{create_oc}: Deprecated function name
}}

\seealso{
\code{\link{rmf_read_oc}}, \code{\link{rmf_write_oc}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?oc.htm}
}
