% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-pval.R
\name{rmf_create_pval}
\alias{rmf_create_pval}
\title{Create an \code{RMODFLOW} pval object}
\usage{
rmf_create_pval(parnam, parval)
}
\arguments{
\item{parnam}{character vector specifying the parameter names; defaults to NULL}

\item{parval}{numeric vector specifying the parameter values; defaults to NULL+}
}
\value{
an \code{RMODFLOW} pval object
}
\description{
\code{rmf_create_pval} creates an \code{RMODFLOW} pval object
}
\details{
parnam & parval should be of the same length
}
\seealso{
\code{\link{rmf_read_pval}}, \code{\link{rmf_write_pval}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?parameter_value_file.htm}
}
