% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-create-rch.R
\name{rmf_create_rch}
\alias{rmf_create_rch}
\title{Create an \code{RMODFLOW} rch object.}
\usage{
rmf_create_rch(..., dis, nrchop = 3, irchcb = 0, irch = NULL)
}
\arguments{
\item{...}{\code{rmf_2d_array's} (possibly of class \code{rmf_parm}) or a single \code{list} with \code{rmf_2d_array's} (possibly of class \code{rmf_parm}) elements; defines the recharge values. See details.}

\item{dis}{\code{RMODFLOW} dis object}

\item{nrchop}{recharge option code; defaults to 3 (recharge is applied to the highest active cell in each vertical column)}

\item{irchcb}{flag and unit number for writing cell-by-cell flow terms; defaults to 0}

\item{irch}{list of \code{rmf_2d_array's} specifying the layer numbers defining in which layer recharge is applied. The \code{'kper'} attribute of the arrays define the stress period in which the array is active. Only used when \code{nrchop = 2}. Defaults to NULL}
}
\value{
\code{RMODFLOW} rch object
}
\description{
\code{rmf_create_rch} creates an \code{RMODFLOW} rch object
}
\details{
the \code{rmf_2d_array's} should have \code{kper} attributes specifying the stress period in which they are active. This is also true for the irch arrays.
}
\seealso{
\code{\link{rmf_read_rch}}, \code{\link{rmf_write_rch}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?rch.htm}
}
