% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-create-riv.R
\name{rmf_create_riv}
\alias{rmf_create_riv}
\title{Create an \code{RMODFLOW} riv object.}
\usage{
rmf_create_riv(..., dis, irivcb = 0, noprint = FALSE, aux = NULL)
}
\arguments{
\item{...}{\code{rmf_list} (possibly of class \code{rmf_parm}) objects or a single \code{list} with \code{rmf_list} objects (possibly of class \code{rmf_parm}) elements; defines the rivers.}

\item{dis}{dis object}

\item{irivcb}{flag and unit number for writing cell-by-cell flow terms; defaults to 0 (cell-by-cell flow terms will not be written)}

\item{noprint}{logical, should the printing of RIV cells to the listing file be suppressed ? Defaults to \code{FALSE}}

\item{aux}{optional character vector specifying the names of the auxiliary variables. These variables should also be included in the \code{rmf_list} objects that are supplied; defaults to \code{NULL}}
}
\value{
\code{RMODFLOW} riv object
}
\description{
\code{rmf_create_riv} creates an \code{RMODFLOW} riv object
}
\seealso{
\code{\link{rmf_read_riv}}, \code{\link{rmf_write_riv}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?riv.htm}
}
