% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-create-zon.R
\name{rmf_create_zon}
\alias{rmf_create_zon}
\title{Create an \code{RMODFLOW} zon object}
\usage{
rmf_create_zon(nzn = 1, zonnam = "ZONE", izon = array(1, dim = c(10, 10,
  1)))
}
\arguments{
\item{nzn}{number of zone arrays to be defined; defaults to 1}

\item{zonnam}{character vector of length \code{nzn} specifying the names of zone arrays; defaults to 'ZONE'}

\item{izon}{numeric 3D array of dimensions \code{dis$nrow x dis$ncol x nzn} specifying the zone arrays; defaults to 1 for all cells}
}
\value{
an \code{RMODFLOW} zon object
}
\description{
\code{rmf_create_zon} creates an \code{RMODFLOW} zon object
}
\seealso{
\code{\link{rmf_read_zon}}, \code{\link{rmf_write_zon}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?zone.htm}
}
