% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-install.R
\name{rmf_install}
\alias{rmf_install}
\title{Install external software}
\usage{
rmf_install(name = "MODFLOW-2005")
}
\arguments{
\item{name}{Character.}
}
\description{
Install external software
}
\details{
Currently supported software: All MODFLOW variants documented at
\url{https://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/}, except for
MODFLOW-2000.I.e. MODFLOW-2005, MODFLOW-OWHM, MODFLOW-NWT, MODFLOW-LGR and
MODFLOW-CFP. The zip files with windows binaries hosted at the USGS websites
are downloaded and extracted in \file{C:/WRDAPP/} each time. The main folder
names are modified in order to have more consistency.
}
\examples{
rmf_install("MODFLOW-2005")
}
