% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_performance}
\alias{rmf_performance}
\alias{rmf_performance.default}
\alias{rmf_performance.hpr}
\title{Generic function to get model performance measures}
\usage{
rmf_performance(...)

\method{rmf_performance}{default}(sim, obs, na_value = -888,
  measures = c("ssq", "mse", "mae", "me", "r2", "nse", "rmse", "pbias",
  "kge"), ...)

\method{rmf_performance}{hpr}(hpr, hobdry = -888, measures = c("ssq",
  "mse", "mae", "me", "r2", "nse", "rmse", "pbias", "kge"), ...)
}
\arguments{
\item{...}{arguments passes to \code{\link{hydroGOF::gof}}}

\item{sim}{numeric vector with simulated values}

\item{obs}{numeric vector with observed values corresponding to sim}

\item{na_value}{numeric; flags values to remove from the calculations; defaults to -888}

\item{measures}{character vector with the required performance measures. Possible values are any of the measures present in \code{\link{hydroGOF::gof}} + 'ssq' (sum of squared errors)}

\item{hpr}{head predictions file object}

\item{hobdry}{value used to flag dry cells; defaults to -888}

\item{measures}{character vector with the required performance measures. Possible values are any of the measures present in \code{\link{hydroGOF::gof}} + 'ssq' (sum of squared errors)}

\item{...}{arguments passes to \code{\link{hydroGOF::gof}}}
}
\value{
data.frame with performance measures

data.frame with performance measures
}
\description{
Generic function to get model performance measures

Get model performance measures

Get model performance measures from a hpr object
}
