% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-performance.R, R/rmf-performance.hpr.R
\name{rmf_performance}
\alias{rmf_performance}
\alias{rmf_performance.hpr}
\title{Generic function to get model performance measures}
\usage{
rmf_performance(...)

\method{rmf_performance}{hpr}(hpr)
}
\arguments{
\item{hpr}{head predictions file object}
}
\value{
performance measures
}
\description{
Generic function to get model performance measures

Get model performance measures from a hpr object
}
