% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmf_plot.evt}
\alias{rmf_plot.evt}
\title{Plot a RMODFLOW evt object}
\usage{
\method{rmf_plot}{evt}(evt, dis, kper = NULL, variable = "evt", ...)
}
\arguments{
\item{evt}{an \code{RMODFLOW} evt object}

\item{dis}{a \code{RMODFLOW} dis object}

\item{kper}{integer specifying the stress-period to plot}

\item{variable}{character specifying which variable to plot. Possible values are 'evt' (default), 'surf', 'exdp' and 'ievt' if defined.}

\item{...}{additional arguments passed to \code{\link{rmf_plot.rmf_2d_array}}}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
Plot a RMODFLOW evt object
}
