% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-plot.ghb.R
\name{rmf_plot.ghb}
\alias{rmf_plot.ghb}
\title{Plot a 2D section of an RMODFLOW ghb object}
\usage{
\method{rmf_plot}{ghb}(ghb, dis, all_parm = F, variable = "identity",
  instnum = NULL, l = NULL, sp = NULL, ...)
}
\arguments{
\item{ghb}{an \code{RMODFLOW} ghb object}

\item{dis}{an \code{RMODFLOW} dis object}

\item{all_parm}{logical, should all parameters defined by plotted (i.e. indepedent of stress periods); defaults to FALSE}

\item{variable}{character, what data should be plotted. Possible values are: "identity" (default; plots the head-boundary locations), "layer", "row", "column", "bhead", "condfact" (for parameter data), "conductance", "parnam" (for parameter data), "instnam" (for time-varying parameter data) and "parval" (for parameter data); defaults to "identity"}

\item{instnum}{numeric vector of length \code{npghb} holding the instance numbers for each time-varying parameter which need to be plotted. Only one instance per parameter is allowed. If a certain parameter \code{i} is not time-varying, specify instnum[i] as '1'; defaults to NULL}

\item{l}{time step number for selecting which stress period to plot; defaults to NULL (last stress period)}

\item{sp}{optional stress period number to plot; will override the stress period calculated from \code{l}; defaults to NULL}

\item{...}{additional arguments passed to \code{rmf_plot.rmf-3d-array}}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
\code{rmf_plot.ghb} plots a 2D section of an \code{RMODFLOW} ghb object using \code{rmf_plot.rmf-3d-array}
}
