% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmf_plot.hob}
\alias{rmf_plot.hob}
\title{Plot a RMODFLOW hob object}
\usage{
\method{rmf_plot}{hob}(hob, dis, i = NULL, j = NULL, k = NULL, prj = NULL, kper = NULL, ...)
}
\arguments{
\item{hob}{\code{RMODFLOW} hob object}

\item{dis}{\code{RMODFLOW} dis object}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{prj}{projection file object; defaults to NULL}

\item{kper}{integer specifying the stress-period to plot}

\item{...}{additional arguments passed to \code{\link{rmf_plot.rmf_list}}}
}
\value{
ggplot2 object or layer
}
\description{
Plot a RMODFLOW hob object
}
\details{
specifying all of the \code{i, j & k} arguments will plot a time series at that cell location. The observations are grouped by name.
}
