% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmf_plot.hpr}
\alias{rmf_plot.hpr}
\title{Plot a MODFLOW head predictions file}
\usage{
\method{rmf_plot}{hpr}(hpr, type = "scatter")
}
\arguments{
\item{hpr}{head predictions file object}

\item{type}{plot type: 'scatter' or 'residual'}
}
\description{
Plot a MODFLOW head predictions file
}
