% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-plot.rmf-3d-array.R
\name{rmf_plot.rmf_3d_array}
\alias{rmf_plot.rmf_3d_array}
\title{Plot a 2D section through a MODFLOW 3D array}
\usage{
\method{rmf_plot}{rmf_3d_array}(array, i = NULL, j = NULL, k = NULL,
  dis, bas = NULL, mask = rmfi_ifelse0(is.null(bas), array * 0 + 1,
  rmfi_ifelse0(bas$xsection, aperm(bas$ibound, c(3, 2, 1)), bas$ibound)),
  zlim = range(array[rmfi_ifelse0(is.null(i), c(1:dim(array)[1]), i),
  rmfi_ifelse0(is.null(j), c(1:dim(array)[2]), j), rmfi_ifelse0(is.null(k),
  c(1:dim(array)[3]), k)][as.logical(mask[rmfi_ifelse0(is.null(i),
  c(1:dim(array)[1]), i), rmfi_ifelse0(is.null(j), c(1:dim(array)[2]), j),
  rmfi_ifelse0(is.null(k), c(1:dim(array)[3]), k)])], finite = TRUE),
  colour_palette = rmfi_rev_rainbow, nlevels = 7, type = "fill",
  levels = NULL, gridlines = FALSE, add = FALSE, title = NULL,
  crop = FALSE, hed = NULL, l = NULL, prj = NULL, crs = NULL,
  ...)
}
\arguments{
\item{array}{an object of class rmf_3d_array}

\item{i}{row number to plot}

\item{j}{column number to plot}

\item{k}{layer number to plot}

\item{dis}{discretization file object}

\item{bas}{basic file object; optional}

\item{mask}{a 3D array with 0 or F indicating inactive cells optional; defaults to having all cells active or, if bas is provided, bas$ibound}

\item{zlim}{vector of minimum and maximum value for the colour scale}

\item{colour_palette}{a colour palette for imaging the array values}

\item{nlevels}{number of levels for the colour scale; defaults to 7}

\item{type}{plot type: 'fill' (default), 'factor' or 'grid'}

\item{levels}{labels that should be used on the factor legend; if NULL the array factor levels are used}

\item{gridlines}{logical; should grid lines be plotted? alternatively, provide colour of the grid lines.}

\item{title}{plot title}

\item{crop}{logical; should plot be cropped by dropping NA values (as set by mask)}

\item{hed}{hed object for only plotting the saturated part of the grid; possibly subsetted with time step number; by default, last time step is used}

\item{l}{time step number for subsetting the hed object}

\item{prj}{projection file object}

\item{crs}{coordinate reference system for the plot}

\item{...}{parameters provided to plot.rmf_2d_array}
}
\value{
ggplot2 object or layer; if plot3D is TRUE, nothing is returned and the plot is made directly
}
\description{
\code{rmf_plot.rmf_3d_array} plots a 2D section through a MODFLOW 3D array.
}
