% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rmf_plot.rmf_list}
\alias{rmf_plot.rmf_list}
\title{Plot a RMODFLOW list object}
\usage{
\method{rmf_plot}{rmf_list}(obj, dis, variable = "id",
  active_only = FALSE, fun = sum, ...)
}
\arguments{
\item{obj}{a \code{RMODFLOW} object of class \code{rmf_list}}

\item{dis}{a \code{RMODFLOW} dis object}

\item{variable}{single character or numeric indicating which column in the \code{rmf_list} object to plot. Defaults to 'id', which plots the locations of the cells.}

\item{active_only}{logical; indicating if only the active cells should be plotted. Non-active cells are set to NA. Defaults to FALSE.}

\item{fun}{function to compute values in the case multiple values are defined for the same MODFLOW cell. Typically either \code{mean} or \code{sum}. Defaults to sum.}

\item{...}{additional arguments passed to \code{\link{rmf_plot.rmf_3d_array}}}
}
\value{
ggplot2 object or layer
}
\description{
Plot a RMODFLOW list object
}
\details{
the rmf_list is converted to a rmf_3d_array using \code{\link{rmf_as_array.rmf_list}}. The sparse argument is set to FALSE.
}
