% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-read-bas.R
\name{rmf_read_bas}
\alias{rmf_read_bas}
\alias{read_bas}
\title{Read a MODFLOW basic file}
\usage{
rmf_read_bas(file = {     cat("Please select bas file ...\\n")    
  file.choose() }, dis = {    
  cat("Please select corresponding dis file ...\\n")    
  rmf_read_dis(file.choose()) })

read_bas(...)
}
\arguments{
\item{file}{filename; typically '*.bas'}

\item{dis}{discretization file object; defaults to that with the same filename but with extension '.dis'}
}
\value{
object of class bas
}
\description{
\code{read_bas} reads in a MODFLOW basic file and returns it as an \code{\link{RMODFLOW}} bas object.
}
\section{Functions}{
\itemize{
\item \code{read_bas}: Deprecated function name
}}

