% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-read-bud.R
\name{rmf_read_bud}
\alias{rmf_read_bud}
\alias{rmf_read_bud}
\title{Read a MODFLOW budget file}
\usage{
rmf_read_bud(...)

rmf_read_bud(...)
}
\arguments{
\item{file}{filename; typically 'Budget.out'}

\item{binary}{logical; is source file binary?}
}
\value{
object of class bud
}
\description{
\code{read_bud} reads in a MODFLOW budget file
}
\section{Functions}{
\itemize{
\item \code{rmf_read_bud}: Deprecated function name
}}

