% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{rmf_read_bud}
\alias{rmf_read_bud}
\alias{rmf_read_budget}
\title{Reads the volumetric budget from a MODFLOW listing file}
\usage{
rmf_read_bud(
  file = {
     cat("Please select listing file ...\\n")
     file.choose()
 }
)

rmf_read_budget(...)
}
\arguments{
\item{file}{path to the listing file; typically '*.lst'}
}
\value{
an object of class bud which is a list with two data frames: one with cumulative fluxes and one with rates
}
\description{
\code{rmf_read_bud} reads a volumetric budget from a MODFLOW listing file and returns it as a list with data frame elements
}
