% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-read-hed.R
\name{rmf_read_hed}
\alias{rmf_read_hed}
\alias{read_hed}
\alias{rmf_read_fhd}
\alias{rmf_read_bhd}
\title{Read a MODFLOW head file}
\usage{
rmf_read_hed(file = {     cat("Please select hed file ...\\n")    
  file.choose() }, dis = {     cat("Please select dis file ...\\n")    
  rmf_read_dis(file.choose()) }, bas = {    
  cat("Please select bas file ...\\n")     rmf_read_bas(file.choose(), dis =
  dis) }, huf = NULL, convert_hnoflo_to_NA = TRUE, binary = TRUE,
  precision = "single")

read_hed(...)

rmf_read_fhd(...)

rmf_read_bhd(...)
}
\arguments{
\item{file}{filename; typically '*.hed'}

\item{dis}{discretization file object}

\item{bas}{basic file object}

\item{huf}{huf object; optional; provide only if huf heads are being read}

\item{convert_hnoflo_to_NA}{logical; should hnoflo values be converted to NA?}
}
\value{
object of class hed
}
\description{
\code{read_hed} reads in a MODFLOW head file and returns it as an \code{\link{RMODFLOW}} hed object.
}
\section{Functions}{
\itemize{
\item \code{read_hed}: Deprecated function name

\item \code{rmf_read_fhd}: Compatible with default ModelMuse file extension

\item \code{rmf_read_bhd}: Compatible with default ModelMuse file extension
}}

