% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-hob.R
\name{rmf_read_hob}
\alias{rmf_read_hob}
\alias{rmf_read_ob_hob}
\title{Read a MODFLOW head observations file}
\usage{
rmf_read_hob(
  file = {
     cat("Please select hob file ...\\n")
     file.choose()
 },
  ...
)

rmf_read_ob_hob(...)
}
\arguments{
\item{file}{filename; typically '*.hob'}

\item{...}{arguments passed to \code{rmfi_parse_variables}. Can be ignored when input is 'free' format.}
}
\value{
object of class hob
}
\description{
\code{read_hob} reads in a MODFLOW head observations file and returns it as an \code{\link{RMODFLOW}} hob object.
}
\section{Functions}{
\itemize{
\item \code{rmf_read_ob_hob()}: Compatible with default ModelMuse file extensions

}}
\seealso{
\code{\link{rmf_create_hob}}, \code{\link{rmf_write_hob}} and \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?hob.htm}
}
