% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-hob.R
\name{rmf_read_hob}
\alias{rmf_read_hob}
\alias{read_hob}
\alias{rmf_read_ob_hob}
\title{Read a MODFLOW head observations file}
\usage{
rmf_read_hob(file = {     cat("Please select hob file ...\\n")    
  file.choose() })

read_hob(...)

rmf_read_ob_hob(...)
}
\arguments{
\item{file}{filename; typically '*.hob'}

\item{...}{arguments passed to \code{rmfi_parse_variables}. Can be ignored when input is 'free' format.}
}
\value{
object of class hob
}
\description{
\code{read_hob} reads in a MODFLOW head observations file and returns it as an \code{\link{RMODFLOW}} hob object.
}
\section{Functions}{
\itemize{
\item \code{read_hob}: Deprecated function name

\item \code{rmf_read_ob_hob}: Compatible with default ModelMuse file extensions
}}

