% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-read-hpr.R
\name{rmf_read_hpr}
\alias{rmf_read_hpr}
\alias{read_hpr}
\alias{rmf_read_hob_out}
\title{Read a MODFLOW head predictions file}
\usage{
rmf_read_hpr(file = {     cat("Please select hpr file ...\\n")    
  file.choose() })

read_hpr(...)

rmf_read_hob_out(...)
}
\arguments{
\item{file}{filename; typically '*.hpr'}
}
\value{
object of class hpr
}
\description{
\code{rmf_read_hpr} reads in a MODFLOW head predictions file and returns it as an \code{\link{RMODFLOW}} hpr object.
}
\section{Functions}{
\itemize{
\item \code{read_hpr}: Deprecated function name

\item \code{rmf_read_hob_out}: Compatible with default ModelMuse file extension
}}

