% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-huf.R
\name{rmf_read_huf}
\alias{rmf_read_huf}
\alias{read_huf}
\title{Read a MODFLOW hydrogeologic unit flow file}
\usage{
rmf_read_huf(file = {     cat("Please select huf file ...\\n")    
  file.choose() }, dis = {    
  cat("Please select corresponding dis file ...\\n")    
  rmf_read_dis(file.choose()) }, ...)

read_huf(...)
}
\arguments{
\item{file}{filename; typically '*.huf'}

\item{dis}{discretization file object; defaults to that with the same filename but with extension '.dis'}

\item{...}{arguments passed to \code{rmfi_parse_array}. Can be ignored when input arrays are free-format and INTERNAL or CONSTANT.}
}
\value{
object of class huf
}
\description{
\code{read_huf} reads in a MODFLOW hydrogeologic unit flow file and returns it as an \code{\link{RMODFLOW}} huf object.
}
\section{Functions}{
\itemize{
\item \code{read_huf}: Deprecated function name
}}

