% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-huf.R
\name{rmf_read_huf}
\alias{rmf_read_huf}
\title{Read a MODFLOW hydrogeologic unit flow file}
\usage{
rmf_read_huf(
  file = {     cat("Please select huf file ...\\n")     file.choose() },
  dis = {     cat("Please select corresponding dis file ...\\n")    
    rmf_read_dis(file.choose()) },
  mlt = NULL,
  zon = NULL,
  ...
)
}
\arguments{
\item{file}{filename; typically '*.huf'}

\item{dis}{discretization file object; defaults to that with the same filename but with extension '.dis'}

\item{mlt}{a \code{RMODFLOW} mlt object. Only needed when reading parameter arrays defined by multiplier arrays}

\item{zon}{a \code{RMODFLOW} zon object. Only needed when reading parameter arrays defined by zone arrays}

\item{...}{arguments passed to \code{rmfi_parse_array}. Can be ignored when input arrays are free-format and INTERNAL or CONSTANT.}
}
\value{
object of class huf
}
\description{
\code{read_huf} reads in a MODFLOW hydrogeologic unit flow file and returns it as an \code{\link{RMODFLOW}} huf object.
}
\details{
\code{\link{rmf_convert_huf_to_grid}} can be used to convert parameters defined on the HUF grid to the numerical grid
}
\seealso{
\code{\link{rmf_convert_huf_to_grid}}, \code{\link{rmf_create_huf}}, \code{\link{rmf_write_huf}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?huf.htm}
}
