% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-read-pcg.R
\name{rmf_read_pcg}
\alias{rmf_read_pcg}
\alias{read_pcg}
\title{Read a MODFLOW preconditioned conjugate-gradient package file}
\usage{
rmf_read_pcg(file = {     cat("Please select pcg file ...\\n")    
  file.choose() }, ...)

read_pcg(...)
}
\arguments{
\item{file}{filename; typically '*.pcg'}
}
\value{
object of class pcg
}
\description{
\code{read_pcg} reads in a MODFLOW preconditioned conjugate-gradient package file and returns it as an \code{\link{RMODFLOW}} pcg object.
#' @param ... arguments passed to \code{rmfi_parse_variables}. Can be ignored when input is 'free' format.
}
\section{Functions}{
\itemize{
\item \code{read_pcg}: Deprecated function name
}}

\seealso{
\code{\link{write_pcg}}, \code{\link{create_pcg}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?pcg.htm}
}
