% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmf_read_prj}
\alias{rmf_read_prj}
\title{Read a projection file}
\usage{
rmf_read_prj(
  file = {     cat("Please select prj file ...\\n")     file.choose() }
)
}
\arguments{
\item{file}{filename; typically '*.prj'}
}
\value{
object of class prj
}
\description{
\code{read_prj} reads in projection file and returns it as a prj object.
}
