% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-pval.R
\name{rmf_read_pval}
\alias{rmf_read_pval}
\title{Read a MODFLOW parameter value file}
\usage{
rmf_read_pval(
  file = {     cat("Please select pval file ...\\n")     file.choose() },
  read_all = FALSE
)
}
\arguments{
\item{file}{filename; typically '*.pval'}

\item{read_all}{logical, indicating if \code{np} parameters should be read, or the full parameter table (only relevant if external codes use the pval file for storing additional parameters).}
}
\value{
object of class pval
}
\description{
\code{rmf_read_pval} reads in a MODFLOW parameter value file and returns it as an \code{\link{RMODFLOW}} pval object.
}
\seealso{
\code{\link{rmf_create_pval}}, \code{\link{rmf_write_pval}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?pval.htm}
}
