% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-write-bas.R
\name{rmf_write_bas}
\alias{rmf_write_bas}
\alias{write_bas}
\title{Write a MODFLOW basic file}
\usage{
rmf_write_bas(bas, file = {    
  cat("Please select bas file to overwrite or provide new filename ...\\n")
      file.choose() }, iprn = -1)

write_bas(...)
}
\arguments{
\item{bas}{an \code{\link{RMODFLOW}} bas object}

\item{file}{filename to write to; typically '*.bas'}

\item{iprn}{format code for printing arrays in the listing file; defaults to -1 (no printing)}
}
\value{
\code{NULL}
}
\description{
\code{write_bas} writes a MODFLOW basic file based on an \code{\link{RMODFLOW}} bas object.
}
\section{Functions}{
\itemize{
\item \code{write_bas}: Deprecated function name
}}

