% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-bcf.R
\name{rmf_write_bcf}
\alias{rmf_write_bcf}
\title{Write a MODFLOW block-centered flow file}
\usage{
rmf_write_bcf(
  bcf,
  dis = {
     cat("Please select corresponding dis file ...\\n")
    
    rmf_read_dis(file.choose())
 },
  file = {
    
    cat("Please select bcf file to overwrite or provide new filename ...\\n")
    
    file.choose()
 },
  iprn = -1,
  ...
)
}
\arguments{
\item{bcf}{an \code{RMODFLOW} bcf object}

\item{dis}{an \code{RMODFLOW} dis object}

\item{file}{filename to write to; typically '*.bcf'}

\item{iprn}{format code for printing arrays in the listing file; defaults to -1 (no printing)}

\item{...}{arguments passed to \code{rmfi_write_array} and \code{rmfi_write_variables}. Can be ignored when format is free and arrays are INTERNAL or CONSTANT.}
}
\value{
\code{NULL}
}
\description{
\code{rmf_write_bcf} writes a MODFLOW block-centered flow file based on an \code{RMODFLOW} bcf object
}
\seealso{
\code{\link{rmf_read_bcf}}, \code{\link{rmf_create_bcf}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?bcf.htm}
}
