% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-write-huf.R
\name{rmf_write_huf}
\alias{rmf_write_huf}
\alias{write_huf}
\title{Write a MODFLOW hydrogeologic unit flow file}
\usage{
rmf_write_huf(huf, file = {    
  cat("Please select huf file to overwrite or provide new filename ...\\n")
      file.choose() }, iprn = -1, ...)

write_huf(...)
}
\arguments{
\item{huf}{an \code{\link{RMODFLOW}} huf object}

\item{file}{filename to write to; typically '*.huf'}

\item{iprn}{format code for printing arrays in the listing file; defaults to -1 (no printing)}

\item{...}{arguments passed to \code{rmfi_write_array}. Can be ignored when arrays are INTERNAL or CONSTANT.}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW hydrogeologic unit flow file
}
\section{Functions}{
\itemize{
\item \code{write_huf}: Deprecated function name
}}

