% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-lpf.R
\name{rmf_write_lpf}
\alias{rmf_write_lpf}
\title{Write a MODFLOW layer-property flow file}
\usage{
rmf_write_lpf(
  lpf,
  file = {    
    cat("Please select lpf file to overwrite or provide new filename ...\\n")    
    file.choose() },
  dis = {     cat("Please select corresponding dis file ...\\n")    
    rmf_read_dis(file.choose()) },
  iprn = -1,
  ...
)
}
\arguments{
\item{lpf}{an \code{\link{RMODFLOW}} lpf object}

\item{file}{filename to write to; typically '*.lpf'}

\item{iprn}{format code for printing arrays in the listing file; defaults to -1 (no printing)}

\item{...}{arguments passed to \code{rmfi_write_array}. Can be ignored when arrays are INTERNAL or CONSTANT.}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW layer-property flow file
}
