% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-nam.R
\name{rmf_write_nam}
\alias{rmf_write_nam}
\alias{write_nam}
\title{Write a MODFLOW name file}
\usage{
rmf_write_nam(nam, file = {    
  cat("Please select nam file to overwrite or provide new filename ...\\n")
      file.choose() })

write_nam(...)
}
\arguments{
\item{nam}{an \code{\link{RMODFLOW}} nam object}

\item{file}{filename to write to; typically '*.nam'}
}
\value{
\code{NULL}
}
\description{
\code{rmf_write_nam} writes a MODFLOW name file based on an \code{\link{RMODFLOW}} nam object.
}
\section{Functions}{
\itemize{
\item \code{write_nam}: Deprecated function name
}}

