% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-nam.R
\name{rmf_write_nam}
\alias{rmf_write_nam}
\title{Write a MODFLOW name file}
\usage{
rmf_write_nam(
  nam,
  file = {
    
    cat("Please select nam file to overwrite or provide new filename ...\\n")
    
    file.choose()
 },
  exclude = NULL
)
}
\arguments{
\item{nam}{an \code{\link{RMODFLOW}} nam object}

\item{file}{filename to write to; typically '*.nam'}

\item{exclude}{character vector with packages names to exclude from the simulation. Defaults to NULL}
}
\value{
\code{NULL}
}
\description{
\code{rmf_write_nam} writes a MODFLOW name file based on an \code{\link{RMODFLOW}} nam object.
}
\seealso{
\code{\link{rmf_read_nam}}, \code{\link{rmf_write_nam}} and \url{http://water.usgs.gov/nrp/gwsoftware/modflow2000/MFDOC/index.html?name_file.htm}
}
