% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf-write-prj.R
\name{rmf_write_prj}
\alias{rmf_write_prj}
\alias{write_prj}
\title{Write an RMODFLOW projection file}
\usage{
rmf_write_prj(prj, file = {    
  cat("Please select prj file to overwrite or provide new filename ...\\n")    
  file.choose() })

write_prj(...)
}
\arguments{
\item{prj}{an \code{\link{RMODFLOW}} prj object}

\item{file}{filename to write to; typically '*.prj'}
}
\description{
\code{write.prj} writes a projection file
}
\section{Functions}{
\itemize{
\item \code{write_prj}: Deprecated function name
}}

