% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-pval.R
\name{rmf_write_pval}
\alias{rmf_write_pval}
\title{Write a MODFLOW parameter value file}
\usage{
rmf_write_pval(pval, path)
}
\arguments{
\item{pval}{an \code{\link{RMODFLOW}} pval object}

\item{path}{Path to write the PVAL file. Typically with extension \code{.pval}.}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW parameter value file
}
\seealso{
\code{\link{rmf_create_pval}}, \code{\link{rmf_read_pval}}, \url{https://water.usgs.gov/ogw/modflow/MODFLOW-2005-Guide/index.html?pval.htm}
}
