% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-pval.R
\name{rmf_write_pvl}
\alias{rmf_write_pvl}
\alias{write_pvl}
\title{Write a MODFLOW parameter value file}
\usage{
rmf_write_pvl(pvl, file = {    
  cat("Please select pvl file to overwrite or provide new filename ...\\n")
      file.choose() })

write_pvl(...)
}
\arguments{
\item{pvl}{an \code{\link{RMODFLOW}} pvl object}

\item{file}{filename to write to; typically '*.pvl'}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW parameter value file
}
\section{Functions}{
\itemize{
\item \code{write_pvl}: Deprecated function name
}}

