% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmfi_create_bc_array}
\alias{rmfi_create_bc_array}
\title{Set array input for a MODFLOW boundary condition package}
\usage{
rmfi_create_bc_array(arg, dis)
}
\arguments{
\item{arg}{list of (1) \code{rmf_2d_array's} and/or rmf_parameter array objects or (2) a single nested \code{list} with \code{rmf_2d_array's} and/or rmf_parameter elements or (3) a \code{matrix}; defines the boundary condition input.}

\item{dis}{dis object. If not explicitely suplied, the function will look in the arg argument for an object of class 'dis'.}
}
\value{
list with the parameters, input arrays and the kper argument
}
\description{
Set array input for a MODFLOW boundary condition package
}
\details{
typically, \code{arg} is \code{list(...)} where the ellipsis contains all the input \code{rmf_arrays} for the \code{rmf_create_*} function. When matrix elements are present, they are coerced to rmf_2d_arrays which are active for all stress-periods with a warning.
}
\seealso{
\code{\link{rmfi_create_bc_list}}, \code{\link{rmfi_write_bc_list}}, \code{\link{rmfi_parse_bc_list}}, \code{\link{rmfi_parse_array_parameters}}, \code{\link{rmfi_write_array_parameters}}
}
\keyword{internal}
