% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmfi-parse-array.R
\name{rmfi_parse_array}
\alias{rmfi_parse_array}
\title{Get an array specified by a free-format control record from the text lines analyzed in an \code{\link{RMODFLOW}} \code{read.*} function}
\usage{
rmfi_parse_array(remaining_lines, nrow, ncol, nlay, ndim = NULL,
  skip_header = FALSE)
}
\arguments{
\item{object}{MODFLOW input file text object, starting with the free-format control record}
}
\value{
A list containing the array and the remaining text of the MODFLOW input file
}
\description{
Get an array specified by a free-format control record from the text lines analyzed in an \code{\link{RMODFLOW}} \code{read.*} function
}
\keyword{internal}
