% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmfi-parse-variables.R
\name{rmfi_parse_variables}
\alias{rmfi_parse_variables}
\title{Read modflow variables
If all are numbers, returns numeric, otherwise returns character vector}
\usage{
rmfi_parse_variables(remaining_lines)
}
\description{
Read modflow variables
If all are numbers, returns numeric, otherwise returns character vector
}
