% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmfi-performance-measures.R
\name{rmfi_performance_measures}
\alias{rmfi_performance_measures}
\title{Model performance measures}
\usage{
rmfi_performance_measures(observations, predictions, print = F)
}
\description{
Model performance measures
}
