% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rmfi_write_array}
\alias{rmfi_write_array}
\title{Write modflow array
Internal function used in the write_* functions for writing array datasets}
\usage{
rmfi_write_array(array, file, cnstnt = 1, iprn = -1, append = TRUE,
  external = NULL, fname = NULL, binary = NULL,
  precision = "single", nam = NULL, xsection = FALSE, ...)
}
\arguments{
\item{external}{character vector with names corresponding to the dataset; used to write external arrays}

\item{fname}{character vector with names corresponding to the dataset; used to write open/close arrays}

\item{binary}{character vector with names corresponding to the dataset; used to write external or open/close arrays}

\item{precision}{character: either \code{'single'} (default) or \code{'double'}. Denotes the precision of binary files}

\item{nam}{\code{\link{RMODFLOW}} nam object; used when writing external arrays}

\item{xsection}{logical; does the array represent a NLAY x NCOL cross-section. Passed to \code{rmf_write_array}}

\item{...}{ignored}
}
\description{
Write modflow array
Internal function used in the write_* functions for writing array datasets
}
