% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run_modflow}
\alias{run_modflow}
\title{Run a MODFLOW model}
\usage{
run_modflow(nam, dir = getwd(), mfVersion = "mf2005", doPlot = T)
}
\arguments{
\item{file}{Filename; typically "*.ba6"}
}
\value{
Object of class ba6
}
\description{
\code{read_ba6} reads in a MODFLOW basic file and returns it as an \code{\link{RMODFLOW}} ba6 object.
}

