% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run_modflow}
\alias{run_modflow}
\title{Run a MODFLOW model}
\usage{
run_modflow(file, modflow_executable = "mf2005", par = NULL)
}
\arguments{
\item{file}{Path to name file; typically "*.nam"}

\item{modflow_executable}{name of the MODFLOW executable to use}
}
\description{
\code{run_modflow} runs a MODFLOW model.
}

