% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_modflow.R
\name{run_modflow}
\alias{run_modflow}
\title{Run a MODFLOW model}
\usage{
run_modflow(file, modflow_executable = ifelse(Sys.info()["sysname"] ==
  "Linux", "wine mf2005.exe", "mf2005"), par = NULL)
}
\arguments{
\item{file}{path to name file; typically '*.nam'}

\item{modflow_executable}{name of the MODFLOW executable to use}

\item{par}{vector of parameter value file parameter values to run the model with}
}
\description{
\code{run_modflow} runs a MODFLOW model.
}

