% Generated by roxygen2 (4.0.2): do not edit by hand
\name{str.modflow_package}
\alias{str.modflow_package}
\title{Structure of modflow package object}
\usage{
\method{str}{modflow_package}(modflow_package)
}
\arguments{
\item{modflow_package}{RMODFLOW modflow package object}
}
\description{
Structure of modflow package object
}
\details{
Provides the object comment attribute additional to the standard \code{str} output.
}

