% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write_ba6}
\alias{write_ba6}
\title{Write a MODFLOW basic file}
\usage{
write_ba6(ba6, file, IPRN = -1)
}
\arguments{
\item{ba6}{an \code{\link{RMODFLOW}} ba6 object}

\item{file}{filename to write to; typically '*.ba6'}

\item{IPRN}{format code for printing arrays in the listing file; defaults to -1 (no printing)}
}
\value{
\code{NULL}
}
\description{
\code{write_ba6} writes a MODFLOW basic file based on an \code{\link{RMODFLOW}} ba6 object.
}

