% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write_dis}
\alias{write_dis}
\title{Write a MODFLOW discretization file}
\usage{
write_dis(dis, file, IPRN = -1)
}
\arguments{
\item{dis}{an \code{\link{RMODFLOW}} dis object}

\item{file}{filename to write to; typically '*.dis'}

\item{IPRN}{format code for printing arrays in the listing file; defaults to -1 (no printing)}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW discretization file
}

