% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write_hob}
\alias{write_hob}
\title{Write a MODFLOW head observations file}
\usage{
write_hob(hob, file)
}
\arguments{
\item{hob}{an \code{\link{RMODFLOW}} hob object}

\item{file}{filename to write to; typically '*.hob'}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW head observations file
}

