% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_huf.R
\name{write_huf}
\alias{write_huf}
\title{Write a MODFLOW hydrogeologic unit flow file}
\usage{
write_huf(huf, file = {    
  cat("Please select huf file to overwrite or provide new filename ...\\n")    
  file.choose() }, iprn = -1)
}
\arguments{
\item{huf}{an \code{\link{RMODFLOW}} huf object}

\item{file}{filename to write to; typically '*.huf'}

\item{iprn}{format code for printing arrays in the listing file; defaults to -1 (no printing)}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW hydrogeologic unit flow file
}

