% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_mlt.R
\name{write_mlt}
\alias{write_mlt}
\title{Write a MODFLOW multiplier file}
\usage{
write_mlt(mlt, file, IPRN = -1)
}
\arguments{
\item{mlt}{an \code{\link{RMODFLOW}} mlt object}

\item{file}{filename to write to; typically '*.mlt'}

\item{IPRN}{format code for printing arrays in the listing file; defaults to -1 (no printing)}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW multiplier file
}

