% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{write_modflow_array}
\alias{write_modflow_array}
\title{Write modflow array
Internal function used in the write_* functions for writing array datasets}
\usage{
write_modflow_array(array, file, cnstnt = 1, iprn = -1, append = TRUE)
}
\description{
Write modflow array
Internal function used in the write_* functions for writing array datasets
}

