% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_modflow_array.R
\name{write_modflow_array}
\alias{write_modflow_array}
\title{Write modflow array}
\usage{
write_modflow_array(modflow_array, file, IPRN = -1, append = TRUE)
}
\description{
Write modflow array
}

