% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pvl.R
\name{write_pvl}
\alias{write_pvl}
\title{Write a MODFLOW parameter value file}
\usage{
write_pvl(pvl, file = {    
  cat("Please select pvl file to overwrite or provide new filename ...\\n")    
  file.choose() })
}
\arguments{
\item{pvl}{an \code{\link{RMODFLOW}} pvl object}

\item{file}{filename to write to; typically '*.pvl'}
}
\value{
\code{NULL}
}
\description{
Write a MODFLOW parameter value file
}

